+++ paper/paper.md
@@ -7,6 +7,7 @@ tags:
 authors:
  - name: Cameron Davidson-Pilon
    orcid: 0000-0003-1794-9143
+affiliations:
 date: 07 March 2019
 bibliography: paper.bib
 ---
