+++ paper/paper.bib
@@ -6,7 +6,7 @@
   url = {https://CRAN.R-project.org/package=survival},
 }
 
-@Article{,
+@Article{flexsurv,
   title = {{flexsurv}: A Platform for Parametric Survival Modeling in
     {R}},
   author = {Christopher Jackson},
