### 2019-03-06 #1305 @CamDavidsonPilon
SORRY ABOUT THE NOISE 🔈

I think I got it.
