+++ paper/paper.bib
@@ -1,197 +1,12 @@
-@Manual{survival-package,
-  title = {A Package for Survival Analysis in S},
-  author = {Terry M Therneau},
-  year = {2015},
-  note = {version 2.38},
-  url = {https://CRAN.R-project.org/package=survival},
+@MISC{wiki:ph,
+    AUTHOR = {{Wikipedia contributors}"},
+    TITLE = "Proportional hazards model --- {Wikipedia}{,} The Free Encyclopedia",
+    URL = "https://en.wikipedia.org/w/index.php?title=Proportional_hazards_model&oldid=885553852",
+    NOTE = "[Online; accessed 6-March-2019]",
+    YEAR = "2019",
 }
-
-@Article{flexsurv,
-  title = {{flexsurv}: A Platform for Parametric Survival Modeling in
-    {R}},
-  author = {Christopher Jackson},
-  journal = {Journal of Statistical Software},
-  year = {2016},
-  volume = {70},
-  number = {8},
-  pages = {1--33},
-  doi = {10.18637/jss.v070.i08},
-}
-
-@article{Harrell1996MultivariablePM,
-  title={Multivariable prognostic models: issues in developing models, evaluating assumptions and adequacy, and measuring and reducing errors.},
-  author={Frank E. Harrell and Kerry L. Lee and Daniel B. Mark},
-  journal={Statistics in medicine},
-  year={1996},
-  volume={15 4},
-  pages={
-          361-87
-        }
-}
-
-@Article{Edwards2016,
-author="Edwards, Jessie K.
-and Hester, Laura L.
-and Gokhale, Mugdha
-and Lesko, Catherine R.",
-title="Methodologic Issues when Estimating Risks in Pharmacoepidemiology",
-journal="Current Epidemiology Reports",
-year="2016",
-month="Dec",
-day="01",
-volume="3",
-number="4",
-pages="285--296",
-abstract="Risk is an important parameter to describe the occurrence of health outcomes over time. However, many outcomes of interest in healthcare settings, such as disease incidence, treatment initiation, and cause-specific mortality, may be precluded from occurring by other events, often referred to as competing events. Here, we review straightforward approaches to estimate risk in the presence of competing events.",
-issn="2196-2995",
-doi="10.1007/s40471-016-0089-1",
-url="https://doi.org/10.1007/s40471-016-0089-1"
-}
-
-@article{doi:10.1111/j.1751-5823.2009.00095\_1.x,
-author = {Ghosh, Jayanta K.},
-title = {Survival and Event History Analysis: A Process Point of View by Odd O. Aalen, Ørnulf Borgan, Håkon K. Gjessing},
-journal = {International Statistical Review},
-volume = {77},
-number = {3},
-pages = {463-464},
-doi = {10.1111/j.1751-5823.2009.00095\_1.x},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1751-5823.2009.00095_1.x},
-eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1751-5823.2009.00095_1.x},
-year = {2009}
-}
-
-
-@MISC {133822,
-    TITLE = {Stratified concordance index (survival::survConcordance)},
-    AUTHOR = {Ben Kuhn (https://stats.stackexchange.com/users/60642/ben-kuhn)},
-    HOWPUBLISHED = {Cross Validated},
-    NOTE = {URL:https://stats.stackexchange.com/q/133822 (version: 2015-01-17)},
-    EPRINT = {https://stats.stackexchange.com/q/133822},
-    URL = {https://stats.stackexchange.com/q/133822}
-}
-
[…]
