+++ paper/paper.bib
@@ -6,36 +6,7 @@
     YEAR = "2019",
 }
 @MISC{skipper_seabold_2017_275519,
-    AUTHOR = {{Skipper Seabold and
-                  Josef Perktold and
-                  ChadFulton and
-                  Kerby Shedden and
-                  j-grana6 and
-                  Vincent Arel-Bundock and
-                  Wes McKinney and
-                  Ian Langmore and
-                  Bart Baker and
-                  Ralf Gommers and
-                  Kevin Sheppard and
-                  Enrico Giampieri and
-                  Matthew Brett and
-                  Paul Hobson and
-                  Jarrod Millman and
-                  Vincent and
-                  Peter Quackenbush and
-                  alexbrc and
-                  Tyler Hartley and
-                  Fernando Perez and
-                  Yaroslav Halchenko and
-                  scottpiraino and
-                  Tommy Guy and
-                  Tom Augspurger and
-                  Denis A. Engemann and
-                  Padarn Wilson and
-                  kiyoto and
-                  Pierre Barbier de Reuille and
-                  Chris Jordan-Squire and
-                  Gael Varoquaux}},
+    AUTHOR = {{Skipper Seabold et al.}},
     TITLE = {statsmodels/statsmodels: Version 0.8.0 Release},
     URL = {https://doi.org/10.5281/zenodo.275519},
     DOI = {10.5281/zenodo.275519},
@@ -43,17 +14,7 @@
     MONTH = feb,
 }
 @MISC{nathaniel_j_smith_2018_1472929,
-    AUTHOR = {{Nathaniel J. Smith and
-                  Christian Hudon and
-                  broessli and
-                  Skipper Seabold and
-                  Peter Quackenbush and
-                  Michael Hudson-Doyle and
-                  Max Humber and
-                  Katrin Leinweber and
-                  Hassan Kibirige and
-                  Cameron Davidson-Pilon and
-                  Andrey Portnoy}},
+    AUTHOR = {{Nathaniel J. Smith et al.},
     TITLE = {pydata/patsy: v0.5.1},
     URL = {https://doi.org/10.5281/zenodo.1472929},
     DOI = {10.5281/zenodo.1472929},
