+++ paper/paper.bib
@@ -14,7 +14,7 @@
     MONTH = feb,
 }
 @MISC{nathaniel_j_smith_2018_1472929,
-    AUTHOR = {{Nathaniel J. Smith et al.},
+    AUTHOR = {{Nathaniel J. Smith et al.}},
     TITLE = {pydata/patsy: v0.5.1},
     URL = {https://doi.org/10.5281/zenodo.1472929},
     DOI = {10.5281/zenodo.1472929},
