### 2019-03-08 #1305 @trallard
Hi @CamDavidsonPilon  I will be editing this paper let's ping the reviewers. Once we have found 2 reviewers we can proceed to the review of the software

---

Dear potential reviewers,

Based on your expertise I would like to ask if you can help out with this JOSS submission.

**Title**: lifelines: survival analysis in Python

**Summary**: lifelines is a pure Python implementation of the best parts of survival analysis. We'd love to hear if you are using lifelines, please leave an Issue and let us know your thoughts on the library.

**Article proof**: [https://github.com/openjournals/joss-papers/blob/joss.01305/joss.01305/10.21105.joss.01305.pdf](https://github.com/openjournals/joss-papers/blob/joss.01305/joss.01305/10.21105.joss.01305.pdf)

**Submitting author**: @CamDavidsonPilon (Cameron Davidson-Pilon)

**Version:** 0.20.0

**Repository**: [https://github.com/CamDavidsonPilon/lifelines](https://github.com/CamDavidsonPilon/lifelines)

---

@becarioprecario: Virgilio, are you able to and interested in taking this review?
@sunhwan: Sunhwan, are you able to and interested in taking this review?
