### 2019-03-08 68a31e0
add contributing; close #668
