### 2019-03-12 #1317 @whedon
[REVIEW]: lifelines: survival analysis in Python
**Submitting author:** @CamDavidsonPilon (<a href="http://orcid.org/0000-0003-1794-9143">Cameron Davidson-Pilon</a>)
**Repository:** <a href="https://github.com/camdavidsonpilon/lifelines" target ="_blank">https://github.com/camdavidsonpilon/lifelines</a>
**Version:** v0.22.2
