### 2019-03-12 #1317 @trallard
👋 @becarioprecario and @sunhwan this is the issue for the review. Each of you has a checklist of the items you have to look at while conducting your review.

Please use this issue to comment, make suggestions and discuss anything related to the submission. If you need to request changes or go in detail on something it is advisable to make an issue on the software repo [https://github.com/camdavidsonpilon/lifelines](https://github.com/camdavidsonpilon/lifelines) and reference this issue to keep track of the process.

If any of you or @CamDavidsonPilon have any questions at any point feel free to ping me.

Happy reviewing 🔎💻
