+++ lifelines/fitters/aalen_additive_fitter.py
@@ -481,7 +481,7 @@ It's important to know that the naive variance estimates of the coefficients are
     @property
     def score_(self):
         """
-        The concordance score (also known as the c-index) of the fit.  The c-index is a generalization of the AUC
+        The concordance score (also known as the c-index) of the fit.  The c-index is a generalization of the ROC AUC
         to survival data, including censorships.
 
         For this purpose, the ``score_`` is a measure of the predictive accuracy of the fitted model
