+++ lifelines/fitters/__init__.py
@@ -1288,7 +1288,7 @@ class ParametericRegressionFitter(BaseFitter):
     @property
     def score_(self):
         """
-        The concordance score (also known as the c-index) of the fit.  The c-index is a generalization of the AUC
+        The concordance score (also known as the c-index) of the fit.  The c-index is a generalization of the ROC AUC
         to survival data, including censorships.
 
         For this purpose, the ``score_`` is a measure of the predictive accuracy of the fitted model
