+++ docs/conf.py
@@ -292,3 +292,7 @@ if not on_rtd:  # only import and set the theme if we're building docs locally
 
     html_theme = "sphinx_rtd_theme"
     html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
+
+
+def setup(app):
+    app.add_stylesheet("custom.css")
