### 2019-03-13 99df6ea
some css and use fill to make printing nicer
