### 2019-03-15 3dcc9bd
small performance (~10%) increase for coxph fitter
