### 2019-03-16 e826d9e
More performance increases! We are now 1:1 with R's performance
