+++ lifelines/fitters/coxph_fitter.py
@@ -280,35 +280,32 @@ class CoxPHFitter(BaseFitter):
 
         self._norm_mean = X.mean(0)
         self._norm_std = X.std(0)
+        X_norm = normalize(X, self._norm_mean, self._norm_std)
 
         hazards_ = self._newton_rhaphson(
-            normalize(X, self._norm_mean, self._norm_std),
-            T,
-            E,
-            weights=weights,
-            initial_point=initial_point,
-            show_progress=show_progress,
-            step_size=step_size,
+            X_norm, T, E, weights=weights, initial_point=initial_point, show_progress=show_progress, step_size=step_size
         )
 
         self.hazards_ = pd.Series(hazards_, index=X.columns, name="coef") / self._norm_std
 
         self.variance_matrix_ = -inv(self._hessian_) / np.outer(self._norm_std, self._norm_std)
-        self.standard_errors_ = self._compute_standard_errors(
-            normalize(X, self._norm_mean, self._norm_std), T, E, weights
-        )
+        self.standard_errors_ = self._compute_standard_errors(X_norm, T, E, weights)
         self.confidence_intervals_ = self._compute_confidence_intervals()
-
-        self.baseline_hazard_ = self._compute_baseline_hazards(X, T, E, weights)
-        self.baseline_cumulative_hazard_ = self._compute_baseline_cumulative_hazard()
-        self.baseline_survival_ = self._compute_baseline_survival()
 
         self._predicted_partial_hazards_ = (
             self.predict_partial_hazard(X)
             .rename(columns={0: "P"})
-            .assign(T=self.durations.values, E=self.event_observed.values)
+            .assign(T=self.durations.values, E=self.event_observed.values, W=self.weights.values)
             .set_index(X.index)
         )
+        self.baseline_hazard_ = self._compute_baseline_hazards()
+        self.baseline_cumulative_hazard_ = self._compute_baseline_cumulative_hazard()
+        self.baseline_survival_ = self._compute_baseline_survival()
+
+        if hasattr(self, "_concordance_score_"):
+            # we have already fit the model.
+            del self._concordance_score_
+
         return self
 
     def _preprocess_dataframe(self, df):
@@ -1340,24 +1337,25 @@ See https://stats.stackexchange.com/questions/11109/how-to-deal-with-perfect-sep
         if X is an array, then the column ordering is assumed to be the
         same as the training dataset.
         """
-
         hazard_names = self.hazards_.index
-        if isinstance(X, pd.DataFrame):
-            order = hazard_names
-            X = X[order]
-            check_for_numeric_dtypes_or_raise(X)
-        elif isinstance(X, pd.Series) and ((X.shape[0] == len(hazard_names) + 2) or (X.shape[0] == len(hazard_names))):
+
+        if isinstance(X, pd.Series) and ((X.shape[0] == len(hazard_names) + 2) or (X.shape[0] == len(hazard_names))):
             X = X.to_frame().T
-            order = hazard_names
-            X = X[order]
-            check_for_numeric_dtypes_or_raise(X)
+            return self.predict_log_partial_hazard(X)
         elif isinstance(X, pd.Series):
             assert len(hazard_names) == 1, "Series not the correct argument"
-            X = pd.DataFrame(X)
+            X = X.to_frame().T
+            return self.predict_log_partial_hazard(X)
+
+        index = _get_index(X)
+
+        if isinstance(X, pd.DataFrame):
+            order = hazard_names
+            X = X.reindex(order, axis="columns")
             check_for_numeric_dtypes_or_raise(X)
+            X = X.values
 
         X = X.astype(float)
-        index = _get_index(X)
 
         X = normalize(X, self._norm_mean.values, 1)
[…]
