+++ tests/test_estimation.py
@@ -518,7 +518,7 @@ class TestUnivariateFitters:
             assert dif == 0
 
 
-class TestLogNormal:
+class TestLogNormalFitter:
     @pytest.fixture()
     def lnf(self):
         return LogNormalFitter()
@@ -2906,8 +2906,6 @@ Log-likelihood ratio test = 33.27 on 7 df, -log2(p)=15.37
         rossi = rossi[["week", "arrest", "fin", "age"]]
         cp = CoxPHFitter()
         cp.fit(rossi, "week", "arrest", weights_col="age")
-
-        npt.assert_almost_equal(cp.baseline_cumulative_hazard_["baseline hazard"].loc[0.0], 0.0, decimal=4)
         npt.assert_almost_equal(cp.baseline_cumulative_hazard_["baseline hazard"].loc[1.0], 0.00183466, decimal=4)
         npt.assert_almost_equal(cp.baseline_cumulative_hazard_["baseline hazard"].loc[2.0], 0.005880265, decimal=4)
         npt.assert_almost_equal(cp.baseline_cumulative_hazard_["baseline hazard"].loc[10.0], 0.035425868, decimal=4)
@@ -3226,7 +3224,7 @@ class TestAalenAdditiveFitter:
         assert abs((T_pred.values > T).mean() - 0.5) < 0.05
 
     def test_dataframe_input_with_nonstandard_index(self):
-        aaf = AalenAdditiveFitter(coef_penalizer=2.0)
+        aaf = AalenAdditiveFitter(coef_penalizer=5.0)
         df = pd.DataFrame(
             [(16, True, True), (1, True, True), (4, False, True)],
             columns=["duration", "done_feeding", "white"],
