+++ lifelines/utils/__init__.py
@@ -905,13 +905,17 @@ A very low variance means that the column {cols} completely determines whether a
         warnings.warn(warning_text, ConvergenceWarning)
 
 
+def correlation(x, y):
+    return np.corrcoef(x, y)[1, 0]
+
+
 def check_complete_separation_close_to_perfect_correlation(df, durations):
     # slow for many columns
     THRESHOLD = 0.99
     n, _ = df.shape
 
     if n > 500:
-        # let's sample to speed this n**2 algo up.
+        # let's sample to speed this up.
         df = df.sample(n=500, random_state=0).copy()
         durations = durations.sample(n=500, random_state=0).copy()
 
@@ -919,7 +923,7 @@ def check_complete_separation_close_to_perfect_correlation(df, durations):
     for col, series in df.iteritems():
         with np.errstate(invalid="ignore", divide="ignore"):
             rank_series = series.values.argsort()
-            if abs(np.corrcoef(rank_durations, rank_series)[1, 0]) >= THRESHOLD:
+            if abs(correlation(rank_durations, rank_series)) >= THRESHOLD:
                 warning_text = (
                     "Column %s has high sample correlation with the duration column. This may harm convergence. This could be a form of 'complete separation'. \
     See https://stats.stackexchange.com/questions/11109/how-to-deal-with-perfect-separation-in-logistic-regression"
@@ -943,7 +947,7 @@ def check_nans_or_infs(df_or_array):
             raise TypeError("NaNs were detected in the dataset. Try using pd.isnull to find the problematic values.")
     # isinf check is done after isnull check since np.isinf doesn't work on None values
     if isinstance(df_or_array, (pd.Series, pd.DataFrame)):
-        infs = df_or_array.values.astype(float) == np.Inf
+        infs = df_or_array.values == np.Inf
     else:
         infs = np.isinf(df_or_array)
 
