+++ docs/Changelog.rst
@@ -1,12 +1,42 @@
 Changelog
 ~~~~~~~~~
 
+0.20.1
+^^^^^^
+
+-  Some performance improvements to ``CoxPHFitter`` (about 30%). I know
+   it may seem silly, but we are now about the same or slighty faster
+   than the Cox model in R’s ``survival`` package (for some testing
+   datasets and some configurations). This is a big deal, because 1)
+   lifelines does more error checking prior, 2) R’s cox model is written
+   in C, and we are still pure Python/NumPy, 3) R’s cox model has
+   decades of development.
+-  suppressed unimportant warnings
+
+API changes
+'''''''''''
+
+-  Previously, lifelines *always* added a 0 row to
+   ``cph.baseline_hazard_``, even if there were no event at this time.
+   This is no longer the case. A 0 will still be added if there is a
+   duration (observed or not) at 0 occurs however.
+
+.. _section-1:
+
 0.20.0
 ^^^^^^
 
 -  Starting with 0.20.0, only Python3 will be supported. Over 75% of
    recent installs where Py3.
 -  Updated minimum dependencies, specifically Matplotlib and Pandas.
+
+New features
+''''''''''''
+
+-  smarter initialization for AFT models which should improve
+   convergence.
+
+.. _api-changes-1:
 
 API changes
 '''''''''''
@@ -23,10 +53,12 @@ Bug fixes
 
 -  Fixed a bug with plotting and ``check_assumptions``.
 
-.. _section-1:
+.. _section-2:
 
 0.19.5
 ^^^^^^
+
+.. _new-features-1:
 
 New features
 ''''''''''''
@@ -36,7 +68,7 @@ New features
    features or categorical variables.
 -  Convergence improvements for AFT models.
 
-.. _section-2:
+.. _section-3:
 
 0.19.4
 ^^^^^^
@@ -48,12 +80,12 @@ Bug fixes
 
 -  remove some bad print statements in ``CoxPHFitter``.
 
-.. _section-3:
+.. _section-4:
 
 0.19.3
 ^^^^^^
 
-.. _new-features-1:
+.. _new-features-2:
 
 New features
 ''''''''''''
@@ -65,12 +97,12 @@ New features
 -  Performance increase to ``print_summary`` in the ``CoxPHFitter`` and
    ``CoxTimeVaryingFitter`` model.
 
-.. _section-4:
+.. _section-5:
 
 0.19.2
 ^^^^^^
 
-.. _new-features-2:
+.. _new-features-3:
 
 New features
 ''''''''''''
@@ -88,12 +120,12 @@ Bug fixes
 -  Univariate fitters are more flexiable and can allow 2-d and
    DataFrames as inputs.
 
-.. _section-5:
+.. _section-6:
 
 0.19.1
 ^^^^^^
 
-.. _new-features-3:
+.. _new-features-4:
 
 New features
 ''''''''''''
@@ -101,7 +133,7 @@ New features
 -  improved stability of ``LogNormalFitter``
 -  Matplotlib for Python3 users are not longer forced to use 2.x.
 
-.. _api-changes-1:
+.. _api-changes-2:
 
 API changes
 '''''''''''
@@ -110,12 +142,12 @@ API changes
    ``PiecewiseExponential`` to the same as ``ExponentialFitter`` (from
[…]
