### 2019-03-16 182e0a7
v0.20.1 (#669)
* small performance (~10%) increase for coxph fitter

* More performance increases! We are now 1:1 with R's performance

* clean up

* bump version
