+++ lifelines/fitters/coxph_fitter.py
@@ -57,7 +57,7 @@ class BatchVsSingle:
             # https://github.com/CamDavidsonPilon/lifelines/issues/591 for original issue.
             # new values from from perf/batch_vs_single script.
             (batch_mode is None)
-            and (0.713869 + -0.000028 * n_total + 0.684229 * frac_dups + 0.000201 * n_total * frac_dups < 1)
+            and (0.712085 + -0.000025 * n_total + 0.579359 * frac_dups + 0.000044 * n_total * frac_dups < 1)
         ):
             return "batch"
         return "single"
@@ -280,35 +280,32 @@ class CoxPHFitter(BaseFitter):
 
         self._norm_mean = X.mean(0)
         self._norm_std = X.std(0)
+        X_norm = normalize(X, self._norm_mean, self._norm_std)
 
         hazards_ = self._newton_rhaphson(
-            normalize(X, self._norm_mean, self._norm_std),
-            T,
-            E,
-            weights=weights,
-            initial_point=initial_point,
-            show_progress=show_progress,
-            step_size=step_size,
+            X_norm, T, E, weights=weights, initial_point=initial_point, show_progress=show_progress, step_size=step_size
         )
 
         self.hazards_ = pd.Series(hazards_, index=X.columns, name="coef") / self._norm_std
 
         self.variance_matrix_ = -inv(self._hessian_) / np.outer(self._norm_std, self._norm_std)
-        self.standard_errors_ = self._compute_standard_errors(
-            normalize(X, self._norm_mean, self._norm_std), T, E, weights
-        )
+        self.standard_errors_ = self._compute_standard_errors(X_norm, T, E, weights)
         self.confidence_intervals_ = self._compute_confidence_intervals()
-
-        self.baseline_hazard_ = self._compute_baseline_hazards(X, T, E, weights)
-        self.baseline_cumulative_hazard_ = self._compute_baseline_cumulative_hazard()
-        self.baseline_survival_ = self._compute_baseline_survival()
 
         self._predicted_partial_hazards_ = (
             self.predict_partial_hazard(X)
             .rename(columns={0: "P"})
-            .assign(T=self.durations.values, E=self.event_observed.values)
+            .assign(T=self.durations.values, E=self.event_observed.values, W=self.weights.values)
             .set_index(X.index)
         )
+        self.baseline_hazard_ = self._compute_baseline_hazards()
+        self.baseline_cumulative_hazard_ = self._compute_baseline_cumulative_hazard()
+        self.baseline_survival_ = self._compute_baseline_survival()
+
+        if hasattr(self, "_concordance_score_"):
+            # we have already fit the model.
+            del self._concordance_score_
+
         return self
 
     def _preprocess_dataframe(self, df):
@@ -683,15 +680,17 @@ See https://stats.stackexchange.com/questions/11109/how-to-deal-with-perfect-sep
     @staticmethod
     def _trivial_log_likelihood_batch(T, E, weights):
         # used for log-likelihood test
+        n = T.shape[0]
         log_lik = 0
-        unique_death_times = np.unique(T)
+        _, counts = np.unique(-T, return_counts=True)
         risk_phi = 0
+        pos = n
 
-        for t in reversed(unique_death_times):
+        for count_of_removals in counts:
 
-            ix = T == t
+            slice_ = slice(pos - count_of_removals, pos)
 
-            weights_at_t = weights[ix]
+            weights_at_t = weights[slice_]
 
             phi_i = weights_at_t
 
@@ -699,11 +698,12 @@ See https://stats.stackexchange.com/questions/11109/how-to-deal-with-perfect-sep
             risk_phi = risk_phi + array_sum_to_scalar(phi_i)
 
[…]
