### 2019-03-18 e436fd3
clean up some typos and move step size to a more obvious place
