+++ lifelines/datasets/__init__.py
@@ -58,7 +58,7 @@ def load_multicenter_aids_cohort_study(**kwargs):
 
         AIDSY: date of AIDS diagnosis
         W: years from AIDS diagnosis to study entry
-        T: years from AIDS diagnosis to minimum of death or censoring 
+        T: years from AIDS diagnosis to minimum of death or censoring
         D: indicator of death during follow up
 
 
@@ -136,9 +136,9 @@ def load_canadian_senators(**kwargs):
 
 def load_dd(**kwargs):
     """
-    Classification of political regimes as democracy and dictatorship. 
-    Classification of democracies as parliamentary, semi-presidential (mixed) and presidential. 
-    Classification of dictatorships as military, civilian and royal. 
+    Classification of political regimes as democracy and dictatorship.
+    Classification of democracies as parliamentary, semi-presidential (mixed) and presidential.
+    Classification of dictatorships as military, civilian and royal.
     Coverage: 202 countries, from 1946 or year of independence to 2008.::
 
         Size: (1808, 12)
@@ -187,12 +187,12 @@ def load_larynx(**kwargs):
 
         Size: (89,6)
         Example:
-                time  age  death  Stage II  Stage III  Stage IV
-            0    0.6   77      1         0          0         0
-            1    1.3   53      1         0          0         0
-            2    2.4   45      1         0          0         0
-            3    2.5   57      0         0          0         0
-            4    3.2   58      1         0          0         0
+            time  age  death  Stage II  Stage III  Stage IV
+             0.6   77      1         0          0         0
+             1.3   53      1         0          0         0
+             2.4   45      1         0          0         0
+             2.5   57      0         0          0         0
+             3.2   58      1         0          0         0
 
     """
     return _load_dataset("larynx.csv", **kwargs)
@@ -204,12 +204,12 @@ def load_lung(**kwargs):
 
         Size: (288,10)
         Example:
-           inst  time  status  age  sex  ph.ecog  ph.karno  pat.karno  meal.cal  wt.loss
-        0   3.0   306       2   74    1      1.0      90.0      100.0    1175.0      NaN
-        1   3.0   455       2   68    1      0.0      90.0       90.0    1225.0     15.0
-        2   3.0  1010       1   56    1      0.0      90.0       90.0       NaN     15.0
-        3   5.0   210       2   57    1      1.0      90.0       60.0    1150.0     11.0
-        4   1.0   883       2   60    1      0.0     100.0       90.0       NaN      0.0
+         inst  time  status  age  sex  ph.ecog  ph.karno  pat.karno  meal.cal  wt.loss
+          3.0   306       2   74    1      1.0      90.0      100.0    1175.0      NaN
+          3.0   455       2   68    1      0.0      90.0       90.0    1225.0     15.0
+          3.0  1010       1   56    1      0.0      90.0       90.0       NaN     15.0
+          5.0   210       2   57    1      1.0      90.0       60.0    1150.0     11.0
+          1.0   883       2   60    1      0.0     100.0       90.0       NaN      0.0
 
     References
     -----------
@@ -225,12 +225,12 @@ def load_panel_test(**kwargs):
 
         Size: (28,5)
         Example:
-               id  t  E  var1  var2
-            0   1  1  0   0.0     1
-            1   1  2  0   0.0     1
-            2   1  3  0   4.0     3
-            3   1  4  1   8.0     4
-            4   2  1  0   1.2     1
+            id  t  E  var1  var2
+             1  1  0   0.0     1
+             1  2  0   0.0     1
+             1  3  0   4.0     3
+             1  4  1   8.0     4
+             2  1  0   1.2     1
 
     """
[…]
