+++ tests/test_estimation.py
@@ -17,6 +17,8 @@ import pandas as pd
 import pytest
 from scipy.stats import weibull_min, norm, logistic
 
+from flaky import flaky
+
 from pandas.util.testing import assert_frame_equal, assert_series_equal
 import numpy.testing as npt
 
@@ -162,6 +164,7 @@ class TestBaseFitter:
 
 
 class TestParametricUnivariateFitters:
+    @flaky
     def test_confidence_interval_is_expected(self):
 
         from autograd.scipy.special import logit
@@ -316,7 +319,6 @@ class TestUnivariateFitters:
             fitter.fit(T)
             assert hasattr(fitter, "plot")
 
-    @pytest.mark.xfail
     def test_univariate_fitters_ok_if_given_timedelta(self, univariate_fitters):
         t = pd.Series(
             [pd.to_datetime("2015-01-01 12:00"), pd.to_datetime("2015-01-02"), pd.to_datetime("2015-01-02 12:00")]
@@ -706,6 +708,30 @@ class TestLogLogisticFitter:
 
 
 class TestWeibullFitter:
+    @flaky
+    @pytest.mark.parametrize("N", [50, 100, 500, 1000])
+    def test_left_censorship_inference(self, N):
+        T_actual = 0.5 * np.random.weibull(5, size=N)
+
+        MIN_0 = np.percentile(T_actual, 5)
+        MIN_1 = np.percentile(T_actual, 10)
+        MIN_2 = np.percentile(T_actual, 30)
+        MIN_3 = np.percentile(T_actual, 50)
+
+        T = T_actual.copy()
+        ix = np.random.randint(4, size=N)
+
+        T = np.where(ix == 0, np.maximum(T, MIN_0), T)
+        T = np.where(ix == 1, np.maximum(T, MIN_1), T)
+        T = np.where(ix == 2, np.maximum(T, MIN_2), T)
+        T = np.where(ix == 3, np.maximum(T, MIN_3), T)
+        E = T_actual == T
+
+        wf = WeibullFitter().fit(T, E, left_censorship=True)
+
+        assert wf.summary.loc["rho_", "lower 0.95"] < 5 < wf.summary.loc["rho_", "upper 0.95"]
+        assert wf.summary.loc["lambda_", "lower 0.95"] < 0.5 < wf.summary.loc["lambda_", "upper 0.95"]
+
     def test_weibull_with_delayed_entries(self):
         # note the the independence of entry and final time is really important
         # (also called non-informative)
@@ -777,10 +803,20 @@ class TestWeibullFitter:
 class TestExponentialFitter:
     def test_fit_computes_correct_lambda_(self):
         T = np.array([10, 10, 10, 10], dtype=float)
-        E = np.array([1, 0, 0, 0], dtype=float)
+        E = np.array([1, 1, 1, 0], dtype=float)
         enf = ExponentialFitter()
         enf.fit(T, E)
         assert abs(enf.lambda_ - (T.sum() / E.sum())) < 10e-6
+
+    def test_fit_computes_correct_asymptotic_variance(self):
+        N = 5000
+        T = np.random.exponential(size=N)
+        C = np.random.exponential(size=N)
+        E = T < C
+        T = np.minimum(T, C)
+        enf = ExponentialFitter()
+        enf.fit(T, E)
+        assert abs(enf.summary.loc["lambda_", "se(coef)"] ** 2 - (T.sum() / E.sum()) ** 2 / N) < 1e-3
 
 
 class TestKaplanMeierFitter:
