### 2019-03-19 03fd138
Experiment with left censorship (#671)
* fix the types (again: it feels like I do this every N months)

* wip
