+++ tests/test_estimation.py
@@ -17,6 +17,8 @@ import pandas as pd
 import pytest
 from scipy.stats import weibull_min, norm, logistic
 
+from flaky import flaky
+
 from pandas.util.testing import assert_frame_equal, assert_series_equal
 import numpy.testing as npt
 
@@ -162,6 +164,7 @@ class TestBaseFitter:
 
 
 class TestParametricUnivariateFitters:
+    @flaky
     def test_confidence_interval_is_expected(self):
 
         from autograd.scipy.special import logit
@@ -316,18 +319,18 @@ class TestUnivariateFitters:
             fitter.fit(T)
             assert hasattr(fitter, "plot")
 
-    @pytest.mark.xfail
     def test_univariate_fitters_ok_if_given_timedelta(self, univariate_fitters):
         t = pd.Series(
             [pd.to_datetime("2015-01-01 12:00"), pd.to_datetime("2015-01-02"), pd.to_datetime("2015-01-02 12:00")]
         )
         T = pd.to_datetime("2015-01-03") - t
         for fitter in univariate_fitters:
-            f = fitter().fit(T)
-            try:
-                npt.assert_allclose(f.timeline, 1e9 * 12 * 60 * 60 * np.array([0, 1, 2, 3]))
-            except:
-                npt.assert_allclose(f.timeline, 1e9 * 12 * 60 * 60 * np.array([1, 2, 3]))
+            with pytest.warns(UserWarning, match="convert"):
+                f = fitter().fit(T)
+                try:
+                    npt.assert_allclose(f.timeline, 1e9 * 12 * 60 * 60 * np.array([0, 1, 2, 3]))
+                except:
+                    npt.assert_allclose(f.timeline, 1e9 * 12 * 60 * 60 * np.array([1, 2, 3]))
 
     def test_univariate_fitters_okay_if_given_boolean_col_with_object_dtype(self, univariate_fitters):
         df = pd.DataFrame({"T": [1, 2, 3, 4, 5], "E": [True, True, True, True, None]})
@@ -336,7 +339,8 @@ class TestUnivariateFitters:
         assert df["E"].dtype == object
 
         for fitter in univariate_fitters:
-            fitter().fit(df["T"], df["E"])
+            with pytest.warns(UserWarning, match="convert"):
+                fitter().fit(df["T"], df["E"])
 
     def test_predict_methods_returns_a_scalar_or_a_array_depending_on_input(
         self, positive_sample_lifetimes, univariate_fitters
@@ -704,6 +708,30 @@ class TestLogLogisticFitter:
 
 
 class TestWeibullFitter:
+    @flaky
+    @pytest.mark.parametrize("N", [50, 100, 500, 1000])
+    def test_left_censorship_inference(self, N):
+        T_actual = 0.5 * np.random.weibull(5, size=N)
+
+        MIN_0 = np.percentile(T_actual, 5)
+        MIN_1 = np.percentile(T_actual, 10)
+        MIN_2 = np.percentile(T_actual, 30)
+        MIN_3 = np.percentile(T_actual, 50)
+
+        T = T_actual.copy()
+        ix = np.random.randint(4, size=N)
+
+        T = np.where(ix == 0, np.maximum(T, MIN_0), T)
+        T = np.where(ix == 1, np.maximum(T, MIN_1), T)
+        T = np.where(ix == 2, np.maximum(T, MIN_2), T)
+        T = np.where(ix == 3, np.maximum(T, MIN_3), T)
+        E = T_actual == T
+
+        wf = WeibullFitter().fit(T, E, left_censorship=True)
+
[…]
