### 2019-03-19 2a72100
improve convergence of some uni fitters
