+++ lifelines/fitters/log_logistic_fitter.py
@@ -74,3 +74,7 @@ class LogLogisticFitter(KnownModelParametericUnivariateFitter):
     def _cumulative_hazard(self, params, times):
         alpha_, beta_ = params
         return np.log1p((times / alpha_) ** beta_)
+
+    def _log_1m_sf(self, params, times):
+        alpha_, beta_ = params
+        return -np.log1p((times / alpha_) ** -beta_)
