### 2019-03-19 56ae827
removed inversion of plotting lifetimes
