### 2019-03-19 53ec5a5
removed inversion of plotting lifetimes (#673)
