+++ docs/Survival analysis with lifelines.rst
@@ -612,8 +612,8 @@ To aid model selection, _lifelines_ has provided qq-plots, `Selecting a parametr
 Other types of censoring
 ''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
 
-Left censored data
-~~~~~~~~~~~~~~~~~~~~~~~~~~
+Left censored data and non-detection
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 We've mainly been focusing on *right-censoring*, which describes cases where we do not observe the death event.
 This situation is the most common one. Alternatively, there are situations where we do not observe the *birth* event
@@ -621,43 +621,92 @@ occurring. Consider the case where a doctor sees a delayed onset of symptoms of 
 is unsure *when* the disease was contracted (birth), but knows it was before the discovery.
 
 Another situation where we have left-censored data is when measurements have only an upper bound, that is, the measurements
-instruments could only detect the measurement was *less* than some upper bound. This bound is often called the limit of detection (LOD). In practice, there could be more than one LOD. One very important statistical lesson: don't "fill-in" this value naively. It's tempting to use something like one-half the LOD, but this will cause *lots* of bias in downstream analysis.
+instruments could only detect the measurement was *less* than some upper bound. This bound is often called the limit of detection (LOD). In practice, there could be more than one LOD. One very important statistical lesson: don't "fill-in" this value naively. It's tempting to use something like one-half the LOD, but this will cause *lots* of bias in downstream analysis. An example dataset is below:
+
+.. code:: python
+
+    from lifelines.datasets import load_nh4
+    df = load_nh4()[['NH4.Orig.mg.per.L', 'NH4.mg.per.L', 'Censored']]
+    print(df.head())
+
+    """
+      NH4.Orig.mg.per.L  NH4.mg.per.L  Censored
+    1            <0.006         0.006      True
+    2            <0.006         0.006      True
+    3             0.006         0.006     False
+    4             0.016         0.016     False
+    5            <0.006         0.006      True
+    """
 
 *lifelines* has support for left-censored datasets in most univariate models, including the ``KaplanMeierFitter`` class, by adding the keyword ``left_censoring=True`` (default ``False``) to the call to ``fit``.
 
 .. code:: python
 
-    from lifelines.datasets import load_lcd
-    lcd_dataset = load_lcd()
 
-    T = lcd_dataset['T']
-    E = lcd_dataset['E'] #boolean array, True if observed.
+    T, E = df['NH4.mg.per.L'], ~df['Censored']
 
     kmf = KaplanMeierFitter()
     kmf.fit(T, E, left_censorship=True)
 
-Instead of producing a survival function, left-censored data is more interested in the cumulative density function
-of time to birth. This is available as the ``cumulative_density_`` property after fitting the data.
+Instead of producing a survival function, left-censored data analysis is more interested in the cumulative density function. This is available as the ``cumulative_density_`` property after fitting the data.
 
 .. code:: python
 
-    print(kmf.cumulative_density_)
+    print(kmf.cumulative_density_.head())
+
     kmf.plot() #will plot the CDF
+    plt.xlabel("Concentration of NH_4")
+
+    """
+              KM_estimate
+    timeline
+    0.000        0.379897
+    0.006        0.401002
+    0.007        0.464319
+    0.008        0.478828
+    0.009        0.536868
+    """
 
 
 .. image:: images/lifelines_intro_lcd.png
 
[…]
