### 2019-03-21 e3ba90e
last few things and bump version
