+++ docs/Changelog.rst
@@ -1,6 +1,45 @@
 Changelog
 ~~~~~~~~~
 
+0.20.2
+^^^^^^
+
+New features
+''''''''''''
+
+-  Left censoring is now supported in univariate parametric models:
+   ``.fit(..., left_censorship=True)``. Examples are in the docs.
+-  new dataset: ``lifelines.datasets.load_nh4()``
+-  Univariate parametric models now include, by default, support for the
+   cumulative density function: ``.cumulative_density_``,
+   ``.confidence_interval_cumulative_density_``,
+   ``plot_cumulative_density()``, ``cumulative_density_at_times(t)``.
+-  add a ``lifelines.plotting.qq_plot`` for univariate parametric models
+   that handles censored data.
+
+API changes
+'''''''''''
+
+-  ``plot_lifetimes`` no longer reverses the order when plotting. Thanks
+   @vpolimenov!
+-  The ``C`` column in ``load_lcd`` dataset is renamed to ``E``.
+
+Bug fixes
+'''''''''
+
+-  fixed a naming error in ``KaplanMeierFitter`` when
+   ``left_censorship`` was set to True, ``plot_cumulative_density_()``
+   is now ``plot_cumulative_density()``.
+-  added some error handling when passing in timedeltas. Ideally, users
+   don’t pass in timedeltas, as the scale is ambiguous. However, the
+   error message before was not obvious, so we do some conversion, warn
+   the user, and pass it through.
+-  ``qth_survival_times`` for a truncated CDF would return ``np.inf`` if
+   the q parameter was below the truncation limit. This should have been
+   ``-np.inf``
+
+.. _section-1:
+
 0.20.1
 ^^^^^^
 
@@ -13,6 +52,8 @@ Changelog
    decades of development.
 -  suppressed unimportant warnings
 
+.. _api-changes-1:
+
 API changes
 '''''''''''
 
@@ -21,7 +62,7 @@ API changes
    This is no longer the case. A 0 will still be added if there is a
    duration (observed or not) at 0 occurs however.
 
-.. _section-1:
+.. _section-2:
 
 0.20.0
 ^^^^^^
@@ -30,13 +71,15 @@ API changes
    recent installs where Py3.
 -  Updated minimum dependencies, specifically Matplotlib and Pandas.
 
+.. _new-features-1:
+
 New features
 ''''''''''''
 
 -  smarter initialization for AFT models which should improve
    convergence.
 
-.. _api-changes-1:
+.. _api-changes-2:
 
 API changes
 '''''''''''
@@ -48,17 +91,19 @@ API changes
    transposed now (previous parameters where columns, now parameters are
    rows).
 
+.. _bug-fixes-1:
+
 Bug fixes
 '''''''''
 
 -  Fixed a bug with plotting and ``check_assumptions``.
 
-.. _section-2:
+.. _section-3:
 
 0.19.5
 ^^^^^^
 
-.. _new-features-1:
+.. _new-features-2:
 
 New features
 ''''''''''''
@@ -68,24 +113,24 @@ New features
    features or categorical variables.
 -  Convergence improvements for AFT models.
 
-.. _section-3:
+.. _section-4:
 
 0.19.4
[…]
