### 2019-03-21 f7ca123
v0.20.2 (#672)
* clean up some typos and move step size to a more obvious place

* Experiment with left censorship (#671)

* fix the types (again: it feels like I do this every N months)

* wip

* improve convergence of some uni fitters

* shitty qq plot

* wip

* removed inversion of plotting lifetimes (#673)

* adding plotting tests

* docs

* some docs

* some docs

* image

* last few things and bump version
