+++ lifelines/utils/__init__.py
@@ -54,7 +54,7 @@ def qth_survival_times(q, survival_functions, cdf=False):
 
     Parameters
     ----------
-    q: float
+    q: float or array
       a float between 0 and 1 that represents the time when the survival function hits the qth percentile.
     survival_functions: a (n,d) DataFrame or numpy array.
       If DataFrame, will return index values (actual times)
@@ -129,7 +129,7 @@ def qth_survival_time(q, survival_function, cdf=False):
         survival_function = survival_function.T.squeeze()
     if cdf:
         if survival_function.iloc[0] > q:
-            return np.inf
+            return -np.inf
         v = survival_function.index[survival_function.searchsorted([q])[0]]
     else:
         if survival_function.iloc[-1] > q:
@@ -869,7 +869,7 @@ df.groupby(level=1).apply(lambda g: g.index.get_level_values(0).is_non_overlappi
 
 
 def _low_var(df):
-    return df.var(0) < 10e-5
+    return df.var(0) < 1e-4
 
 
 def check_low_var(df, prescript="", postscript=""):
@@ -938,25 +938,27 @@ def check_complete_separation(df, events, durations, event_col):
 
 
 def check_nans_or_infs(df_or_array):
-    nulls = pd.isnull(df_or_array)
-    if hasattr(nulls, "values"):
-        if nulls.values.any():
-            raise TypeError("NaNs were detected in the dataset. Try using pd.isnull to find the problematic values.")
-    else:
-        if nulls.any():
-            raise TypeError("NaNs were detected in the dataset. Try using pd.isnull to find the problematic values.")
-    # isinf check is done after isnull check since np.isinf doesn't work on None values
     if isinstance(df_or_array, (pd.Series, pd.DataFrame)):
-        infs = df_or_array.values == np.Inf
-    else:
-        infs = np.isinf(df_or_array)
+        return check_nans_or_infs(df_or_array.values)
 
-    if hasattr(infs, "values"):
-        if infs.values.any():
-            raise TypeError("Infs were detected in the dataset. Try using np.isinf to find the problematic values.")
-    else:
-        if infs.any():
-            raise TypeError("Infs were detected in the dataset. Try using np.isinf to find the problematic values.")
+    if pd.isnull(df_or_array).any():
+        raise TypeError("NaNs were detected in the dataset. Try using pd.isnull to find the problematic values.")
+
+    try:
+        infs = np.isinf(df_or_array)
+    except TypeError:
+        warning_text = (
+            """Attempting to convert an unexpected datatype '%s' to float. Suggestion: 1) use `lifelines.utils.datetime_to_durations` to do conversions or 2) manually convert to floats/booleans."""
+            % df_or_array.dtype
+        )
+        warnings.warn(warning_text)
+        try:
+            infs = np.isinf(df_or_array.astype(float))
+        except:
+            raise TypeError("Wrong dtype '%s'." % df_or_array.dtype)
+
+    if infs.any():
+        raise TypeError("Infs were detected in the dataset. Try using np.isinf to find the problematic values.")
 
 
 def to_episodic_format(df, duration_col, event_col, id_col=None, time_gaps=1):
