+++ lifelines/fitters/__init__.py
@@ -215,6 +215,10 @@ class UnivariateFitter(BaseFitter):
         raise NotImplementedError
 
     @_must_call_fit_first
+    def cumulative_density_at_times(self, times, label=None):
+        raise NotImplementedError
+
+    @_must_call_fit_first
     def plot_cumulative_hazard(self, **kwargs):
         raise NotImplementedError()
 
@@ -226,6 +230,10 @@ class UnivariateFitter(BaseFitter):
     def plot_hazard(self, **kwargs):
         raise NotImplementedError()
 
+    @_must_call_fit_first
+    def plot_cumulative_density(self, **kwargs):
+        raise NotImplementedError()
+
 
 class ParametericUnivariateFitter(UnivariateFitter):
     """
@@ -234,7 +242,7 @@ class ParametericUnivariateFitter(UnivariateFitter):
     """
 
     _KNOWN_MODEL = False
-    _MIN_PARAMETER_VALUE = 1e-09
+    _MIN_PARAMETER_VALUE = 1e-9
 
     def __init__(self, *args, **kwargs):
         super(ParametericUnivariateFitter, self).__init__(*args, **kwargs)
@@ -324,12 +332,31 @@ class ParametericUnivariateFitter(UnivariateFitter):
     def _survival_function(self, params, times):
         return anp.exp(-self._cumulative_hazard(params, times))
 
+    def _cumulative_density(self, params, times):
+        return 1 - anp.exp(-self._cumulative_hazard(params, times))
+
     def _log_hazard(self, params, times):
         hz = self._hazard(params, times)
         hz = anp.clip(hz, 1e-18, np.inf)
         return anp.log(hz)
 
-    def _negative_log_likelihood(self, params, T, E, entry):
+    def _log_1m_sf(self, params, times):
+        # equal to log(cdf), but often easier to express with sf.
+        cum_haz = self._cumulative_hazard(params, times)
+        return anp.log1p(-anp.exp(-cum_haz))
+
+    def _negative_log_likelihood_left_censoring(self, params, T, E, _entry):
+
+        n = T.shape[0]
+        log_hz = self._log_hazard(params, T)
+        cum_haz = self._cumulative_hazard(params, T)
+        log_1m_sf = self._log_1m_sf(params, T)
+
+        ll = (E * (log_hz - cum_haz - log_1m_sf)).sum() + log_1m_sf.sum()
+
+        return -ll / n
+
+    def _negative_log_likelihood_right_censoring(self, params, T, E, entry):
 
         n = T.shape[0]
         log_hz = self._log_hazard(params, T[E])
@@ -386,11 +413,17 @@ class ParametericUnivariateFitter(UnivariateFitter):
     def _fit_model(self, T, E, entry, show_progress=True):
 
         non_zero_entries = entry[entry > 0]
+
+        if self.left_censorship:
+            negative_log_likelihood = self._negative_log_likelihood_left_censoring
+        else:
+            negative_log_likelihood = self._negative_log_likelihood_right_censoring
+
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
 
             results = minimize(
-                value_and_grad(self._negative_log_likelihood),  # pylint: disable=no-value-for-parameter
+                value_and_grad(negative_log_likelihood),  # pylint: disable=no-value-for-parameter
                 self._initial_values,
                 jac=True,
                 method="L-BFGS-B",
@@ -401,8 +434,10 @@ class ParametericUnivariateFitter(UnivariateFitter):
 
             if results.success:
                 # pylint: disable=no-value-for-parameter
-                hessian_ = hessian(self._negative_log_likelihood)(results.x, T, E, non_zero_entries)
[…]
