+++ lifelines/fitters/kaplan_meier_fitter.py
@@ -107,6 +107,8 @@ class KaplanMeierFitter(UnivariateFitter):
         if event_observed is not None:
             self._check_values(event_observed)
 
+        self.left_censorship = left_censorship
+
         if weights is not None:
             if (weights.astype(int) != weights).any():
                 warnings.warn(
@@ -155,7 +157,7 @@ class KaplanMeierFitter(UnivariateFitter):
         self._predict_label = label
         self._update_docstrings()
 
-        setattr(self, "plot_" + estimate_name, self.plot)
+        setattr(self, "plot_" + estimate_name.rstrip("_"), self.plot)
         return self
 
     def _check_values(self, array):
@@ -166,12 +168,6 @@ class KaplanMeierFitter(UnivariateFitter):
         Plot :math:`\log(S(t))` against :math:`\log(t)`
         """
         return plot_loglogs(self, *args, **kwargs)
-
-    def plot_survival_function(self, **kwargs):
-        """
-        Alias of ``plot``
-        """
-        return self.plot(**kwargs)
 
     def survival_function_at_times(self, times, label=None):
         """

[…]
