+++ lifelines/fitters/kaplan_meier_fitter.py
@@ -15,7 +15,7 @@ from lifelines.utils import (
     StatisticalWarning,
     coalesce,
 )
-from lifelines.plotting import plot_loglogs
+from lifelines.plotting import plot_loglogs, _plot_estimate
 
 
 class KaplanMeierFitter(UnivariateFitter):
@@ -108,6 +108,7 @@ class KaplanMeierFitter(UnivariateFitter):
             self._check_values(event_observed)
 
         self.left_censorship = left_censorship
+        self._label = label
 
         if weights is not None:
             if (weights.astype(int) != weights).any():
@@ -121,13 +122,15 @@ class KaplanMeierFitter(UnivariateFitter):
                 )
 
         # if the user is interested in left-censorship, we return the cumulative_density_, no survival_function_,
-        estimate_name = "survival_function_" if not left_censorship else "cumulative_density_"
-        v = _preprocess_inputs(durations, event_observed, timeline, entry, weights)
-        self.durations, self.event_observed, self.timeline, self.entry, self.event_table = v
+        primary_estimate_name = "survival_function_" if not self.left_censorship else "cumulative_density_"
+        secondary_estimate_name = "cumulative_density_" if not self.left_censorship else "survival_function_"
 
-        self._label = label
+        self.durations, self.event_observed, self.timeline, self.entry, self.event_table = _preprocess_inputs(
+            durations, event_observed, timeline, entry, weights
+        )
+
         alpha = alpha if alpha else self.alpha
-        log_survival_function, cumulative_sq_ = _additive_estimate(
+        log_estimate, cumulative_sq_ = _additive_estimate(
             self.event_table, self.timeline, self._additive_f, self._additive_var, left_censorship
         )
 
@@ -145,19 +148,23 @@ class KaplanMeierFitter(UnivariateFitter):
                 )
 
         # estimation
-        setattr(self, estimate_name, pd.DataFrame(np.exp(log_survival_function), columns=[self._label]))
-        self.__estimate = getattr(self, estimate_name)
+        setattr(self, primary_estimate_name, pd.DataFrame(np.exp(log_estimate), columns=[self._label]))
+        setattr(self, secondary_estimate_name, pd.DataFrame(1 - np.exp(log_estimate), columns=[self._label]))
+
+        self.__estimate = getattr(self, primary_estimate_name)
         self.confidence_interval_ = self._bounds(cumulative_sq_[:, None], alpha, ci_labels)
         self.median_ = median_survival_times(self.__estimate, left_censorship=left_censorship)
         self._cumulative_sq_ = cumulative_sq_
 
+        setattr(self, "confidence_interval_" + primary_estimate_name, self.confidence_interval_)
+        setattr(self, "confidence_interval_" + secondary_estimate_name, 1 - self.confidence_interval_)
+
         # estimation methods
-        self._estimation_method = estimate_name
-        self._estimate_name = estimate_name
+        self._estimation_method = primary_estimate_name
+        self._estimate_name = primary_estimate_name
         self._predict_label = label
         self._update_docstrings()
 
-        setattr(self, "plot_" + estimate_name.rstrip("_"), self.plot)
         return self
 
     def _check_values(self, array):
@@ -185,6 +192,22 @@ class KaplanMeierFitter(UnivariateFitter):
         label = coalesce(label, self._label)
         return pd.Series(self.predict(times), index=_to_array(times), name=label)
 
+    def plot_survival_function(self, **kwargs):
+        return _plot_estimate(
+            self,
+            estimate=self.survival_function_,
+            confidence_intervals=self.confidence_interval_survival_function_,
+            **kwargs
+        )
+
+    def plot_cumulative_density(self, **kwargs):
[…]
