+++ lifelines/fitters/log_normal_fitter.py
@@ -36,6 +36,10 @@ class LogNormalFitter(KnownModelParametericUnivariateFitter):
         The estimated survival function (with custom timeline if provided)
     confidence_interval_survival_function_ : DataFrame
         The lower and upper confidence intervals for the survival function
+    cumumlative_density_ : DataFrame
+        The estimated cumulative density function (with custom timeline if provided)
+    confidence_interval_cumumlative_density_ : DataFrame
+        The lower and upper confidence intervals for the cumulative density
     variance_matrix_ : numpy array
         The variance matrix of the coefficients
     median_: float
