+++ lifelines/plotting.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import warnings
+from textwrap import dedent
 import numpy as np
 from lifelines.utils import coalesce
 from scipy import stats
@@ -484,7 +485,7 @@ def _plot_estimate(
              .plot(iloc=slice(0,10))
           will plot the first 10 time points.
     invert_y_axis: bool
-        boolean to invert the y-axis, useful to show cumulative graphs instead of survival graphs.
+        boolean to invert the y-axis, useful to show cumulative graphs instead of survival graphs. (Deprecated, use ``plot_cumulative_density()``)
 
     Returns
     -------
@@ -494,6 +495,21 @@ def _plot_estimate(
     plot_estimate_config = PlotEstimateConfig(
         cls, estimate, confidence_intervals, loc, iloc, show_censors, censor_styles, **kwargs
     )
+
+    if invert_y_axis:
+        warnings.warn(
+            dedent(
+                """
+            The invert_y_axis will be removed in lifelines 0.21.0. Likely you are trying to plot the cumulative density function?
+            That's now part of the KaplanMeierFitter,
+
+            >>> kmf.plot_cumulative_density()
+            >>> # nice
+
+        """
+            ),
+            PendingDeprecationWarning,
+        )
 
     dataframe_slicer = create_dataframe_slicer(iloc, loc)
 
