### 2019-03-23 2f5ab3c
docs on plotting
