+++ lifelines/fitters/__init__.py
@@ -93,6 +93,47 @@ class UnivariateFitter(BaseFitter):
 
     @_must_call_fit_first
     def plot(self, **kwargs):
+        """
+        Plots a pretty figure of the model
+
+        Matplotlib plot arguments can be passed in inside the kwargs, plus
+
+        Parameters
+        -----------
+        show_censors: bool
+            place markers at censorship events. Default: False
+        censor_styles: bool
+            If show_censors, this dictionary will be passed into the plot call.
+        ci_alpha: bool
+            the transparency level of the confidence interval. Default: 0.3
+        ci_force_lines: bool
+            force the confidence intervals to be line plots (versus default shaded areas). Default: False
+        ci_show: bool
+            show confidence intervals. Default: True
+        ci_legend: bool
+            if ci_force_lines is True, this is a boolean flag to add the lines' labels to the legend. Default: False
+        at_risk_counts: bool
+            show group sizes at time points. See function ``add_at_risk_counts`` for details. Default: False
+        loc: slice
+            specify a time-based subsection of the curves to plot, ex:
+
+            >>> model.plot(loc=slice(0.,10.))
+
+            will plot the time values between t=0. and t=10.
+        iloc: slice
+            specify a location-based subsection of the curves to plot, ex:
+
+            >>> model.plot(iloc=slice(0,10))
+
+            will plot the first 10 time points.
+        invert_y_axis: bool
+            boolean to invert the y-axis, useful to show cumulative graphs instead of survival graphs. (Deprecated, use ``plot_cumulative_density()``)
+
+        Returns
+        -------
+        ax:
+            a pyplot axis object
+        """
         return _plot_estimate(
             self, estimate=getattr(self, self._estimate_name), confidence_intervals=self.confidence_interval_, **kwargs
         )
