### 2019-03-25 c3bbf8e
Add variance formula
