### 2019-03-26 3d79464
some doc changes, requirements in setup.py changes, closing #679
