### 2019-03-26 929660f
some doc changes, requirements in setup.py changes, closing #679
