+++ lifelines/fitters/piecewise_exponential_fitter.py
@@ -77,7 +77,6 @@ class PiecewiseExponentialFitter(KnownModelParametericUnivariateFitter):
             raise ValueError("Do not add inf to the breakpoints.")
 
         if breakpoints and breakpoints[0] <= 0:
-            raise ValueError("First breakpoint must be greater than 0.")
 
         self.breakpoints = np.append(breakpoints, [np.inf])
         n_breakpoints = len(self.breakpoints)
