+++ lifelines/fitters/piecewise_exponential_regression_fitter.py
@@ -1,0 +1,845 @@
+# -*- coding: utf-8 -*-
+from textwrap import dedent
+from autograd import numpy as np
+from scipy.special import gamma
+import pandas as pd
+from autograd import hessian, value_and_grad, elementwise_grad as egrad, grad
+from datetime import datetime
+from lifelines.utils import _get_index, coalesce, qth_survival_times
+from lifelines.fitters import BaseFitter
+
+import warnings
+from lifelines.utils import (
+    qth_survival_times,
+    _to_array,
+    _to_list,
+    dataframe_interpolate_at_times,
+    ConvergenceError,
+    inv_normal_cdf,
+    string_justify,
+    format_floats,
+    format_p_value,
+    format_exp_floats,
+    coalesce,
+    check_nans_or_infs,
+    pass_for_numeric_dtypes_or_raise_array,
+    check_for_numeric_dtypes_or_raise,
+    check_complete_separation,
+    check_low_var,
+    StatisticalWarning,
+    StatError,
+    median_survival_times,
+    normalize,
+    concordance_index,
+)
+from autograd.differential_operators import make_jvp_reversemode
+from scipy.optimize import minimize
+from scipy import stats
+import pandas as pd
+from numpy.linalg import inv, pinv
+
+
+class PiecewiseExponentialRegressionFitter(BaseFitter):
+    r"""
+    This class implements an Piecewise Exponential model for univariate data. The model has parameterized
+    hazard rate:
+
+    .. math::  h(t) = \begin{cases}
+                        1/\lambda_0,  & \text{if $t \le \tau_0$} \\
+                        1/\lambda_1 & \text{if $\tau_0 < t \le \tau_1$} \\
+                        1/\lambda_2 & \text{if $\tau_1 < t \le \tau_2$} \\
+                        ...
+                      \end{cases}
+
+    You specify the breakpoints, :math:`\tau_i`, and *lifelines* will find the
+    optional values for the parameters.
+
+    After calling the `.fit` method, you have access to properties like: ``survival_function_``, ``plot``, ``cumulative_hazard_``
+    A summary of the fit is available with the method ``print_summary()``
+
+    Important
+    ----------
+    The parameterization of this model changed in lifelines 0.19.1. Previously, the cumulative hazard looked like
+    :math:`\lambda_i t`. The parameterization is now the reciprocal of :math:`\lambda_i`.
+
+    Attributes
+    ----------
+    cumulative_hazard_ : DataFrame
+        The estimated cumulative hazard (with custom timeline if provided)
+    confidence_interval_cumulative_hazard_ : DataFrame
+        The lower and upper confidence intervals for the cumulative hazard
+    hazard_ : DataFrame
+        The estimated hazard (with custom timeline if provided)
+    confidence_interval_hazard_ : DataFrame
+        The lower and upper confidence intervals for the hazard
+    survival_function_ : DataFrame
+        The estimated survival function (with custom timeline if provided)
+    confidence_interval_survival_function_ : DataFrame
+        The lower and upper confidence intervals for the survival function
+    cumumlative_density_ : DataFrame
+        The estimated cumulative density function (with custom timeline if provided)
+    confidence_interval_cumumlative_density_ : DataFrame
+        The lower and upper confidence intervals for the cumulative density
+    variance_matrix_ : numpy array
+        The variance matrix of the coefficients
+    median_: float
+        The median time to event
+    lambda_i_: float
+        The fitted parameter in the model, for i = 0, 1 ... n-1 breakpoints
+    durations: array
+        The durations provided
+    event_observed: array
+        The event_observed variable provided
+    timeline: array
+        The time line to use for plotting and indexing
+    entry: array or None
[…]
