+++ lifelines/fitters/piecewise_exponential_regression_fitter.py
@@ -1,14 +1,15 @@
 # -*- coding: utf-8 -*-
 from textwrap import dedent
+import warnings
+from datetime import datetime
 from autograd import numpy as np
 from scipy.special import gamma
 import pandas as pd
 from autograd import hessian, value_and_grad, elementwise_grad as egrad, grad
-from datetime import datetime
 from lifelines.utils import _get_index, coalesce, qth_survival_times
 from lifelines.fitters import BaseFitter
+from lifelines.plotting import _plot_estimate, set_kwargs_drawstyle, set_kwargs_ax
 
-import warnings
 from lifelines.utils import (
     qth_survival_times,
     _to_array,
@@ -44,23 +45,19 @@ class PiecewiseExponentialRegressionFitter(BaseFitter):
     This class implements an Piecewise Exponential model for univariate data. The model has parameterized
     hazard rate:
 
-    .. math::  h(t) = \begin{cases}
-                        1/\lambda_0,  & \text{if $t \le \tau_0$} \\
-                        1/\lambda_1 & \text{if $\tau_0 < t \le \tau_1$} \\
-                        1/\lambda_2 & \text{if $\tau_1 < t \le \tau_2$} \\
+    .. math::  h(t\;|\;x) = \begin{cases}
+                        1/\lambda_0(x),  & \text{if $t \le \tau_0$} \\
+                        1/\lambda_1(x) & \text{if $\tau_0 < t \le \tau_1$} \\
+                        1/\lambda_2(x) & \text{if $\tau_1 < t \le \tau_2$} \\
                         ...
                       \end{cases}
 
     You specify the breakpoints, :math:`\tau_i`, and *lifelines* will find the
     optional values for the parameters.
 
-    After calling the `.fit` method, you have access to properties like: ``survival_function_``, ``plot``, ``cumulative_hazard_``
+    After calling the `.fit` method, you have access to properties like: ``params_``
     A summary of the fit is available with the method ``print_summary()``
 
-    Important
-    ----------
-    The parameterization of this model changed in lifelines 0.19.1. Previously, the cumulative hazard looked like
-    :math:`\lambda_i t`. The parameterization is now the reciprocal of :math:`\lambda_i`.
 
     Attributes
     ----------
@@ -124,9 +121,9 @@ class PiecewiseExponentialRegressionFitter(BaseFitter):
         M = np.minimum(np.tile(bp, (n, 1)), T)
         M = np.hstack([M[:, tuple([0])], np.diff(M, axis=1)])
         lambdas_ = np.array(
-            [np.exp(np.dot(X, params[self._LOOKUP_SLICE["lambda_%d_" % i]])) for i in range(self.n_breakpoints)]
+            [np.exp(-np.dot(X, params[self._LOOKUP_SLICE["lambda_%d_" % i]])) for i in range(self.n_breakpoints)]
         )
-        return M * (1 / lambdas_.T)
+        return M * lambdas_.T
 
     def _log_hazard(self, params, T, X):
         # can be overwritten to improve convergence, see WeibullAFTFitter
@@ -327,7 +324,7 @@ class PiecewiseExponentialRegressionFitter(BaseFitter):
     def _fit_model(self, T, E, weights, X, show_progress=False, initial_point=None):
 
         if initial_point is None:
-            initial_point = np.ones((X.shape[1] * self.n_breakpoints))
+            initial_point = np.zeros((X.shape[1] * self.n_breakpoints))
 
         results = minimize(
             # using value_and_grad is much faster (takes advantage of shared computations) than spitting.
@@ -427,16 +424,16 @@ class PiecewiseExponentialRegressionFitter(BaseFitter):
 
     @property
     def _ll_null(self):
[…]
