+++ experiments/piecewise_exponential_regression_example.py
@@ -9,24 +9,24 @@ from lifelines.datasets import load_regression_dataset
 from lifelines.generate_datasets import piecewise_exponential_survival_data
 
 
-N, d = 2000, 1
+N, d = 2500, 2
 
-breakpoints = (1, 31, 34, 62, 65, 93, 96)  # initial purchase  # second bill  # third bill  # fourth bill
+breakpoints = (1, 31, 34, 62, 65, 93, 96)
 
 betas = np.array(
     [
-        [-1.25, np.log(15)],
-        [-2.25, np.log(333)],
-        [-1.1, np.log(18)],
-        [-2.1, np.log(500)],
-        [-1.0, np.log(20)],
-        [-1.8, np.log(500)],
-        [-0.5, np.log(20)],
-        [-1.5, np.log(250)],
+        [-1.25, 5.0, np.log(15)],
+        [-2.25, 6.0, np.log(333)],
+        [-1.1, 7.0, np.log(18)],
+        [-2.1, 8.0, np.log(500)],
+        [-1.0, 9.0, np.log(20)],
+        [-1.8, 10.0, np.log(500)],
+        [-0.5, 11.0, np.log(20)],
+        [-1.5, 12.0, np.log(250)],
     ]
 )
 
-X = 0.1 * np.random.randn(N, d)
+X = 0.1 * np.random.exponential(size=(N, d))
 X = np.c_[X, np.ones(N)]
 
 T = np.empty(N)
@@ -34,7 +34,7 @@ for i in range(N):
     lambdas = np.exp(-betas.dot(X[i, :]))
     T[i] = piecewise_exponential_survival_data(1, breakpoints, lambdas)[0]
 
-T_censor = np.minimum(0.9 * T.mean() * np.random.exponential(size=N), 110)
+T_censor = np.minimum(T.mean() * np.random.exponential(size=N), 110)
 
 df = pd.DataFrame(X)
 df["T"] = np.minimum(T, T_censor)
@@ -46,4 +46,34 @@ pew = PiecewiseExponentialRegressionFitter(breakpoints=breakpoints, penalizer=0.
 )
 pew.print_summary()
 
-kmf = KaplanMeierFitter().fit(df["T"], df["E"])
+"""
+ps = 10.0**np.arange(-5, 5)
+results0 = []
+results1 = []
+results2 = []
+results3 = []
+results4 = []
+results5 = []
+results6 = []
+
+
+for p in ps:
+    pew = PiecewiseExponentialRegressionFitter(breakpoints=breakpoints, penalizer=p, fit_intercept=False).fit(df, "T", "E")
+    results0.append(pew.summary.loc[('lambda_0_', 0), ['coef', 'lower 0.95', 'upper 0.95']])
+    results1.append(pew.summary.loc[('lambda_1_', 0), ['coef', 'lower 0.95', 'upper 0.95']])
+    results2.append(pew.summary.loc[('lambda_2_', 0), ['coef', 'lower 0.95', 'upper 0.95']])
+    results3.append(pew.summary.loc[('lambda_3_', 0), ['coef', 'lower 0.95', 'upper 0.95']])
+    results4.append(pew.summary.loc[('lambda_4_', 0), ['coef', 'lower 0.95', 'upper 0.95']])
+    results5.append(pew.summary.loc[('lambda_5_', 0), ['coef', 'lower 0.95', 'upper 0.95']])
+    results6.append(pew.summary.loc[('lambda_6_', 0), ['coef', 'lower 0.95', 'upper 0.95']])
+
+
+
+ax = pd.DataFrame(results0).assign(x=ps).plot('x', 'coef', logx=True)
[…]
