+++ lifelines/utils/__init__.py
@@ -1360,12 +1360,14 @@ def format_p_value(decimals):
     return lambda p: "<%s" % threshold if p < threshold else "{:4.{prec}f}".format(p, prec=decimals)
 
 
-def format_exp_floats():
+def format_exp_floats(decimals):
     """
     sometimes the exp. column can be too large
     """
     threshold = 10 ** 5
-    return lambda n: "{:.2e}".format(n) if n > threshold else str(n)
+    return (
+        lambda n: "{:.{prec}e}".format(n, prec=decimals) if n > threshold else "{:4.{prec}f}".format(n, prec=decimals)
+    )
 
 
 def format_floats(decimals):
