### 2019-03-27 c5a1069
adding left truncation support to AFT models
