+++ lifelines/fitters/__init__.py
@@ -41,6 +41,7 @@ from lifelines.utils import (
     check_for_numeric_dtypes_or_raise,
     check_complete_separation,
     check_low_var,
+    check_positivity,
     StatisticalWarning,
     StatError,
     median_survival_times,
@@ -655,10 +656,7 @@ class ParametericUnivariateFitter(UnivariateFitter):
             check_nans_or_infs(event_observed)
 
         self.durations = np.asarray(pass_for_numeric_dtypes_or_raise_array(durations))
-        if np.any(self.durations <= 0):
-            raise ValueError(
-                "This model does not allow for non-positive durations. Suggestion: add a small positive value to zero elements."
-            )
+        check_positivity(self.durations)
 
         if not self._KNOWN_MODEL:
             self._check_cumulative_hazard_is_monotone_and_positive(self.durations, self._initial_values)
@@ -898,9 +896,9 @@ class KnownModelParametericUnivariateFitter(ParametericUnivariateFitter):
     _KNOWN_MODEL = True
 
 
-class ParametericRegressionFitter(BaseFitter):
+class ParametericAFTRegressionFitter(BaseFitter):
     def __init__(self, alpha=0.05, penalizer=0.0, l1_ratio=0.0, fit_intercept=True):
-        super(ParametericRegressionFitter, self).__init__(alpha=alpha)
+        super(ParametericAFTRegressionFitter, self).__init__(alpha=alpha)
         self._hazard = egrad(self._cumulative_hazard, argnum=1)  # pylint: disable=unexpected-keyword-arg
         self.penalizer = penalizer
         self.l1_ratio = l1_ratio
@@ -913,10 +911,17 @@ class ParametericRegressionFitter(BaseFitter):
         hz = anp.clip(hz, 1e-20, np.inf)
         return anp.log(hz)
 
-    def _negative_log_likelihood(self, params, T, E, W, *Xs):
+    def _negative_log_likelihood(self, params, T, E, W, starts, *Xs):
         warnings.simplefilter(action="ignore", category=FutureWarning)
 
-        ll = (W * E * self._log_hazard(params, T, *Xs)).sum() - (W * self._cumulative_hazard(params, T, *Xs)).sum()
+        non_zero_starts = starts > 0
+
+        ll = (
+            (W * E * self._log_hazard(params, T, *Xs)).sum()
+            - (W * self._cumulative_hazard(params, T, *Xs)).sum()
+            + self._cumulative_hazard(params, starts[non_zero_starts], *(X[non_zero_starts, :] for X in Xs)).sum()
+        )
+
         if self.penalizer > 0:
             penalty = self.l1_ratio * anp.abs(params).sum() + 0.5 * (1.0 - self.l1_ratio) * (params ** 2).sum()
         else:
@@ -936,6 +941,7 @@ class ParametericRegressionFitter(BaseFitter):
         weights_col=None,
         robust=False,
         initial_point=None,
+        entry_col=None,
     ):
         """
         Fit the accelerated failure time model to a dataset.
@@ -980,6 +986,8 @@ class ParametericRegressionFitter(BaseFitter):
             initialize the starting point of the iterative
             algorithm. Default is the zero vector.
 
+        entry_col: TODO
+
         Returns
         -------
         self:
@@ -1015,6 +1023,7 @@ class ParametericRegressionFitter(BaseFitter):
         self.duration_col = duration_col
         self.event_col = event_col
         self.weights_col = weights_col
+        self.entry_col = entry_col
         self._n_examples = df.shape[0]
         self.timeline = timeline
         self.robust = robust
@@ -1032,43 +1041,34 @@ class ParametericRegressionFitter(BaseFitter):
             if (self.weights_col is not None)
[…]
