+++ perf_tests/weibull_aft_perf.py
@@ -5,16 +5,18 @@
 if __name__ == "__main__":
     import pandas as pd
     import time
+    import numpy as np
 
     from lifelines import WeibullAFTFitter
     from lifelines.datasets import load_rossi
 
     df = load_rossi()
     df = pd.concat([df] * 1)
+    df["start"] = np.random.randint(0, 3, size=df.shape[0])
     # df = df.reset_index()
     # df['week'] = np.random.exponential(1, size=df.shape[0])
     wp = WeibullAFTFitter()
     start_time = time.time()
-    wp.fit(df, duration_col="week", event_col="arrest")
+    wp.fit(df, duration_col="week", event_col="arrest", start_col="start")
     print("--- %s seconds ---" % (time.time() - start_time))
     wp.print_summary()
