### 2019-03-27 f9bfab0
v0.20.4 (#680)
* some doc changes, requirements in setup.py changes, closing #679

* fix formatting

* adding left truncation support to AFT models

* bump version
