+++ lifelines/fitters/coxph_fitter.py
@@ -38,6 +38,7 @@ from lifelines.utils import (
     string_justify,
     format_p_value,
     format_floats,
+    format_exp_floats,
     dataframe_interpolate_at_times,
 )
 
@@ -333,6 +334,23 @@ class CoxPHFitter(BaseFitter):
             else pd.Series(np.ones((self._n_examples,)), index=df.index, name="weights")
         )
 
+        _clusters = df.pop(self.cluster_col).values if self.cluster_col else None
+
+        X = df.astype(float)
+        T = T.astype(float)
+        E = E.astype(bool)
+
+        self._check_values(X, T, E, W)
+
+        return X, T, E, W, original_index, _clusters
+
+    def _check_values(self, X, T, E, W):
+        check_for_numeric_dtypes_or_raise(X)
+        check_nans_or_infs(T)
+        check_nans_or_infs(E)
+        check_nans_or_infs(X)
+        check_low_var(X)
+        check_complete_separation(X, E, T, self.event_col)
         # check to make sure their weights are okay
         if self.weights_col:
             if (W.astype(int) != W).any() and not self.robust:
@@ -345,25 +363,6 @@ estimate the variances. See paper "Variance estimation when using inverse probab
                 )
             if (W <= 0).any():
                 raise ValueError("values in weight column %s must be positive." % self.weights_col)
-
-        _clusters = df.pop(self.cluster_col).values if self.cluster_col else None
-
-        X = df.astype(float)
-        T = T.astype(float)
-        E = E.astype(bool)
-
-        self._check_values(X, T, E, self.event_col)
-
-        return X, T, E, W, original_index, _clusters
-
-    @staticmethod
-    def _check_values(X, T, E, event_col):
-        check_for_numeric_dtypes_or_raise(X)
-        check_nans_or_infs(T)
-        check_nans_or_infs(E)
-        check_nans_or_infs(X)
-        check_low_var(X)
-        check_complete_separation(X, E, T, event_col)
 
     def _newton_rhaphson(
         self,
@@ -1257,7 +1256,12 @@ See https://stats.stackexchange.com/questions/11109/how-to-deal-with-perfect-sep
 
         df = self.summary
         # Significance codes as last column
-        print(df.to_string(float_format=format_floats(decimals), formatters={"p": format_p_value(decimals)}))
+        print(
+            df.to_string(
+                float_format=format_floats(decimals),
+                formatters={"p": format_p_value(decimals), "exp(coef)": format_exp_floats(decimals)},
+            )
+        )
 
         # Significance code explanation
         print("---")
