### 2019-03-27 117942b
example + some improvements
