+++ lifelines/fitters/piecewise_exponential_regression_fitter.py
@@ -52,48 +52,50 @@ class PiecewiseExponentialRegressionFitter(BaseFitter):
                         ...
                       \end{cases}
 
-    You specify the breakpoints, :math:`\tau_i`, and *lifelines* will find the
-    optional values for the parameters.
+    and $\lambda_i(x) = \exp(\mathbf{\beta}_i x^T), \;\; \mathbf{\beta}_i = (\beta_{i,1}, \beta_{i,2}, ...)$. That is, each period has a hazard rate, $\lambda_i$ the is the exponential of a linear model. The parameters of each linear model are unique to that period - different periods have different parameters (later we will generalize this).
+
+    Why do I want a model like this? Well, it offers lots of flexibility (at the cost of efficiency though), but importantly I can see:
+
+    1. Influence of variables over time.
+    2. Looking at important variables at specific "drops" (or regime changes). For example, what variables cause the large drop at the start? What variables prevent death at the second billing?
+    3. Predictive power: since we model the hazard more accurately (we hope) than a simpler parametric form, we have better estimates of a subjects survival curve.
 
     After calling the `.fit` method, you have access to properties like: ``params_``
     A summary of the fit is available with the method ``print_summary()``
 
 
+    Parameters
+    -----------
+    breakpoints: list
+        a list of times when a new exponential model is constructed.
+
+    alpha: float, optional (default=0.05)
+        the level in the confidence intervals.
+
+    fit_intercept: boolean, optional (default=True)
+        Allow lifelines to add an intercept column of 1s to df, and ancillary_df if applicable.
+
+    penalizer: float, optional (default=0.0)
+        the penalizer coefficient to the size of the coefficients. See `l1_ratio`. Must be equal to or greater than 0.
+
     Attributes
     ----------
-    cumulative_hazard_ : DataFrame
-        The estimated cumulative hazard (with custom timeline if provided)
-    confidence_interval_cumulative_hazard_ : DataFrame
-        The lower and upper confidence intervals for the cumulative hazard
-    hazard_ : DataFrame
-        The estimated hazard (with custom timeline if provided)
-    confidence_interval_hazard_ : DataFrame
-        The lower and upper confidence intervals for the hazard
-    survival_function_ : DataFrame
-        The estimated survival function (with custom timeline if provided)
-    confidence_interval_survival_function_ : DataFrame
-        The lower and upper confidence intervals for the survival function
-    cumumlative_density_ : DataFrame
-        The estimated cumulative density function (with custom timeline if provided)
-    confidence_interval_cumumlative_density_ : DataFrame
-        The lower and upper confidence intervals for the cumulative density
+    params_ : DataFrame
+        The estimated coefficients
+    confidence_intervals_ : DataFrame
+        The lower and upper confidence intervals for the coefficients
+    durations: Series
+        The event_observed variable provided
+    event_observed: Series
+        The event_observed variable provided
+    weights: Series
+        The event_observed variable provided
     variance_matrix_ : numpy array
         The variance matrix of the coefficients
-    median_: float
-        The median time to event
-    lambda_i_: float
-        The fitted parameter in the model, for i = 0, 1 ... n-1 breakpoints
-    durations: array
-        The durations provided
-    event_observed: array
-        The event_observed variable provided
-    timeline: array
-        The time line to use for plotting and indexing
-    entry: array or None
-        The entry array provided, or None
-    breakpoints: array
-        The provided breakpoints
-
+    standard_errors_: Series
[…]
