+++ perf_tests/weibull_aft_perf.py
@@ -11,12 +11,11 @@ if __name__ == "__main__":
     from lifelines.datasets import load_rossi
 
     df = load_rossi()
-    df = pd.concat([df] * 1)
-    df["start"] = np.random.randint(0, 3, size=df.shape[0])
+    df = pd.concat([df] * 20)
     # df = df.reset_index()
     # df['week'] = np.random.exponential(1, size=df.shape[0])
     wp = WeibullAFTFitter()
     start_time = time.time()
-    wp.fit(df, duration_col="week", event_col="arrest", start_col="start")
+    wp.fit(df, duration_col="week", event_col="arrest")
     print("--- %s seconds ---" % (time.time() - start_time))
     wp.print_summary()
