+++ lifelines/fitters/coxph_fitter.py
@@ -1809,7 +1809,7 @@ the following on the original dataset, df: `df.groupby(%s).size()`. Expected is 
 
         for variable in self.hazards_.index:
             minumum_observed_p_value = test_results.summary.loc[variable, "p"].min()
-            if minumum_observed_p_value > p_value_threshold:
+            if np.round(minumum_observed_p_value, 2) > p_value_threshold:
                 continue
 
             counter += 1
@@ -1847,7 +1847,7 @@ the following on the original dataset, df: `df.groupby(%s).size()`. Expected is 
 
                 # Arbitrary chosen 10 and 4 to check for ability to use strata col.
                 # This should capture dichotomous / low cardinality values.
-                if n_uniques <= 10 and value_counts.min() >= 4:
+                if n_uniques <= 10 and value_counts.min() >= 5:
                     print(
                         fill(
                             "   Advice: with so few unique values (only {0}), you can include `strata=['{1}', ...]` in the call in `.fit`. See documentation in link [B] below.".format(
@@ -1859,7 +1859,7 @@ the following on the original dataset, df: `df.groupby(%s).size()`. Expected is 
                 else:
                     print(
                         fill(
-                            """   Advice: try binning the variable '{var}' using pd.cut, and then specify it in `strata=['{var}', ...]` in the call in `.fit`. See documentation in link [B] below.""".format(
+                            """   Advice 1: the functional form of the variable '{var}' might be incorrect. That is, there may be non-linear terms missing. The proportional hazard test used is very sensitive to incorrect functional forms. See documentation in link [D] below on how to specify a functional form.""".format(
                                 var=variable
                             ),
                             width=100,
@@ -1867,7 +1867,15 @@ the following on the original dataset, df: `df.groupby(%s).size()`. Expected is 
                     )
                     print(
                         fill(
-                            """   Alternative Advice: try adding an interaction term with your time variable. See documentation in link [A] and specifically link [C] below.""".format(
+                            """   Advice 2: try binning the variable '{var}' using pd.cut, and then specify it in `strata=['{var}', ...]` in the call in `.fit`. See documentation in link [B] below.""".format(
+                                var=variable
+                            ),
+                            width=100,
+                        )
+                    )
+                    print(
+                        fill(
+                            """   Advice 3: try adding an interaction term with your time variable. See documentation in link [A] and specifically link [C] below.""".format(
                                 var=variable
                             ),
                             width=100,
