### 2019-04-03 b92122b
Some documentation updates (#691)
* docs

* more docs

* ipynb links

* fix docs

* change rst changelog too
