+++ CHANGELOG.md
@@ -15,14 +15,14 @@
  - `PiecewiseExponentialFitter` is available with `from lifelines import *`.
 
 
-#### 0.20.3
+#### 0.20.3 - 2019-03-23
 
 ##### New features
  - Now `cumulative_density_` & `survival_function_` are _always_ present on a fitted `KaplanMeierFitter`.
  - New attributes/methods on `KaplanMeierFitter`: `plot_cumulative_density()`, `confidence_interval_cumulative_density_`, `plot_survival_function` and `confidence_interval_survival_function_`.
 
 
-#### 0.20.2
+#### 0.20.2 - 2019-03-21
 
 ##### New features
  - Left censoring is now supported in univariate parametric models: `.fit(..., left_censorship=True)`. Examples are in the docs.
@@ -40,7 +40,8 @@
  - `qth_survival_times` for a truncated CDF would return `np.inf` if the q parameter was below the truncation limit. This should have been `-np.inf`
 
 
-#### 0.20.1
+#### 0.20.1 - 2019-03-16
+
  - Some performance improvements to `CoxPHFitter` (about 30%). I know it may seem silly, but we are now about the same or slighty faster than the Cox model in R's `survival` package (for some testing datasets and some configurations). This is a big deal, because 1) lifelines does more error checking prior, 2) R's cox model is written in C, and we are still pure Python/NumPy, 3) R's cox model has decades of development.
  - suppressed unimportant warnings
 
@@ -48,7 +49,8 @@
  - Previously, lifelines _always_ added a 0 row to `cph.baseline_hazard_`, even if there were no event at this time. This is no longer the case. A 0 will still be added if there is a duration (observed or not) at 0 occurs however.
 
 
-#### 0.20.0
+#### 0.20.0 - 2019-03-05
+
  - Starting with 0.20.0, only Python3 will be supported. Over 75% of recent installs where Py3.
  - Updated minimum dependencies, specifically Matplotlib and Pandas.
 
@@ -65,18 +67,18 @@
 
 
 
-#### 0.19.5
+#### 0.19.5 - 2019-02-26
 
 ##### New features
  -  `plot_covariate_group` can accept multiple covariates to plot. This is useful for columns that have implicit correlation like polynomial features or categorical variables.
  - Convergence improvements for AFT models.
 
-#### 0.19.4
+#### 0.19.4 - 2019-02-25
 
 ##### Bug fixes
  - remove some bad print statements in `CoxPHFitter`.
 
-#### 0.19.3
+#### 0.19.3 - 2019-02-25
 
 ##### New features
  - new AFT models: `LogNormalAFTFitter` and `LogLogisticAFTFitter`.
@@ -84,7 +86,7 @@
  - Robust errors (sandwich errors) are now avilable in AFT models using the `robust=True` kwarg in `fit`.
  - Performance increase to `print_summary` in the `CoxPHFitter` and `CoxTimeVaryingFitter` model.
 
-#### 0.19.2
+#### 0.19.2 - 2019-02-22
 
 ##### New features
  - `ParametricUnivariateFitters`, like `WeibullFitter`, have smoothed plots when plotting (vs stepped plots)
@@ -93,7 +95,7 @@
  - The `ExponentialFitter` log likelihood _value_ was incorrect - inference was correct however.
  - Univariate fitters are more flexiable and can allow 2-d and DataFrames as inputs.
 
-#### 0.19.1
+#### 0.19.1 - 2019-02-21
 
 ##### New features
  - improved stability of `LogNormalFitter`
@@ -103,7 +105,7 @@
[…]

[…]
