### 2019-04-07 eb89b39
updates to AJF model, including fixing #688
