+++ lifelines/fitters/aalen_johansen_fitter.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from textwrap import dedent
 import numpy as np
 import pandas as pd
 import warnings
@@ -103,8 +104,10 @@ class AalenJohansenFitter(UnivariateFitter):
 
         if ties:
             warnings.warn(
-                """Tied event times were detected. The Aalen-Johansen estimator cannot handle tied event times.
-                To resolve ties, data is randomly jittered.""",
+                dedent(
+                    """Tied event times were detected. The Aalen-Johansen estimator cannot handle tied event times.
+                To resolve ties, data is randomly jittered."""
+                ),
                 Warning,
             )
             durations = self._jitter(
@@ -114,28 +117,32 @@ class AalenJohansenFitter(UnivariateFitter):
                 seed=self._seed,
             )
 
+        alpha = alpha if alpha else self.alpha
+
         # Creating label for event of interest & indicator for that event
-        cmprisk_label = "CIF_" + str(int(event_of_interest))
-        self.label_cmprisk = "observed_" + str(int(event_of_interest))
+        event_of_interest = int(event_of_interest)
+        cmprisk_label = "CIF_" + str(event_of_interest)
+        self.label_cmprisk = "observed_" + str(event_of_interest)
 
         # Fitting Kaplan-Meier for either event of interest OR competing risk
-        km = KaplanMeierFitter()
-        km.fit(durations, event_observed=event_observed, timeline=timeline, entry=entry, weights=weights)
+        km = KaplanMeierFitter().fit(
+            durations, event_observed=event_observed, timeline=timeline, entry=entry, weights=weights
+        )
         aj = km.event_table
         aj["overall_survival"] = km.survival_function_
         aj["lagged_overall_survival"] = aj["overall_survival"].shift()
 
         # Setting up table for calculations and to return to user
-        event_spec = np.where(pd.Series(event_observed) == event_of_interest, 1, 0)
-        event_spec_proc = _preprocess_inputs(
+        event_spec = pd.Series(event_observed) == event_of_interest
+        self.durations, self.event_observed, *_, event_table = _preprocess_inputs(
             durations=durations, event_observed=event_spec, timeline=timeline, entry=entry, weights=weights
         )
-        event_spec_times = event_spec_proc[-1]["observed"]
+        event_spec_times = event_table["observed"]
         event_spec_times = event_spec_times.rename(self.label_cmprisk)
         aj = pd.concat([aj, event_spec_times], axis=1).reset_index()
 
         # Estimator of Cumulative Incidence (Density) Function
-        aj[cmprisk_label] = ((aj[self.label_cmprisk]) / (aj["at_risk"]) * aj["lagged_overall_survival"]).cumsum()
+        aj[cmprisk_label] = (aj[self.label_cmprisk] / aj["at_risk"] * aj["lagged_overall_survival"]).cumsum()
         aj.loc[0, cmprisk_label] = 0  # Setting initial CIF to be zero
         aj = aj.set_index("event_at")
 
@@ -145,35 +152,38 @@ class AalenJohansenFitter(UnivariateFitter):
         self._predict_label = label
         self._update_docstrings()
 
-        alpha = alpha if alpha else self.alpha
         self._label = label
         self.cumulative_density_ = pd.DataFrame(aj[cmprisk_label])
+
         # Technically, cumulative incidence, but consistent with KaplanMeierFitter
         self.event_table = aj[
             ["removed", "observed", self.label_cmprisk, "censored", "entrance", "at_risk"]
         ]  # Event table
+
         if self._calc_var:
-            self.variance, self.confidence_interval_ = self._bounds(
[…]
