### 2019-04-07 f882ad9
fixing loglikelihood ratio test #687 and some good perf improvements there too
