### 2019-04-08 3e63f74
fixing loglikelihood ratio test #687 and some good perf improvements there too (#697)
