### 2019-04-08 085f744
this function wasn't showing up in docs, and I've made modifications to return a weight array too
