### 2019-04-08 0f2c437
this function wasn't showing up in docs, and I've made modifications … (#699)
* this function wasn't showing up in docs, and I've made modifications to return a weight array too

* improve docs
