### 2019-04-08 e62b889
fix error handling in Piecewise Exp model and lint
