+++ docs/Changelog.rst
@@ -1,8 +1,35 @@
 Changelog
 ~~~~~~~~~
 
-0.20.4
-^^^^^^
+0.20.5 - 2019-04-08
+^^^^^^^^^^^^^^^^^^^
+
+New features
+''''''''''''
+
+-  performance improvements for ``print_summary``.
+
+API changes
+'''''''''''
+
+-  ``utils.survival_events_from_table`` returns an integer weight vector
+   as well as durations and censoring vector.
+-  in ``AalenJohansenFitter``, the ``variance`` parameter is renamed to
+   ``variance_`` to align with the usual lifelines convention.
+
+Bug fixes
+'''''''''
+
+-  Fixed an error in the ``CoxTimeVaryingFitter``\ ’s likelihood ratio
+   test when using strata.
+-  Fixed some plotting bugs with ``AalenJohansenFitter``
+
+.. _section-1:
+
+0.20.4 - 2019-03-27
+^^^^^^^^^^^^^^^^^^^
+
+.. _new-features-1:
 
 New features
 ''''''''''''
@@ -13,11 +40,15 @@ New features
    generating piecewise exp. data
 -  Faster ``print_summary`` for AFT models.
 
+.. _api-changes-1:
+
 API changes
 '''''''''''
 
 -  Pandas is now correctly pinned to >= 0.23.0. This was always the
    case, but not specified in setup.py correctly.
+
+.. _bug-fixes-1:
 
 Bug fixes
 '''''''''
@@ -26,12 +57,12 @@ Bug fixes
 -  ``PiecewiseExponentialFitter`` is available with
    ``from lifelines import *``.
 
-.. _section-1:
+.. _section-2:
 
 0.20.3 - 2019-03-23
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-1:
+.. _new-features-2:
 
 New features
 ''''''''''''
@@ -44,12 +75,12 @@ New features
    ``plot_survival_function`` and
    ``confidence_interval_survival_function_``.
 
-.. _section-2:
+.. _section-3:
 
 0.20.2 - 2019-03-21
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-2:
+.. _new-features-3:
 
 New features
 ''''''''''''
@@ -64,7 +95,7 @@ New features
 -  add a ``lifelines.plotting.qq_plot`` for univariate parametric models
    that handles censored data.
 
-.. _api-changes-1:
+.. _api-changes-2:
 
 API changes
 '''''''''''
@@ -73,7 +104,7 @@ API changes
    @vpolimenov!
 -  The ``C`` column in ``load_lcd`` dataset is renamed to ``E``.
 
-.. _bug-fixes-1:
+.. _bug-fixes-2:
 
 Bug fixes
 '''''''''
@@ -89,7 +120,7 @@ Bug fixes
    the q parameter was below the truncation limit. This should have been
    ``-np.inf``
 
-.. _section-3:
+.. _section-4:
 
 0.20.1 - 2019-03-16
 ^^^^^^^^^^^^^^^^^^^
@@ -103,7 +134,7 @@ Bug fixes
    decades of development.
 -  suppressed unimportant warnings
 
-.. _api-changes-2:
+.. _api-changes-3:
 
 API changes
 '''''''''''
@@ -113,7 +144,7 @@ API changes
    This is no longer the case. A 0 will still be added if there is a
[…]
