### 2019-04-08 15f81a7
v0.20.5 (#701)
* fix error handling in Piecewise Exp model and lint

* adjust flaky test
