+++ lifelines/fitters/__init__.py
@@ -890,6 +890,8 @@ class ParametericUnivariateFitter(UnivariateFitter):
             self, estimate=getattr(self, "hazard_"), confidence_intervals=self.confidence_interval_hazard_, **kwargs
         )
 
+    fit_right_censoring = fit
+
 
 class KnownModelParametericUnivariateFitter(ParametericUnivariateFitter):
 
@@ -905,35 +907,97 @@ class ParametericAFTRegressionFitter(BaseFitter):
         self.fit_intercept = fit_intercept
         self._fitted_parameter_names = [self._primary_parameter_name, self._ancillary_parameter_name]
 
+    def _check_values(self, df, T, E, weights, entries):
+        check_for_numeric_dtypes_or_raise(df)
+        check_nans_or_infs(df)
+        check_nans_or_infs(T)
+        check_nans_or_infs(E)
+        check_positivity(T)
+        check_complete_separation(df, E, T, self.event_col)
+
+        if self.weights_col:
+            if (weights.astype(int) != weights).any() and not self.robust:
+                warnings.warn(
+                    dedent(
+                        """It appears your weights are not integers, possibly propensity or sampling scores then?
+                                        It's important to know that the naive variance estimates of the coefficients are biased. Instead a) set `robust=True` in the call to `fit`, or b) use Monte Carlo to
+                                        estimate the variances. See paper "Variance estimation when using inverse probability of treatment weighting (IPTW) with survival analysis"""
+                    ),
+                    StatisticalWarning,
+                )
+            if (weights <= 0).any():
+                raise ValueError("values in weight column %s must be positive." % self.weights_col)
+
+        if self.entry_col:
+            count_invalid_rows = (entries > T).sum()
+            if count_invalid_rows:
+                warnings.warn("""There exist %d rows where entry > duration.""")
+
+        if self.fit_intercept:
+            check_low_var(df)
+
     def _log_hazard(self, params, T, *Xs):
         # can be overwritten to improve convergence, see WeibullAFTFitter
         hz = self._hazard(params, T, *Xs)
         hz = anp.clip(hz, 1e-20, np.inf)
         return anp.log(hz)
 
-    def _negative_log_likelihood(self, params, T, E, W, starts, *Xs):
+    def _log_1m_sf(self, params, T, *Xs):
+        # equal to log(cdf), but often easier to express with sf.
+        # TODO: can be overwritten when log(cdf) is simple to express
+        cum_haz = self._cumulative_hazard(params, times, *Xs)
+        return anp.log1p(-anp.exp(-cum_haz))
+
+    def _survival_function(self, params, T, *Xs):
+        return anp.exp(-self._cumulative_hazard(params, T, *Xs))
+
+    def _log_likelihood_right_censoring(self, params, Ts, E, W, entries, Xs):
         warnings.simplefilter(action="ignore", category=FutureWarning)
 
-        non_zero_starts = starts > 0
-
+        non_zero_entries = entries > 0
+        T = Ts[1]
         ll = (
             (W * E * self._log_hazard(params, T, *Xs)).sum()
             - (W * self._cumulative_hazard(params, T, *Xs)).sum()
-            + self._cumulative_hazard(params, starts[non_zero_starts], *(X[non_zero_starts, :] for X in Xs)).sum()
+            + self._cumulative_hazard(params, entries[non_zero_entries], *(X[non_zero_entries, :] for X in Xs)).sum()
         )
-
-        if self.penalizer > 0:
-            penalty = self.l1_ratio * anp.abs(params).sum() + 0.5 * (1.0 - self.l1_ratio) * (params ** 2).sum()
-        else:
-            penalty = 0
-
         ll = ll / np.sum(W)
-        return -ll + self.penalizer * penalty
+        return ll
+
[…]
