### 2019-04-09 dadedba
interval inference with right censored data works
