+++ lifelines/fitters/__init__.py
@@ -52,6 +52,12 @@ from lifelines.utils import (
 __all__ = []
 
 
+class CensoringType:
+    LEFT = "left"
+    INTERVAL = "interval"
+    RIGHT = "right"
+
+
 def _must_call_fit_first(func):
     @wraps(func)
     def error_wrapper(*args, **kwargs):
@@ -82,6 +88,12 @@ class BaseFitter(object):
         except AttributeError:
             s = """<lifelines.%s>""" % classname
         return s
+
+    def fit(*args, **kwargs):
+        raise NotImplementedError()
+
+    def fit_right_censoring(self, *args, **kwargs):
+        return self.fit(*args, **kwargs)
 
 
 class UnivariateFitter(BaseFitter):
@@ -579,7 +591,7 @@ class ParametericUnivariateFitter(UnivariateFitter):
         """
         justify = string_justify(18)
         print(self)
-        print("{} = {}".format(justify("number of subjects"), self.durations.shape[0]))
+        print("{} = {}".format(justify("number of subjects"), self.event_observed.shape[0]))
         print("{} = {}".format(justify("number of events"), np.where(self.event_observed)[0].shape[0]))
         print("{} = {:.3f}".format(justify("log-likelihood"), self._log_likelihood))
         if self.left_censorship:
@@ -890,8 +902,6 @@ class ParametericUnivariateFitter(UnivariateFitter):
             self, estimate=getattr(self, "hazard_"), confidence_intervals=self.confidence_interval_hazard_, **kwargs
         )
 
-    fit_right_censoring = fit
-
 
 class KnownModelParametericUnivariateFitter(ParametericUnivariateFitter):
 
@@ -949,31 +959,43 @@ class ParametericAFTRegressionFitter(BaseFitter):
         return anp.log1p(-anp.exp(-cum_haz))
 
     def _survival_function(self, params, T, *Xs):
-        return anp.exp(-self._cumulative_hazard(params, T, *Xs))
+        return anp.clip(anp.exp(-self._cumulative_hazard(params, T, *Xs)), 1e-25, 1.0)
 
     def _log_likelihood_right_censoring(self, params, Ts, E, W, entries, Xs):
         warnings.simplefilter(action="ignore", category=FutureWarning)
 
+        T = Ts[0]
         non_zero_entries = entries > 0
-        T = Ts[1]
-        ll = (
-            (W * E * self._log_hazard(params, T, *Xs)).sum()
-            - (W * self._cumulative_hazard(params, T, *Xs)).sum()
-            + self._cumulative_hazard(params, entries[non_zero_entries], *(X[non_zero_entries, :] for X in Xs)).sum()
+
+        log_hz = self._log_hazard(params, T, *Xs)
+        cum_hz = self._cumulative_hazard(params, T, *Xs)
+        delayed_entries = self._cumulative_hazard(
+            params, entries[non_zero_entries], *(X[non_zero_entries, :] for X in Xs)
         )
+
+        ll = 0
+        ll = ll + (W * E * log_hz).sum()
+        ll = ll + -(W * cum_hz).sum()
+        ll = ll + (W[non_zero_entries] * delayed_entries).sum()
         ll = ll / np.sum(W)
         return ll
 
     def _log_likelihood_left_censoring(self, params, Ts, E, W, entries, Xs):
         warnings.simplefilter(action="ignore", category=FutureWarning)
 
-        # non_zero_starts = entries > 0
-        T = Ts[0]
+        T = Ts[1]
+        non_zero_entries = entries > 0
+
         log_hz = self._log_hazard(params, T, *Xs)
         cum_haz = self._cumulative_hazard(params, T, *Xs)
[…]
