### 2019-04-09 bbbd300
okay I think univariate models can do interval censoring now
