+++ lifelines/fitters/__init__.py
@@ -382,13 +382,13 @@ class ParametericUnivariateFitter(UnivariateFitter):
                 yield (lb + self._MIN_PARAMETER_VALUE, ub - self._MIN_PARAMETER_VALUE)
 
     def _cumulative_hazard(self, params, times):
-        raise NotImplementedError
+        raise NotImplementedError("Subclass implements this.")
 
     def _survival_function(self, params, times):
         return anp.exp(-self._cumulative_hazard(params, times))
 
     def _cumulative_density(self, params, times):
-        return 1 - anp.exp(-self._cumulative_hazard(params, times))
+        return 1 - self._survival_function(params, times)
 
     def _log_hazard(self, params, times):
         hz = self._hazard(params, times)
@@ -400,8 +400,9 @@ class ParametericUnivariateFitter(UnivariateFitter):
         cum_haz = self._cumulative_hazard(params, times)
         return anp.log1p(-anp.exp(-cum_haz))
 
-    def _negative_log_likelihood_left_censoring(self, params, T, E, _entry):
+    def _negative_log_likelihood_left_censoring(self, params, Ts, E, _entry):
 
+        T = Ts[1]
         n = T.shape[0]
         log_hz = self._log_hazard(params, T)
         cum_haz = self._cumulative_hazard(params, T)
@@ -411,12 +412,21 @@ class ParametericUnivariateFitter(UnivariateFitter):
 
         return -ll / n
 
-    def _negative_log_likelihood_right_censoring(self, params, T, E, entry):
-
+    def _negative_log_likelihood_right_censoring(self, params, Ts, E, entry):
+        T = Ts[0]
         n = T.shape[0]
         log_hz = self._log_hazard(params, T[E])
+        cum_haz = self._cumulative_hazard(params, T).sum()
 
-        ll = log_hz.sum() - self._cumulative_hazard(params, T).sum() + self._cumulative_hazard(params, entry).sum()
+        ll = log_hz.sum() - cum_haz.sum() + self._cumulative_hazard(params, entry).sum()
+        return -ll / n
+
+    def _negative_log_likelihood_interval_censoring(self, params, Ts, E, entry):
+        start, stop = Ts
+        n = start.shape[0]
+        ll = self._log_hazard(params, start[E]).sum() - self._cumulative_hazard(params, start[E]).sum()
+        ll = ll + anp.log(self._survival_function(params, start[~E]) - self._survival_function(params, stop[~E])).sum()
+        ll = ll + self._cumulative_hazard(params, entry).sum()
         return -ll / n
 
     def _compute_confidence_bounds_of_cumulative_hazard(self, alpha, ci_labels):
@@ -465,12 +475,14 @@ class ParametericUnivariateFitter(UnivariateFitter):
         df[ci_labels[1]] = transform(self._fitted_parameters_, self.timeline) - z * std_cumulative_hazard
         return df
 
-    def _fit_model(self, T, E, entry, show_progress=True):
+    def _fit_model(self, Ts, E, entry, show_progress=True):
 
         non_zero_entries = entry[entry > 0]
 
-        if self.left_censorship:
+        if self._censoring_type == CensoringType.LEFT:
             negative_log_likelihood = self._negative_log_likelihood_left_censoring
+        elif self._censoring_type == CensoringType.INTERVAL:
+            negative_log_likelihood = self._negative_log_likelihood_interval_censoring
         else:
             negative_log_likelihood = self._negative_log_likelihood_right_censoring
 
@@ -482,15 +494,15 @@ class ParametericUnivariateFitter(UnivariateFitter):
                 self._initial_values,
                 jac=True,
                 method="L-BFGS-B",
-                args=(T, E, non_zero_entries),
+                args=(Ts, E, non_zero_entries),
[…]
