+++ lifelines/fitters/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-
-
 import collections
 from functools import wraps
 import sys
@@ -93,6 +91,7 @@ class BaseFitter(object):
         raise NotImplementedError()
 
     def fit_right_censoring(self, *args, **kwargs):
+        self._censoring_type = CensoringType.RIGHT
         return self.fit(*args, **kwargs)
 
 
@@ -400,34 +399,45 @@ class ParametericUnivariateFitter(UnivariateFitter):
         cum_haz = self._cumulative_hazard(params, times)
         return anp.log1p(-anp.exp(-cum_haz))
 
-    def _negative_log_likelihood_left_censoring(self, params, Ts, E, _entry):
-
+    def _negative_log_likelihood_left_censoring(self, params, Ts, E, entry, W):
         T = Ts[1]
-        n = T.shape[0]
+        non_zero_entries = entry > 0
+
         log_hz = self._log_hazard(params, T)
         cum_haz = self._cumulative_hazard(params, T)
         log_1m_sf = self._log_1m_sf(params, T)
 
-        ll = (E * (log_hz - cum_haz - log_1m_sf)).sum() + log_1m_sf.sum()
+        ll = (E * W * (log_hz - cum_haz - log_1m_sf)).sum() + (W * log_1m_sf).sum()
+        ll = ll + (W[non_zero_entries] * self._cumulative_hazard(params, entry[non_zero_entries])).sum()
 
-        return -ll / n
+        return -ll / W.sum()
 
-    def _negative_log_likelihood_right_censoring(self, params, Ts, E, entry):
+    def _negative_log_likelihood_right_censoring(self, params, Ts, E, entry, W):
         T = Ts[0]
-        n = T.shape[0]
+        non_zero_entries = entry > 0
+
         log_hz = self._log_hazard(params, T[E])
-        cum_haz = self._cumulative_hazard(params, T).sum()
+        cum_haz = self._cumulative_hazard(params, T)
 
-        ll = log_hz.sum() - cum_haz.sum() + self._cumulative_hazard(params, entry).sum()
-        return -ll / n
+        ll = (W * log_hz).sum() - (W * cum_haz).sum()
+        ll = ll + (W[non_zero_entries] * self._cumulative_hazard(params, entry[non_zero_entries])).sum()
+        return -ll / W.sum()
 
-    def _negative_log_likelihood_interval_censoring(self, params, Ts, E, entry):
+    def _negative_log_likelihood_interval_censoring(self, params, Ts, E, entry, W):
         start, stop = Ts
-        n = start.shape[0]
-        ll = self._log_hazard(params, start[E]).sum() - self._cumulative_hazard(params, start[E]).sum()
-        ll = ll + anp.log(self._survival_function(params, start[~E]) - self._survival_function(params, stop[~E])).sum()
-        ll = ll + self._cumulative_hazard(params, entry).sum()
-        return -ll / n
+        non_zero_entries = entry > 0
+
+        ll = (W[E] * self._log_hazard(params, start[E])).sum() - (
+            W[E] * self._cumulative_hazard(params, start[E])
+        ).sum()
+        ll = (
+            ll
+            + (
+                W[~E] * anp.log(self._survival_function(params, start[~E]) - self._survival_function(params, stop[~E]))
+            ).sum()
+        )
+        ll = ll + (W[non_zero_entries] * self._cumulative_hazard(params, entry[non_zero_entries])).sum()
+        return -ll / W.sum()
 
     def _compute_confidence_bounds_of_cumulative_hazard(self, alpha, ci_labels):
[…]
