+++ lifelines/fitters/log_normal_aft_fitter.py
@@ -84,6 +84,17 @@ class LogNormalAFTFitter(ParametericAFTRegressionFitter):
 
         return logpdf(Z) - log_sigma_ - np.log(T) - logsf(Z)
 
+    def _log_1m_sf(self, params, T, *Xs):
+        mu_params = params[self._LOOKUP_SLICE["mu_"]]
+        mu_ = np.dot(Xs[0], mu_params)
+
+        sigma_params = params[self._LOOKUP_SLICE["sigma_"]]
+
+        log_sigma_ = np.dot(Xs[1], sigma_params)
+        sigma_ = np.exp(log_sigma_)
+        Z = (np.log(T) - mu_) / sigma_
+        return norm.logcdf(Z, loc=0, scale=1)
+
     def predict_percentile(self, X, ancillary_X=None, p=0.5):
         """
         Returns the median lifetimes for the individuals, by default. If the survival curve of an
