### 2019-04-10 786b20c
starting to add tests against R
