+++ lifelines/fitters/__init__.py
@@ -419,7 +419,7 @@ class ParametericUnivariateFitter(UnivariateFitter):
         log_hz = self._log_hazard(params, T[E])
         cum_haz = self._cumulative_hazard(params, T)
 
-        ll = (W * log_hz).sum() - (W * cum_haz).sum()
+        ll = (W[E] * log_hz).sum() - (W * cum_haz).sum()
         ll = ll + (W[non_zero_entries] * self._cumulative_hazard(params, entry[non_zero_entries])).sum()
         return -ll / W.sum()
 
@@ -677,9 +677,9 @@ class ParametericUnivariateFitter(UnivariateFitter):
             self with new properties like ``cumulative_hazard_``, ``survival_function_``
 
         """
-        check_nans_or_infs(durations)
-        check_positivity(durations)
         self.durations = np.asarray(pass_for_numeric_dtypes_or_raise_array(durations))
+        check_nans_or_infs(self.durations)
+        check_positivity(self.durations)
         self._censoring_type = CensoringType.RIGHT
 
         return self._fit(
@@ -704,6 +704,7 @@ class ParametericUnivariateFitter(UnivariateFitter):
         ci_labels=None,
         show_progress=False,
         entry=None,
+        weights=None,
     ):  # pylint: disable=too-many-arguments
         """
         Parameters
@@ -735,10 +736,10 @@ class ParametericUnivariateFitter(UnivariateFitter):
 
         """
         # TODO: new name
-        check_nans_or_infs(durations)
-        check_positivity(durations)
 
         self.durations = np.asarray(pass_for_numeric_dtypes_or_raise_array(durations))
+        check_nans_or_infs(self.durations)
+        check_positivity(self.durations)
         self._censoring_type = CensoringType.LEFT
         return self._fit(
             (None, self.durations),
@@ -1200,7 +1201,7 @@ class ParametericAFTRegressionFitter(BaseFitter):
         entry_col=None,
     ):
         """
-        Fit the accelerated failure time model to a dataset.
+        Fit the accelerated failure time model to a right-censored dataset.
 
         Parameters
         ----------
@@ -1312,6 +1313,84 @@ class ParametericAFTRegressionFitter(BaseFitter):
         initial_point=None,
         entry_col=None,
     ):
+        """
+        Fit the accelerated failure time model to a left-censored dataset.
+
+        Parameters
+        ----------
+        df: DataFrame
+            a Pandas DataFrame with necessary columns `duration_col` and
+            `event_col` (see below), covariates columns, and special columns (weights).
+            `duration_col` refers to
+            the lifetimes of the subjects. `event_col` refers to whether
+            the 'death' events was observed: 1 if observed, 0 else (censored).
+
+        start_col: string
+            the name of the column in DataFrame that contains the subjects'
+            left-most observation.
+
+        stop_col: string
+            the name of the column in DataFrame that contains the subjects'
+            right-most observation.
+
+        event_col: string, optional
+            the  name of the column in DataFrame that contains the subjects' death
+            observation. If left as None, assume all individuals are uncensored.
+
+        show_progress: boolean, optional (default=False)
+            since the fitter is iterative, show convergence
+            diagnostics. Useful if convergence is failing.
+
+        ancillary_df: None, boolean, or DataFrame, optional (default=None)
+            Choose to model the ancillary parameters.
+            If None or False, explicitly do not fit the ancillary parameters using any covariates.
+            If True, model the ancillary parameters with the same covariates as ``df``.
[…]
