+++ docs/index.rst
@@ -68,6 +68,7 @@ Contents:
 
   Gitter channel <https://gitter.im/python-lifelines/Lobby>
   Create a GitHub issue <https://github.com/camdavidsonpilon/lifelines/issues>
+  Development blog <https://dataorigami.net/blogs/napkin-folding/tagged/lifelines>
 
 Installation
 ------------------------------
