### 2019-04-10 CamDavidsonPilon/lifelines#703 @sunhwan
minor issue
* `examples/Customer Churn.ipynb` does not have matplotlib import.

* `examples/Modelling time-lagged conversion rates.ipynb` fails at 5th code cell:

```
lifelines/lifelines/utils/__init__.py in check_positivity(array)
    879     if np.any(array <= 0):
    880         raise ValueError(
--> 881             "This model does not allow for non-positive durations. Suggestion: add a small positive value to zero elements."
    882         )
    883

ValueError: This model does not allow for non-positive durations. Suggestion: add a small positive value to zero elements.
```

* Requirement missing for developer: Package `dill` is missing during pytest run. Maybe add requirements.txt file for developer who may want to run tests?

* `perf_tests/weibull_aft_perf.py` fails.

```
Traceback (most recent call last):
  File "weibull_aft_perf.py", line 20, in <module>
    wp.fit(df, duration_col="week", event_col="arrest", start_col="start")
TypeError: fit() got an unexpected keyword argument 'start_col'
```
