### 2019-04-10 #1317 @sunhwan
@trallard I went through the checklist and I was satisfied with most of the item in the list. I found a few problems running some script/file. I left a comment in the repository. https://github.com/CamDavidsonPilon/lifelines/issues/703

This package looks great and I'm okay to accept after the author address the minor issue. Are reviewers expected to write a separate review?
