### 2019-04-10 CamDavidsonPilon/lifelines#703 @CamDavidsonPilon
Awesome, thanks for letting me know about these!
