### 2019-04-11 dcd95ee
Fix types in docstring
