+++ lifelines/fitters/__init__.py
@@ -421,14 +421,20 @@ class ParametericUnivariateFitter(UnivariateFitter):
     def _negative_log_likelihood_interval_censoring(self, params, Ts, E, entry, W):
         start, stop = Ts
         non_zero_entries = entry > 0
+        observed_weights, censored_weights = W[E], W[~E]
+        censored_starts = start[~E]
+        observed_stops, censored_stops = stop[E], stop[~E]
 
-        ll = (W[E] * self._log_hazard(params, start[E])).sum() - (
-            W[E] * self._cumulative_hazard(params, start[E])
+        ll = (observed_weights * self._log_hazard(params, observed_stops)).sum() - (
+            observed_weights * self._cumulative_hazard(params, observed_stops)
         ).sum()
         ll = (
             ll
             + (
-                W[~E] * anp.log(self._survival_function(params, start[~E]) - self._survival_function(params, stop[~E]))
+                censored_weights
+                * anp.log(
+                    self._survival_function(params, censored_starts) - self._survival_function(params, censored_stops)
+                )
             ).sum()
         )
         ll = ll + (W[non_zero_entries] * self._cumulative_hazard(params, entry[non_zero_entries])).sum()
@@ -485,7 +491,7 @@ class ParametericUnivariateFitter(UnivariateFitter):
             negative_log_likelihood = self._negative_log_likelihood_left_censoring
         elif self._censoring_type == CensoringType.INTERVAL:
             negative_log_likelihood = self._negative_log_likelihood_interval_censoring
-        else:
+        elif self._censoring_type == CensoringType.RIGHT:
             negative_log_likelihood = self._negative_log_likelihood_right_censoring
 
         with warnings.catch_warnings():
@@ -712,9 +718,11 @@ class ParametericUnivariateFitter(UnivariateFitter):
         weights=None,
     ):  # pylint: disable=too-many-arguments
         """
+        Fit the model to a left-censored dataset
+
         Parameters
         ----------
-        durations: an array, or pd.Series TODO: new name
+        durations: an array, or pd.Series
           length n, duration subject was observed for
         event_observed: numpy array or pd.Series, optional
           length n, True if the the death was observed, False if the event was lost (right-censored). Defaults all True if event_observed==None
@@ -740,7 +748,6 @@ class ParametericUnivariateFitter(UnivariateFitter):
             self with new properties like ``cumulative_hazard_``, ``survival_function_``
 
         """
-        # TODO: new name
 
         self.durations = np.asarray(pass_for_numeric_dtypes_or_raise_array(durations))
         check_nans_or_infs(self.durations)
@@ -772,12 +779,17 @@ class ParametericUnivariateFitter(UnivariateFitter):
         weights=None,
     ):  # pylint: disable=too-many-arguments
         """
+        Fit the model to an interval censored dataset.
+
         Parameters
         ----------
-        durations: an array, or pd.Series
-          length n, duration subject was observed for
+        start: an array, or pd.Series
+          length n, the start of the period the subject experienced the event in.
+        stop: an array, or pd.Series
+          length n, the end of the period the subject experienced the event in. If the value is equal to the corresponding value in start, then
+          the individual's event was observed (not censored).
         event_observed: numpy array or pd.Series, optional
-          length n, True if the the death was observed, False if the event was lost (right-censored). Defaults all True if event_observed==None
+          length n, if left optional, infer from ``start`` and ``stop`` (if start==stop then event observed, if start < stop, then event censored)
         timeline: list, optional
             return the estimate at the values in timeline (positively increasing)
[…]
