+++ lifelines/fitters/__init__.py
@@ -767,8 +767,8 @@ class ParametericUnivariateFitter(UnivariateFitter):
 
     def fit_interval_censoring(
         self,
-        start,
-        stop,
+        lower_bound,
+        upper_bound,
         event_observed=None,
         timeline=None,
         label=None,
@@ -783,13 +783,13 @@ class ParametericUnivariateFitter(UnivariateFitter):
 
         Parameters
         ----------
-        start: an array, or pd.Series
+        lower_bound: an array, or pd.Series
           length n, the start of the period the subject experienced the event in.
-        stop: an array, or pd.Series
-          length n, the end of the period the subject experienced the event in. If the value is equal to the corresponding value in start, then
+        upper_bound: an array, or pd.Series
+          length n, the end of the period the subject experienced the event in. If the value is equal to the corresponding value in lower_bound, then
           the individual's event was observed (not censored).
         event_observed: numpy array or pd.Series, optional
-          length n, if left optional, infer from ``start`` and ``stop`` (if start==stop then event observed, if start < stop, then event censored)
+          length n, if left optional, infer from ``lower_bound`` and ``upper_cound`` (if lower_bound==upper_bound then event observed, if lower_bound < upper_bound, then event censored)
         timeline: list, optional
             return the estimate at the values in timeline (positively increasing)
         label: string, optional
@@ -812,27 +812,27 @@ class ParametericUnivariateFitter(UnivariateFitter):
             self with new properties like ``cumulative_hazard_``, ``survival_function_``
 
         """
-        check_nans_or_infs(start)
-        check_positivity(stop)
+        check_nans_or_infs(lower_bound)
+        check_positivity(upper_bound)
 
-        self.stop = np.asarray(pass_for_numeric_dtypes_or_raise_array(stop))
-        self.start = np.asarray(pass_for_numeric_dtypes_or_raise_array(start))
+        self.upper_bound = np.asarray(pass_for_numeric_dtypes_or_raise_array(upper_bound))
+        self.lower_bound = np.asarray(pass_for_numeric_dtypes_or_raise_array(lower_bound))
 
-        if (self.stop < self.start).any():
-            raise ValueError("All stop times must be greater than or equal to start times.")
+        if (self.upper_bound < self.lower_bound).any():
+            raise ValueError("All upper_bound times must be greater than or equal to lower_bound times.")
 
         if event_observed is None:
-            event_observed = self.stop == self.start
+            event_observed = self.upper_bound == self.lower_bound
 
-        if ((self.start == self.stop) != event_observed).any():
+        if ((self.lower_bound == self.upper_bound) != event_observed).any():
             raise ValueError(
-                "For all rows, start == stop if and only if event observed = 1 (uncensored). Likewise, start < stop if and only if event observed = 0 (censored)"
+                "For all rows, lower_bound == upper_bound if and only if event observed = 1 (uncensored). Likewise, lower_bound < upper_bound if and only if event observed = 0 (censored)"
             )
 
         self._censoring_type = CensoringType.INTERVAL
 
         return self._fit(
-            (np.clip(self.start, 1e-20, 1e25), np.clip(self.stop, 1e-20, 1e25)),
+            (np.clip(self.lower_bound, 1e-20, 1e25), np.clip(self.upper_bound, 1e-20, 1e25)),
             event_observed=event_observed,
             timeline=timeline,
             label=label,
@@ -1327,8 +1327,8 @@ class ParametericAFTRegressionFitter(BaseFitter):
     def fit_interval_censoring(
         self,
         df,
-        start_col,
-        stop_col,
+        lower_bound_col,
+        upper_bound_col,
         event_col=None,
[…]
