+++ docs/Changelog.rst
@@ -1,13 +1,51 @@
 Changelog
 ~~~~~~~~~
 
+0.21.0
+^^^^^^
+
+New features
+''''''''''''
+
+-  ``weights`` is now a optional kwarg for parametric univariate models.
+-  all univariate and multivariate parametric models now have ability to
+   handle left, right and interval censored data (the former two being
+   special cases of the latter). Users can use the
+   ``fit_right_censoring`` (which is an alias for ``fit``),
+   ``fit_left_censoring`` and ``fit_interval_censoring``.
+-  a new interval censored dataset is available under
+   ``lifelines.datasets.load_diabetes``
+
+API changes
+'''''''''''
+
+-  ``left_censorship`` on all univariate fitters has been deprecated.
+   Please use the new api ``model.fit_left_censoring(...)``.
+-  ``invert_y_axis`` in ``model.plot(...`` has been removed.
+-  ``entries`` property in multivariate parametric models has a new
+   Series name: ``entry``
+
+Bug fixes
+'''''''''
+
+-  lifelines was silently converting any NaNs in the event vector to
+   True. An error is now thrown instead.
+-  Fixed an error that didn’t let users use Numpy arrays in prediction
+   for AFT models
+
+.. _section-1:
+
 0.20.5 - 2019-04-08
 ^^^^^^^^^^^^^^^^^^^
+
+.. _new-features-1:
 
 New features
 ''''''''''''
 
 -  performance improvements for ``print_summary``.
+
+.. _api-changes-1:
 
 API changes
 '''''''''''
@@ -17,6 +55,8 @@ API changes
 -  in ``AalenJohansenFitter``, the ``variance`` parameter is renamed to
    ``variance_`` to align with the usual lifelines convention.
 
+.. _bug-fixes-1:
+
 Bug fixes
 '''''''''
 
@@ -24,12 +64,12 @@ Bug fixes
    test when using strata.
 -  Fixed some plotting bugs with ``AalenJohansenFitter``
 
-.. _section-1:
+.. _section-2:
 
 0.20.4 - 2019-03-27
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-1:
+.. _new-features-2:
 
 New features
 ''''''''''''
@@ -40,7 +80,7 @@ New features
    generating piecewise exp. data
 -  Faster ``print_summary`` for AFT models.
 
-.. _api-changes-1:
+.. _api-changes-2:
 
 API changes
 '''''''''''
@@ -48,7 +88,7 @@ API changes
 -  Pandas is now correctly pinned to >= 0.23.0. This was always the
    case, but not specified in setup.py correctly.
 
-.. _bug-fixes-1:
+.. _bug-fixes-2:
 
 Bug fixes
 '''''''''
@@ -57,12 +97,12 @@ Bug fixes
 -  ``PiecewiseExponentialFitter`` is available with
    ``from lifelines import *``.
 
-.. _section-2:
+.. _section-3:
 
 0.20.3 - 2019-03-23
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-2:
+.. _new-features-3:
 
 New features
 ''''''''''''
@@ -75,12 +115,12 @@ New features
[…]
