### 2019-04-12 e72ce18
fix left censoring
