### 2019-04-12 059b49b
v0.21.0 (#702)
* tests pass, though things feel slow

* interval inference with right censored data works

* initial tests against icenreg look positive

* okay I think univariate models can do interval censoring now

* univariate models now do interval

* starting to add tests against R

* mo-mo-monster PR

* need to add some more tests

* adding the diabetes dataset

* renam start and stop to lower bound and upper bound

* bump version

* fix left censoring

* fix plotting
