+++ lifelines/fitters/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-
-
 import collections
 from functools import wraps
 import sys
@@ -47,6 +45,7 @@ from lifelines.utils import (
     median_survival_times,
     normalize,
     concordance_index,
+    CensoringType,
 )
 
 __all__ = []
@@ -82,6 +81,13 @@ class BaseFitter(object):
         except AttributeError:
             s = """<lifelines.%s>""" % classname
         return s
+
+    def fit(*args, **kwargs):
+        raise NotImplementedError()
+
+    def fit_right_censoring(self, *args, **kwargs):
+        self._censoring_type = CensoringType.RIGHT
+        return self.fit(*args, **kwargs)
 
 
 class UnivariateFitter(BaseFitter):
@@ -370,13 +376,13 @@ class ParametericUnivariateFitter(UnivariateFitter):
                 yield (lb + self._MIN_PARAMETER_VALUE, ub - self._MIN_PARAMETER_VALUE)
 
     def _cumulative_hazard(self, params, times):
-        raise NotImplementedError
+        raise NotImplementedError("Subclass implements this.")
 
     def _survival_function(self, params, times):
         return anp.exp(-self._cumulative_hazard(params, times))
 
     def _cumulative_density(self, params, times):
-        return 1 - anp.exp(-self._cumulative_hazard(params, times))
+        return 1 - self._survival_function(params, times)
 
     def _log_hazard(self, params, times):
         hz = self._hazard(params, times)
@@ -388,24 +394,51 @@ class ParametericUnivariateFitter(UnivariateFitter):
         cum_haz = self._cumulative_hazard(params, times)
         return anp.log1p(-anp.exp(-cum_haz))
 
-    def _negative_log_likelihood_left_censoring(self, params, T, E, _entry):
+    def _negative_log_likelihood_left_censoring(self, params, Ts, E, entry, W):
+        T = Ts[1]
+        non_zero_entries = entry > 0
 
-        n = T.shape[0]
         log_hz = self._log_hazard(params, T)
         cum_haz = self._cumulative_hazard(params, T)
         log_1m_sf = self._log_1m_sf(params, T)
 
-        ll = (E * (log_hz - cum_haz - log_1m_sf)).sum() + log_1m_sf.sum()
+        ll = (E * W * (log_hz - cum_haz - log_1m_sf)).sum() + (W * log_1m_sf).sum()
+        ll = ll + (W[non_zero_entries] * self._cumulative_hazard(params, entry[non_zero_entries])).sum()
 
-        return -ll / n
+        return -ll / W.sum()
 
-    def _negative_log_likelihood_right_censoring(self, params, T, E, entry):
+    def _negative_log_likelihood_right_censoring(self, params, Ts, E, entry, W):
+        T = Ts[0]
+        non_zero_entries = entry > 0
 
-        n = T.shape[0]
         log_hz = self._log_hazard(params, T[E])
+        cum_haz = self._cumulative_hazard(params, T)
 
-        ll = log_hz.sum() - self._cumulative_hazard(params, T).sum() + self._cumulative_hazard(params, entry).sum()
-        return -ll / n
+        ll = (W[E] * log_hz).sum() - (W * cum_haz).sum()
+        ll = ll + (W[non_zero_entries] * self._cumulative_hazard(params, entry[non_zero_entries])).sum()
+        return -ll / W.sum()
+
+    def _negative_log_likelihood_interval_censoring(self, params, Ts, E, entry, W):
+        start, stop = Ts
+        non_zero_entries = entry > 0
+        observed_weights, censored_weights = W[E], W[~E]
+        censored_starts = start[~E]
+        observed_stops, censored_stops = stop[E], stop[~E]
+
[…]
