+++ tests/test_estimation.py
@@ -66,6 +66,7 @@ from lifelines.datasets import (
     load_regression_dataset,
     load_stanford_heart_transplants,
     load_multicenter_aids_cohort_study,
+    load_diabetes,
 )
 from lifelines.generate_datasets import (
     generate_hazard_rates,
@@ -227,7 +228,7 @@ class TestParametricUnivariateFitters:
         T = np.maximum(T1, T2)
 
         for fitter in known_parametric_univariate_fitters:
-            fitter().fit(T, E, left_censorship=True)
+            fitter().fit_left_censoring(T, E)
 
     def test_parametric_univarite_fitters_can_print_summary(
         self, positive_sample_lifetimes, known_parametric_univariate_fitters
@@ -266,6 +267,17 @@ class TestParametricUnivariateFitters:
 
         with pytest.warns(StatisticalWarning, match="non-decreasing") as w:
             DecreasingFitter().fit([0.01, 0.5, 10.0, 20])
+
+    def test_parameteric_models_all_can_do_interval_censoring(self, known_parametric_univariate_fitters):
+        df = load_diabetes()
+        for fitter in known_parametric_univariate_fitters:
+            f = fitter().fit_interval_censoring(df["left"], df["right"])
+
+    def test_parameteric_models_fail_if_passing_in_bad_event_data(self, known_parametric_univariate_fitters):
+        df = load_diabetes()
+        for fitter in known_parametric_univariate_fitters:
+            with pytest.raises(ValueError, match="lower_bound == upper_bound"):
+                f = fitter().fit_interval_censoring(df["left"], df["right"], event_observed=np.ones_like(df["right"]))
 
 
 class TestUnivariateFitters:
@@ -340,8 +352,8 @@ class TestUnivariateFitters:
             [pd.to_datetime("2015-01-01 12:00"), pd.to_datetime("2015-01-02"), pd.to_datetime("2015-01-02 12:00")]
         )
         T = pd.to_datetime("2015-01-03") - t
-        for fitter in univariate_fitters:
-            with pytest.warns(UserWarning, match="convert"):
+        with pytest.warns(UserWarning, match="Attempting to convert an unexpected"):
+            for fitter in univariate_fitters:
                 f = fitter().fit(T)
                 try:
                     npt.assert_allclose(f.timeline, 1e9 * 12 * 60 * 60 * np.array([0, 1, 2, 3]))
@@ -508,7 +520,7 @@ class TestUnivariateFitters:
                 fitter().fit(T)
 
     def test_typeerror_is_thrown_if_there_is_nans_in_the_event_col(self, univariate_fitters):
-        T = np.arange(5)
+        T = np.arange(1, 5)
         E = [1, 0, None, 1, 1]
         for fitter in univariate_fitters:
             with pytest.raises(TypeError):
@@ -537,6 +549,13 @@ class TestUnivariateFitters:
             dif = (fitter.durations - unpickled.durations).sum()
             assert dif == 0
 
+    def test_all_models_have_censoring_type(self, positive_sample_lifetimes, univariate_fitters):
+        T = positive_sample_lifetimes[0]
+        for f in univariate_fitters:
+            fitter = f()
+            fitter.fit(T)
+            assert hasattr(fitter, "_censoring_type")
+
 
 class TestPiecewiseExponentialFitter:
[…]
