### 2019-04-12 d39fffb
removing axes from errorbar_kwargs if provided
