### 2019-04-12 094574e
removing pyplot calls if axes is present
