### 2019-04-13 225962e
removing axes from errorbar_kwargs if provided (#708)
* removing axes from errorbar_kwargs if provided

* removing pyplot calls if axes is present
