+++ lifelines/fitters/cox_time_varying_fitter.py
@@ -723,7 +723,7 @@ See https://stats.stackexchange.com/questions/11109/how-to-deal-with-perfect-sep
         """
         from matplotlib import pyplot as plt
 
-        ax = errorbar_kwargs.get("ax", None) or plt.figure().add_subplot(111)
+        ax = errorbar_kwargs.pop("ax", None) or plt.figure().add_subplot(111)
 
         errorbar_kwargs.setdefault("c", "k")
         errorbar_kwargs.setdefault("fmt", "s")
@@ -750,8 +750,9 @@ See https://stats.stackexchange.com/questions/11109/how-to-deal-with-perfect-sep
 
         tick_labels = [columns[i] for i in order]
 
-        plt.yticks(yaxis_locations, tick_labels)
-        plt.xlabel("log(HR) (%g%% CI)" % ((1 - self.alpha) * 100))
+        ax.set_yticks(yaxis_locations)
+        ax.set_yticklabels(tick_labels)
+        ax.set_xlabel("log(HR) (%g%% CI)" % ((1 - self.alpha) * 100))
 
         return ax
 
