### 2019-04-14 e0d33f0
Fix minor typo in docs
