### 2019-04-14 94963bf
Add Numpy import (#710)
