### 2019-04-26 538692f
better support for custom cols in time varying dataset creation, better support for interval in survival_table_from_events
