### 2019-04-26 dec9453
small updates to docs
