### 2019-04-26 5b8ddae
promote cdf_ plot
