### 2019-04-26 c6758dc
small code style updates
