### 2019-04-26 994c0a9
add support for arrays as breakpoints
