### 2019-04-26 7bc6ba9
fix some tests
