+++ docs/Changelog.rst
@@ -1,8 +1,33 @@
 Changelog
 ~~~~~~~~~
 
-0.21.0
-^^^^^^
+0.21.1 - 2019-04-26
+^^^^^^^^^^^^^^^^^^^
+
+New features
+''''''''''''
+
+-  users can provided their own start and stop column names in
+   ``add_covariate_to_timeline``
+-  PiecewiseExponentialFitter now allows numpy arrays as breakpoints
+
+API changes
+'''''''''''
+
+-  output of ``survival_table_from_events`` when collapsing rows to
+   intervals now removes the “aggregate” column multi-index.
+
+Bug fixes
+'''''''''
+
+-  fixed bug in CoxTimeVaryingFitter when ax is provided, thanks @j-i-l!
+
+.. _section-1:
+
+0.21.0 - 2019-04-12
+^^^^^^^^^^^^^^^^^^^
+
+.. _new-features-1:
 
 New features
 ''''''''''''
@@ -16,6 +41,8 @@ New features
 -  a new interval censored dataset is available under
    ``lifelines.datasets.load_diabetes``
 
+.. _api-changes-1:
+
 API changes
 '''''''''''
 
@@ -25,6 +52,8 @@ API changes
 -  ``entries`` property in multivariate parametric models has a new
    Series name: ``entry``
 
+.. _bug-fixes-1:
+
 Bug fixes
 '''''''''
 
@@ -33,19 +62,19 @@ Bug fixes
 -  Fixed an error that didn’t let users use Numpy arrays in prediction
    for AFT models
 
-.. _section-1:
+.. _section-2:
 
 0.20.5 - 2019-04-08
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-1:
+.. _new-features-2:
 
 New features
 ''''''''''''
 
 -  performance improvements for ``print_summary``.
 
-.. _api-changes-1:
+.. _api-changes-2:
 
 API changes
 '''''''''''
@@ -55,7 +84,7 @@ API changes
 -  in ``AalenJohansenFitter``, the ``variance`` parameter is renamed to
    ``variance_`` to align with the usual lifelines convention.
 
-.. _bug-fixes-1:
+.. _bug-fixes-2:
 
 Bug fixes
 '''''''''
@@ -64,12 +93,12 @@ Bug fixes
    test when using strata.
 -  Fixed some plotting bugs with ``AalenJohansenFitter``
 
-.. _section-2:
+.. _section-3:
 
 0.20.4 - 2019-03-27
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-2:
+.. _new-features-3:
 
 New features
 ''''''''''''
@@ -80,7 +109,7 @@ New features
    generating piecewise exp. data
 -  Faster ``print_summary`` for AFT models.
 
-.. _api-changes-2:
+.. _api-changes-3:
 
 API changes
 '''''''''''
@@ -88,7 +117,7 @@ API changes
 -  Pandas is now correctly pinned to >= 0.23.0. This was always the
    case, but not specified in setup.py correctly.
 
-.. _bug-fixes-2:
+.. _bug-fixes-3:
 
 Bug fixes
 '''''''''
@@ -97,12 +126,12 @@ Bug fixes
 -  ``PiecewiseExponentialFitter`` is available with
    ``from lifelines import *``.
 
-.. _section-3:
+.. _section-4:
 
 0.20.3 - 2019-03-23
 ^^^^^^^^^^^^^^^^^^^
 
[…]
