+++ lifelines/utils/__init__.py
@@ -1246,7 +1246,7 @@ def add_covariate_to_timeline(
         raise ValueError("delay parameter must be equal to or greater than 0")
 
     if any(col not in long_form_df for col in (id_col, event_col, start_col, stop_col)):
-        raise ValueError("Missing column in long_form_df")
+        raise IndexError("Missing column in long_form_df")
 
     cv[duration_col] += delay
     cv = cv.dropna()
