### 2019-05-02 f3b4ede
Doc improvements
