### 2019-05-02 c7745b3
Adding piecwise regression fitter (no tests yet though)
