### 2019-05-02 805796a
New default behaviour for `predict` in KaplanMeier and NelsonAalen: use the previous value, interpolate with interpolate flag.
