+++ lifelines/fitters/__init__.py
@@ -51,19 +51,6 @@ from lifelines.utils import (
 __all__ = []
 
 
-def _must_call_fit_first(func):
-    @wraps(func)
-    def error_wrapper(*args, **kwargs):
-        self = args[0]
-        try:
-            self._predict_label
-        except AttributeError:
-            raise RuntimeError("Must call `fit` first!")
-        return func(*args, **kwargs)
-
-    return error_wrapper
-
-
 class BaseFitter(object):
     def __init__(self, alpha=0.05):
         if not (0 < alpha <= 1.0):
@@ -91,7 +78,6 @@ class BaseFitter(object):
 
 
 class UnivariateFitter(BaseFitter):
-    @_must_call_fit_first
     def _update_docstrings(self):
         # Update their docstrings
         self.__class__.subtract.__doc__ = self.subtract.__doc__.format(self._estimate_name, self.__class__.__name__)
@@ -99,7 +85,6 @@ class UnivariateFitter(BaseFitter):
         self.__class__.predict.__doc__ = self.predict.__doc__.format(self.__class__.__name__)
         self.__class__.plot.__doc__ = _plot_estimate.__doc__.format(self.__class__.__name__, self._estimate_name)
 
-    @_must_call_fit_first
     def plot(self, **kwargs):
         """
         Plots a pretty figure of the model
@@ -146,7 +131,6 @@ class UnivariateFitter(BaseFitter):
             self, estimate=getattr(self, self._estimate_name), confidence_intervals=self.confidence_interval_, **kwargs
         )
 
-    @_must_call_fit_first
     def subtract(self, other):
         """
         Subtract the {0} of two {1} objects.
@@ -166,7 +150,6 @@ class UnivariateFitter(BaseFitter):
             columns=["diff"],
         )
 
-    @_must_call_fit_first
     def divide(self, other):
         """
         Divide the {0} of two {1} objects.
@@ -229,7 +212,6 @@ class UnivariateFitter(BaseFitter):
         """
         return self._conditional_time_to_event_()
 
-    @_must_call_fit_first
     def _conditional_time_to_event_(self):
         """
         Return a DataFrame, with index equal to survival_function_, that estimates the median
@@ -256,35 +238,27 @@ class UnivariateFitter(BaseFitter):
             - age
         )
 
-    @_must_call_fit_first
     def hazard_at_times(self, times, label=None):
         raise NotImplementedError
 
-    @_must_call_fit_first
     def survival_function_at_times(self, times, label=None):
         raise NotImplementedError
 
-    @_must_call_fit_first
     def cumulative_hazard_at_times(self, times, label=None):
         raise NotImplementedError
 
-    @_must_call_fit_first
     def cumulative_density_at_times(self, times, label=None):
         raise NotImplementedError
 
-    @_must_call_fit_first
     def plot_cumulative_hazard(self, **kwargs):
         raise NotImplementedError()
 
-    @_must_call_fit_first
     def plot_survival_function(self, **kwargs):
         raise NotImplementedError()
 
-    @_must_call_fit_first
     def plot_hazard(self, **kwargs):
         raise NotImplementedError()
 
-    @_must_call_fit_first
     def plot_cumulative_density(self, **kwargs):
         raise NotImplementedError()
 
@@ -292,7 +266,6 @@ class UnivariateFitter(BaseFitter):
 class ParametericUnivariateFitter(UnivariateFitter):
     """
     Without overriding anything, assumes all parameters must be greater than 0.
-
     """
 
     _KNOWN_MODEL = False
@@ -574,7 +547,6 @@ class ParametericUnivariateFitter(UnivariateFitter):
         return (self._fitted_parameters_ - self._initial_values) / self._compute_standard_errors().loc["se"]
 
     @property
-    @_must_call_fit_first
     def summary(self):
         """
[…]
