### 2019-05-11 ecd4c3c
Adding some tests
