+++ Makefile
@@ -9,7 +9,7 @@ else
 endif
 
 test:
-	py.test -rfs --cov=lifelines --block=False --cov-report term-missing
+	py.test tests/ -rfs --cov=lifelines --block=False --cov-report term-missing
 
 lint:
 ifeq ($(TRAVIS_PYTHON_VERSION), 2.7)
