### 2019-05-15 8de4afb
Changed the way printing is done through iterations by flushing stdout
