### 2019-05-16 1b8ffbf
change how I internally track how censoring is kept
