+++ lifelines/fitters/breslow_fleming_harrington_fitter.py
@@ -28,6 +28,7 @@ class BreslowFlemingHarringtonFitter(UnivariateFitter):
 
     """
 
+    @CensoringType.right_censoring
     def fit(
         self,
         durations,
@@ -65,7 +66,6 @@ class BreslowFlemingHarringtonFitter(UnivariateFitter):
           self, with new properties like ``survival_function_``.
 
         """
-        self._censoring_type = CensoringType.RIGHT
         self._label = label
         alpha = coalesce(alpha, self.alpha)
 
