+++ docs/Changelog.rst
@@ -1,8 +1,38 @@
 Changelog
 ~~~~~~~~~
 
+0.21.2 - 2019-04-26
+^^^^^^^^^^^^^^^^^^^
+
+New features
+''''''''''''
+
+-  New regression model: ``PiecewiseExponentialRegressionFitter`` is
+   available. See blog post here:
+   https://dataorigami.net/blogs/napkin-folding/churn
+-  Regression models have a new method ``log_likelihood_ratio_test``
+   that computes, you guessed it, the log-likelihood ratio test.
+   Previously this was an internal API that is being exposed.
+
+API changes
+'''''''''''
+
+-  The default behavior of the ``predict`` method on non-parametric
+   estimators (``KaplanMeierFitter``, etc.) has changed from (previous)
+   linear interpolation to (new) return last value. Linear interpolation
+   is still possible with the ``interpolate`` flag.
+-  removing ``_compute_likelihood_ratio_test`` on regression models. Use
+   ``log_likelihood_ratio_test`` now.
+
+Bug fixes
+'''''''''
+
+.. _section-1:
+
 0.21.1 - 2019-04-26
 ^^^^^^^^^^^^^^^^^^^
+
+.. _new-features-1:
 
 New features
 ''''''''''''
@@ -11,23 +41,27 @@ New features
    ``add_covariate_to_timeline``
 -  PiecewiseExponentialFitter now allows numpy arrays as breakpoints
 
+.. _api-changes-1:
+
 API changes
 '''''''''''
 
 -  output of ``survival_table_from_events`` when collapsing rows to
    intervals now removes the “aggregate” column multi-index.
 
+.. _bug-fixes-1:
+
 Bug fixes
 '''''''''
 
 -  fixed bug in CoxTimeVaryingFitter when ax is provided, thanks @j-i-l!
 
-.. _section-1:
+.. _section-2:
 
 0.21.0 - 2019-04-12
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-1:
+.. _new-features-2:
 
 New features
 ''''''''''''
@@ -41,7 +75,7 @@ New features
 -  a new interval censored dataset is available under
    ``lifelines.datasets.load_diabetes``
 
-.. _api-changes-1:
+.. _api-changes-2:
 
 API changes
 '''''''''''
@@ -52,7 +86,7 @@ API changes
 -  ``entries`` property in multivariate parametric models has a new
    Series name: ``entry``
 
-.. _bug-fixes-1:
+.. _bug-fixes-2:
 
 Bug fixes
 '''''''''
@@ -62,19 +96,19 @@ Bug fixes
 -  Fixed an error that didn’t let users use Numpy arrays in prediction
    for AFT models
 
-.. _section-2:
+.. _section-3:
 
 0.20.5 - 2019-04-08
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-2:
+.. _new-features-3:
 
 New features
 ''''''''''''
 
 -  performance improvements for ``print_summary``.
 
-.. _api-changes-2:
+.. _api-changes-3:
 
 API changes
 '''''''''''
@@ -84,7 +118,7 @@ API changes
 -  in ``AalenJohansenFitter``, the ``variance`` parameter is renamed to
[…]
