### 2019-05-16 85d1f2e
Changed the way printing is done through iterations by flushing stdout (#719)
