### 2019-05-16 af56f7a
V0.21.2 (#715)
* Doc improvements

* Adding piecwise regression fitter (no tests yet though)

* New default behaviour for `predict` in KaplanMeier and NelsonAalen: use the previous value, interpolate with interpolate flag.

* docs and changelog

* Adding some tests

* clean tests

* only tests /tests

* some cleanup

* change how I internally track how censoring is kept

* bump version

* fix tests
