+++ lifelines/fitters/__init__.py
@@ -51,27 +51,15 @@ from lifelines.utils import (
 __all__ = []
 
 
-def _must_call_fit_first(func):
-    @wraps(func)
-    def error_wrapper(*args, **kwargs):
-        self = args[0]
-        try:
-            self._predict_label
-        except AttributeError:
-            raise RuntimeError("Must call `fit` first!")
-        return func(*args, **kwargs)
-
-    return error_wrapper
-
-
 class BaseFitter(object):
     def __init__(self, alpha=0.05):
         if not (0 < alpha <= 1.0):
             raise ValueError("alpha parameter must be between 0 and 1.")
         self.alpha = alpha
+        self._class_name = self.__class__.__name__
 
     def __repr__(self):
-        classname = self.__class__.__name__
+        classname = self._class_name
         try:
             s = """<lifelines.%s: fitted with %d observations, %d censored>""" % (
                 classname,
@@ -82,24 +70,23 @@ class BaseFitter(object):
             s = """<lifelines.%s>""" % classname
         return s
 
+    @CensoringType.right_censoring
     def fit(*args, **kwargs):
         raise NotImplementedError()
 
+    @CensoringType.right_censoring
     def fit_right_censoring(self, *args, **kwargs):
-        self._censoring_type = CensoringType.RIGHT
         return self.fit(*args, **kwargs)
 
 
 class UnivariateFitter(BaseFitter):
-    @_must_call_fit_first
     def _update_docstrings(self):
         # Update their docstrings
-        self.__class__.subtract.__doc__ = self.subtract.__doc__.format(self._estimate_name, self.__class__.__name__)
-        self.__class__.divide.__doc__ = self.divide.__doc__.format(self._estimate_name, self.__class__.__name__)
-        self.__class__.predict.__doc__ = self.predict.__doc__.format(self.__class__.__name__)
-        self.__class__.plot.__doc__ = _plot_estimate.__doc__.format(self.__class__.__name__, self._estimate_name)
+        self.__class__.subtract.__doc__ = self.subtract.__doc__.format(self._estimate_name, self._class_name)
+        self.__class__.divide.__doc__ = self.divide.__doc__.format(self._estimate_name, self._class_name)
+        self.__class__.predict.__doc__ = self.predict.__doc__.format(self._class_name)
+        self.__class__.plot.__doc__ = _plot_estimate.__doc__.format(self._class_name, self._estimate_name)
 
-    @_must_call_fit_first
     def plot(self, **kwargs):
         """
         Plots a pretty figure of the model
@@ -146,7 +133,6 @@ class UnivariateFitter(BaseFitter):
             self, estimate=getattr(self, self._estimate_name), confidence_intervals=self.confidence_interval_, **kwargs
         )
 
-    @_must_call_fit_first
     def subtract(self, other):
         """
         Subtract the {0} of two {1} objects.
@@ -166,7 +152,6 @@ class UnivariateFitter(BaseFitter):
             columns=["diff"],
         )
 
-    @_must_call_fit_first
     def divide(self, other):
         """
         Divide the {0} of two {1} objects.
@@ -188,15 +173,18 @@ class UnivariateFitter(BaseFitter):
         )
         return t
 
-    @_must_call_fit_first
-    def predict(self, times):
+    def predict(self, times, interpolate=False):
         """
         Predict the {0} at certain point in time. Uses a linear interpolation if
         points in time are not in the index.
 
         Parameters
         ----------
-        times: a scalar or an array of times to predict the value of {0} at.
+        times: scalar, or array
+            a scalar or an array of times to predict the value of {0} at.
+        interpolate: boolean, optional (default=False)
+            for methods that produce a stepwise solution (Kaplan-Meier, Nelson-Aalen, etc), turning this to
[…]
