+++ docs/Changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 ~~~~~~~~~
 
-0.21.2 - 2019-04-26
+0.21.2 - 2019-05-16
 ^^^^^^^^^^^^^^^^^^^
 
 New features
