+++ docs/lifelines.fitters.rst
@@ -111,6 +111,13 @@ lifelines.fitters.log\_logistic\_aft\_fitter module
     :undoc-members:
 
 
+lifelines.fitters.piecewise\_exponential\_regression\_fitter module
+-----------------------------------------------------
+
+.. automodule:: lifelines.fitters.piecewise_exponential_regression_fitter
+    :members:
+    :undoc-members:
+
 
 .. automodule:: lifelines.fitters
     :members:
