### 2019-05-27 4504e09
Fixed scipy version to <=1.2.1
Fixed scipy version to <= 1.2.1 until autograd HIPS/autograd#490 is released. After that we could fix autograd to >= whatever version they release with that specific PR and return scipy to >=1.0.
