### 2019-05-28 e671855
Fixed scipy to >=1.0,<=1.2.1, changed base-requirements.txt for travis
