### 2019-05-28 c134ad4
Fixed scipy version to <=1.2.1 (#726)
* Fixed scipy version to <=1.2.1

Fixed scipy version to <= 1.2.1 until autograd HIPS/autograd#490 is released. After that we could fix autograd to >= whatever version they release with that specific PR and return scipy to >=1.0.

* Fixed scipy to >=1.0,<=1.2.1, changed base-requirements.txt for travis
