### 2019-05-31 1ff9eed
Fixing reference in warning text for "unexpected datetype"
The warning text refers to `lifelines.utils.datetime_to_durations` but should refer to `lifelines.utils.datetimes_to_durations` (notice "s" after datetime).
