### 2019-06-03 2ed767c
allowing users to plot HR instead of log(HR)
