### 2019-06-03 57ec5e1
allow users to subset columns to test ph assuptions
