### 2019-06-03 93c462d
make covariates_from_event_matrix handle nulls better
