### 2019-06-04 c28b9e7
Added an adapter allowing to evaluate plug lifelines models to the stuff expecting sklearn models
