### 2019-06-04 bbc685d
Some minor improvements in adapter
