### 2019-06-04 af75167
Added an adapter allowing to plug lifelines models to the stuff expecting sklearn models (#736)
* Added an adapter allowing to evaluate plug lifelines models to the stuff expecting sklearn models

* Some minor improvements in adapter
