### 2019-06-05 e3fd90a
Adding new sklearn adapter + docs
