+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ### Changelog
 
-#### 0.21.3
+#### 0.21.3 - 2019-06-04
 
 ##### New features
+ - include in lifelines is a scikit-learn adapter so lifeline's models can be used with scikit-learn's API. See [documentation here](https://lifelines.readthedocs.io/en/latest/Compatibility%20with%20scikit-learn.html).
  - `CoxPHFitter.plot` now accepts a `hazard_ratios` (boolean) parameter that will plot the hazard ratios (and CIs) instead of the log-hazard ratios.
  - `CoxPHFitter.check_assumptions` now accepts a `columns` parameter to specify only checking a subset of columns.
 
