### 2019-06-05 c58ca85
Replaced setup.py to setup.cfg for storage of metadata. Version is now fetched from git tag. Lifted restriction on scipy version because `worksforme` (tests pass) at the latest one.
