+++ setup.cfg
@@ -1,0 +1,39 @@
+[metadata]
+name = lifelines
+author = Cameron Davidson-Pilon
+author_email = cam.davidson.pilon@gmail.com
+license = MIT
+license_file = LICENSE
+description = Survival analysis in Python, including Kaplan Meier, Nelson Aalen and regression
+keywords = survival, analysis, statistics, data, analysis
+url = https://github.com/CamDavidsonPilon/lifelines
+long_description = file: README.md
+long_description_content_type = text/markdown
+classifiers =
+    Development Status :: 4 - Beta
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python
+    Programming Language :: Python :: 3.5
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Topic :: Scientific/Engineering
+
+[options]
+packages = find:
+install_requires =
+    numpy>=1.6.0
+    scipy>=1.0
+    pandas>=0.23.0
+    matplotlib>=3.0
+    bottleneck>=1.0
+    autograd>=1.2
+setup_requires =
+    setuptools>=30.3.0
+    setuptools_scm
+python_requires = >=3.5
+
+[options.package_data]
+lifelines =
+    ../README.md
+    ../README.txt
+    datasets/*
