### 2019-06-05 a1c65d9
Implemented an adapter allowing to plug lifelines models to the stuff expecting sklearn models, such as  cross_val_predict Fixes #553
