+++ lifelines/utils/sklearn_adapter.py
@@ -7,33 +7,153 @@ try:
 except ImportError:
     raise ImportError("scikit-learn must be installed on the local system to use this utility class.")
 from . import concordance_index
+from ..fitters import BaseFitter
 
-__all__ = ["sklearn_adapter"]
+__all__ = ("sklearn_adapter", "LifelinesSKLearnAdapter")
 
 
 def filter_kwargs(f, kwargs):
-    s = inspect.signature(f)
-    res = {k: kwargs[k] for k in s.parameters if k in kwargs}
-    return res
+    s = inspect.getfullargspec(f)
+    if s.varkw:
+        return kwargs
+    else:
+        allArgs = set()
+        if s.args:
+            allArgs |= set(s.args)
+        if s.kwonlyargs:
+            allArgs |= set(s.kwonlyargs)
+        fedArgs = set(kwargs)
+        redundantArgs = fedArgs - allArgs
+        presentArgs = fedArgs & allArgs
+        #if redundantArgs:
+        #    warnings.warn("Following args are redundant for " + str(f) + str(inspect.signature(f)) + ": " + repr(redundantArgs))
+        res = {k: kwargs[k] for k in presentArgs}
+        return res
 
 
-class _SklearnModel(BaseEstimator, MetaEstimatorMixin, RegressorMixin):
-    def __init__(self, **kwargs):
-        self._params = kwargs
-        self.lifelines_model = self.lifelines_model(**filter_kwargs(self.lifelines_model.__init__, self._params))
+class LifelinesSKLearnAdapter(BaseEstimator, MetaEstimatorMixin, RegressorMixin):
+    def __init__(self, lifelines_model=None, call_params=None, y_arg_names="duration_col", event_arg_name="event_col", remap=None, concordance_remap=None, predict_method="predict_expectation", fit_method="fit", etalon_func=None, **kwargs):
+        """
+        Parameters
+        ----------
 
-        self._params["duration_col"] = "duration_col"
-        self._params["event_col"] = self._event_col
+        lifelines_model: BaseFitter
+            lifelines model to be wrapper. Example: ``CoxPHFitter(...)`` or ``WeibullAFTFitter(...)``
+        params: 
+            args to pass to the functions like `fit`
+        event_col: string
+            The column in your DataFrame that represents (if applicable) the event column
+        predict_method: `str` or `callable`
+            Can be a name of method like ``"predict_median", "predict_expectation"`` or a callable
+        y_arg_names: string or a tuple of strings
+            name(s) of columns considered to be targets. It can be `duration_col` or fields for parametric lifelines_models params or something else. Heavily depends on a lifelines_model.
+        event_arg_name: string or None
+            column name to pass censorship status to a lifelines_model
+        remap: dict or None
+            sometimes you need to rename params.
+        concordance_remap: callable
+            a function to remap concordance. For example, if you compare survival times to hazards you need `lambda sself, c: 1. - c`
+        etalon_func: callable
+            Can be used when `y_arg_names` are not survival times but parameters of a parametric lifelines_model.
+        _sklearnParamsWorkaround: dict
+            don't touch it - it is for sklearn, not for you
+        """
+        self.y_arg_names = y_arg_names
+        self.event_arg_name = event_arg_name
+        if remap is None:
+            remap = {}
+
+        self.remap = remap
+
+        if etalon_func is None:
+
+            def etalon_func(sself, Y):
+                return Y
+
+        self.etalon_func = etalon_func
+
+        if concordance_remap is None:
+
+            def concordance_remap(sself, c):
+                return c
+
+        self.concordance_remap = concordance_remap
+        self.call_params = call_params
+        
+        ##############################################################
+
+        if callable(lifelines_model):
+            if isinstance(lifelines_model, type):
[…]
