### 2019-06-05 574d46b
Revert "Added an adapter allowing to plug lifelines models to the stuff expecting sklearn models (#736)"
This reverts commit af751676711c83008668b0f196432e2001cceb4f.
