### 2019-06-05 ccd70b4
well this works, but I hate the globals() hack
