### 2019-06-06 e7a63f3
lots of updates to sklean adapter
