### 2019-06-06 4f57a39
lots of updates to sklean adapter
