+++ tests/test_estimation.py
@@ -1874,8 +1874,10 @@ class TestWeibullAFTFitter:
         aft.predict_cumulative_hazard(rossi, ancillary_X=rossi)
         aft.predict_survival_function(rossi, ancillary_X=rossi)
 
-        with pytest.raises(ValueError):
-            aft.predict_survival_function(rossi)
+        aft.predict_median(rossi)
+        aft.predict_percentile(rossi)
+        aft.predict_cumulative_hazard(rossi)
+        aft.predict_survival_function(rossi)
 
     def test_passing_in_additional_ancillary_df_in_predict_methods_okay_if_not_fitted_with_one(self, rossi, aft):
 
