### 2019-06-06 9e6d817
Sklearn adapter v0.1 (#738)
* Fixing reference in warning text for "unexpected datetype" (#729)

The warning text refers to `lifelines.utils.datetime_to_durations` but should refer to `lifelines.utils.datetimes_to_durations` (notice "s" after datetime).

* Added an adapter allowing to plug lifelines models to the stuff expecting sklearn models (#736)

* Added an adapter allowing to evaluate plug lifelines models to the stuff expecting sklearn models

* Some minor improvements in adapter

* Adding new sklearn adapter + docs

* well this works, but I hate the globals() hack

* lots of updates to
[…]
