### 2019-06-06 84466e5
change _fitter to lifelines_model
