+++ experiments/sklearn_adapter.py
@@ -12,7 +12,7 @@ rossi = load_rossi()
 Y = rossi.pop("week").values
 X = rossi
 
-base_class = sklearn_adapter(CoxPHFitter, duration_col="week", event_col="arrest")
+base_class = sklearn_adapter(CoxPHFitter, event_col="arrest")
 cph = base_class(penalizer=1.0)
 
 cph.fit(X, Y)
@@ -20,7 +20,7 @@ cph.fit(X, Y)
 scores = cross_val_score(cph, X, Y, cv=5)
 print(scores)
 
-base_model = sklearn_adapter(WeibullAFTFitter, duration_col="week", event_col="arrest")
+base_model = sklearn_adapter(WeibullAFTFitter, event_col="arrest")
 
 clf = GridSearchCV(base_model(), {"penalizer": [0, 1, 10], "model_ancillary": [True, False]}, cv=5)
 clf.fit(X, Y)
