### 2019-06-06 d22e535
V0.21.3 (#731)
* allowing users to plot HR instead of log(HR)

* allow users to subset columns to test ph assuptions

* make covariates_from_event_matrix handle nulls better

* fix this

* update some docs
"

* bump version

* Sklearn adapter v0.1 (#738)

* Fixing reference in warning text for "unexpected datetype" (#729)

The warning text refers to `lifelines.utils.datetime_to_durations` but should refer to `lifelines.utils.datetimes_to_durations` (notice "s" after datetime).

* Added an adapter allowing to plug lifelines models to the stuff expecting sklearn models (#736)

* Added an adapter
[…]
