+++ tests/utils/test_utils.py
@@ -10,9 +10,10 @@ import numpy.testing as npt
 from numpy.linalg import norm, lstsq
 from numpy.random import randn
 
-from lifelines import CoxPHFitter
+from lifelines import CoxPHFitter, WeibullAFTFitter
 from lifelines.datasets import load_regression_dataset, load_larynx, load_waltons, load_rossi
 from lifelines import utils
+from lifelines.utils.sklearn_adapter import sklearn_adapter
 
 
 def test_format_p_values():
@@ -720,6 +721,31 @@ class TestLongDataFrameUtils(object):
         )
         assert_frame_equal(expected, df, check_like=True)
 
+    def test_covariates_from_event_matrix_with_simple_addition(self):
+
+        base_df = pd.DataFrame(
+            [[1, 0, 5, 1], [2, 0, 4, 1], [3, 0, 8, 1], [4, 0, 4, 1]], columns=["id", "start", "stop", "e"]
+        )
+
+        event_df = pd.DataFrame([[1, 1], [2, 2], [3, 3], [4, None]], columns=["id", "poison"])
+        cv = utils.covariates_from_event_matrix(event_df, "id")
+        ldf = utils.add_covariate_to_timeline(base_df, cv, "id", "duration", "e", cumulative_sum=True)
+        assert pd.notnull(ldf).all().all()
+
+        expected = pd.DataFrame(
+            [
+                (0.0, 0.0, 1.0, 1, False),
+                (1.0, 1.0, 5.0, 1, True),
+                (0.0, 0.0, 2.0, 2, False),
+                (2.0, 1.0, 4.0, 2, True),
+                (0.0, 0.0, 3.0, 3, False),
+                (3.0, 1.0, 8.0, 3, True),
+                (0.0, 0.0, 4.0, 4, True),
+            ],
+            columns=["start", "cumsum_poison", "stop", "id", "e"],
+        )
+        assert_frame_equal(expected, ldf, check_dtype=False, check_like=True)
+
     def test_covariates_from_event_matrix(self):
 
         base_df = pd.DataFrame(
@@ -895,3 +921,91 @@ class TestStepSizer:
         assert ss.next() == start
         ss.update(20.0)
         assert ss.next() < start
+
+
+class TestSklearnAdapter:
+    @pytest.fixture
+    def X(self):
+        return load_regression_dataset().drop("T", axis=1)
+
+    @pytest.fixture
+    def Y(self):
+        return load_regression_dataset().pop("T")
+
+    def test_model_has_correct_api(self, X, Y):
+        base_model = sklearn_adapter(CoxPHFitter, event_col="E")
+        cph = base_model()
+        assert hasattr(cph, "fit")
+        cph.fit(X, Y)
+        assert hasattr(cph, "predict")
+        cph.predict(X)
+        assert hasattr(cph, "score")
+        cph.score(X, Y)
+
+    def test_sklearn_cross_val_score_accept_model(self, X, Y):
+        from sklearn.model_selection import cross_val_score
+        from sklearn.model_selection import GridSearchCV
+
+        base_model = sklearn_adapter(WeibullAFTFitter, event_col="E")
+        wf = base_model(penalizer=1.0)
+        assert len(cross_val_score(wf, X, Y, cv=3)) == 3
+
+    def test_sklearn_GridSearchCV_accept_model(self, X, Y):
+        from sklearn.model_selection import cross_val_score
+        from sklearn.model_selection import GridSearchCV
+
+        base_model = sklearn_adapter(WeibullAFTFitter, event_col="E")
+
+        grid_params = {
+            "penalizer": 10.0 ** np.arange(-2, 3),
[…]
