+++ lifelines/utils/__init__.py
@@ -656,6 +656,11 @@ def k_fold_cross_validation(
     -------
     results: list
       (k,1) list of scores for each fold. The scores can be anything.
+
+    See Also
+    ---------
+    lifelines.utils.sklearn_adapter.sklearn_adapter
+
     """
     # Make sure fitters is a list
     fitters = _to_list(fitters)
