### 2019-06-10 fc6c4e6
improve stability of AFT models
