+++ lifelines/fitters/log_logistic_fitter.py
@@ -82,8 +82,8 @@ class LogLogisticFitter(KnownModelParametericUnivariateFitter):
 
     def _cumulative_hazard(self, params, times):
         alpha_, beta_ = params
-        return np.log1p((times / alpha_) ** beta_)
+        return np.logaddexp(beta_ * (np.log(times) - np.log(alpha_)), 0)
 
     def _log_1m_sf(self, params, times):
         alpha_, beta_ = params
-        return -np.log1p((times / alpha_) ** -beta_)
+        return -np.logaddexp(-beta_ * (np.log(times) - np.log(alpha_)), 0)
