### 2019-06-10 81f051a
handle when interval censoring can have 0 as left-hand-side values
