### 2019-06-11 #1317 @becarioprecario
Hi,

I have been a bit stuck with other stuff. Please, may I ask you to send me the guidelines again on how to make the review.

Thanks!

Virgilio
