### 2019-06-12 a2a7bcc
small updates
