### 2019-06-12 a91113f
small doc changes
