+++ docs/conf.py
@@ -15,6 +15,7 @@
 from datetime import date
 import sys
 import os
+import lifelines
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -40,6 +41,7 @@ extensions = [
 
 exclude_patterns = ["_build", "jupyter_notebooks/.ipynb_checkpoints/*.ipynb"]
 
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
@@ -63,8 +65,9 @@ copyright = "2014-{},  Cam Davidson-Pilon".format(date.today().year)
 #
 # The short X.Y version.
 
-version = "0.21.3"
-# The full version, including dev info
+# The short X.Y version.
+version = lifelines.__version__
+# The full version, including alpha/beta/rc tags.
 release = version
 
 # The language for content autogenerated by Sphinx. Refer to documentation
