### 2019-06-17 #1317 @trallard
Hi @becarioprecario  the checklist for the review is at the top of this issue 👆🏼

For reference the guidelines for reviewers are located here: [https://joss.theoj.org/about#reviewer_guidelines](https://joss.theoj.org/about#reviewer_guidelines)
