### 2019-06-17 f9dc4c1
better docs on convergence + inversion
