### 2019-06-17 ac5b7ac
better docs on convergence + inversion
