+++ lifelines/fitters/coxph_fitter.py
@@ -1622,6 +1622,14 @@ the following on the original dataset, df: `df.groupby(%s).size()`. Expected is 
         errorbar_kwargs:
             pass in additional plotting commands to matplotlib errorbar command
 
+        Examples
+        ---------
+
+        >>> from lifelines import datasets, CoxPHFitter
+        >>> rossi = datasets.load_rossi()
+        >>> cph = CoxPHFitter().fit(rossi, 'week', 'arrest')
+        >>> cph.plot(hazard_ratios=True)
+
         Returns
         -------
         ax: matplotlib axis
@@ -1707,7 +1715,15 @@ the following on the original dataset, df: `df.groupby(%s).size()`. Expected is 
         >>> cph.plot_covariate_groups('prio', values=np.arange(0, 15), cmap='coolwarm')
 
         >>> # multiple variables at once
-        >>> cph.plot_covariate_groups(['prio', 'paro'], values=[[0, 0], [5, 0], [10, 0], [0, 1], [5, 1], [10, 1]], cmap='coolwarm')
+        >>> cph.plot_covariate_groups(['prio', 'paro'], values=[
+        >>>  [0,  0],
+        >>>  [5,  0],
+        >>>  [10, 0],
+        >>>  [0,  1],
+        >>>  [5,  1],
+        >>>  [10, 1]
+        >>> ], cmap='coolwarm')
+        >>>
 
         >>> # if you have categorical variables, you can simply things:
         >>> cph.plot_covariate_groups(['dummy1', 'dummy2', 'dummy3'], values=np.eye(3))
