### 2019-06-18 e36dda2
Pad counts based on largest number for the first tick to align group labels, explicitly pop kwargs to allow tick labels configuration
