### 2019-06-18 11b5363
Set xlimit after setting xticks to maintain alignment for at-risk counts
