+++ lifelines/plotting.py
@@ -272,8 +272,8 @@ def add_at_risk_counts(*fitters, **kwargs):
     # Set ticks and labels on bottom
     ax2.xaxis.tick_bottom()
     # Match tick numbers and locations
-    ax2.set_xlim(ax.get_xlim())
     ax2.set_xticks(ax.get_xticks())
+    ax2.set_xlim(ax.get_xlim())
     # Remove ticks, need to do this AFTER moving the ticks
     remove_ticks(ax2, x=True, y=True)
     # Add population size at times
