### 2019-06-18 6f88203
Formatting with black
