### 2019-06-18 f189d93
Move calculation of longest count into first tick check
