### 2019-06-18 91e00ff
Use format for compatibility with python 3.5
