### 2019-06-18 27b1a62
Remove black inserted comma after kwarg in plot_lifetimes
