### 2019-06-18 9fcd60f
Remove another trailing black-inserted comma for compatibility with python 3.5
