### 2019-06-19 52fba6e
Align ticks in at-risk count table (#747)
* Pad counts based on largest number for the first tick to align group labels, explicitly pop kwargs to allow tick labels configuration

* Set xlimit after setting xticks to maintain alignment for at-risk counts

* Formatting with black

* Move calculation of longest count into first tick check

* Use format for compatibility with python 3.5

* Remove black inserted comma after kwarg in plot_lifetimes

* Remove another trailing black-inserted comma for compatibility with python 3.5

* Set at-risk xticks based on ticks falling within plot xlim

* Allow 
[…]
