+++ lifelines/plotting.py
@@ -246,11 +246,11 @@ def add_at_risk_counts(*fitters, **kwargs):
     from matplotlib import pyplot as plt
 
     # Axes and Figure can't be None
-    ax = kwargs.get("ax", None)
+    ax = kwargs.pop("ax", None)
     if ax is None:
         ax = plt.gca()
 
-    fig = kwargs.get("fig", None)
+    fig = kwargs.pop("fig", None)
     if fig is None:
         fig = plt.gcf()
 
@@ -258,7 +258,7 @@ def add_at_risk_counts(*fitters, **kwargs):
         labels = [f._label for f in fitters]
     else:
         # Allow None, in which case no labels should be used
-        labels = kwargs["labels"]
+        labels = kwargs.pop("labels", None)
         if labels is None:
             labels = [None] * len(fitters)
     # Create another axes where we can put size ticks
@@ -271,28 +271,38 @@ def add_at_risk_counts(*fitters, **kwargs):
     remove_spines(ax2, ["top", "right", "bottom", "left"])
     # Set ticks and labels on bottom
     ax2.xaxis.tick_bottom()
-    # Match tick numbers and locations
-    ax2.set_xlim(ax.get_xlim())
-    ax2.set_xticks(ax.get_xticks())
+    # Set limit
+    min_time, max_time = ax.get_xlim()
+    ax2.set_xlim(min_time, max_time)
+    # Set ticks to kwarg or visible ticks
+    xticks = kwargs.pop("xticks", None)
+    if xticks is None:
+        xticks = [xtick for xtick in ax.get_xticks() if min_time <= xtick <= max_time]
+    ax2.set_xticks(xticks)
     # Remove ticks, need to do this AFTER moving the ticks
     remove_ticks(ax2, x=True, y=True)
     # Add population size at times
     ticklabels = []
     for tick in ax2.get_xticks():
         lbl = ""
-        for f, l in zip(fitters, labels):
+        # Get counts at tick
+        counts = [f.durations[f.durations >= tick].shape[0] for f in fitters]
+        # Create tick label
+        for l, c in zip(labels, counts):
             # First tick is prepended with the label
             if tick == ax2.get_xticks()[0] and l is not None:
+                # Get length of largest count
+                max_length = len(str(max(counts)))
                 if is_latex_enabled():
-                    s = "\n{}\\quad".format(l) + "{}"
+                    s = "\n{}\\quad".format(l) + "{{:>{}d}}".format(max_length)
                 else:
-                    s = "\n{}   ".format(l) + "{}"
+                    s = "\n{}   ".format(l) + "{{:>{}d}}".format(max_length)
             else:
                 s = "\n{}"
-            lbl += s.format(f.durations[f.durations >= tick].shape[0])
+            lbl += s.format(c)
         ticklabels.append(lbl.strip())
     # Align labels to the right so numbers can be compared easily
-    ax2.set_xticklabels(ticklabels, ha="right")
+    ax2.set_xticklabels(ticklabels, ha="right", **kwargs)
 
     # Add a descriptive headline.
     ax2.xaxis.set_label_coords(0, ax2_ypos)
