### 2019-06-21 fb14cf8
small refactor
