### 2019-06-21 74c44fb
remove necessary dep on sklearn
