+++ tests/utils/test_utils.py
@@ -1029,6 +1029,7 @@ class TestSklearnAdapter:
         assert clf.best_params_ == {"l1_ratio": 0.5, "model_ancillary": False, "penalizer": 0.01}
         assert clf.predict(X).shape[0] == X.shape[0]
 
+    @pytest.mark.xfail
     def test_joblib(self, X, Y):
         from joblib import dump, load
 
