+++ lifelines/fitters/coxph_fitter.py
@@ -1648,14 +1648,16 @@ the following on the original dataset, df: `df.groupby(%s).size()`. Expected is 
         errorbar_kwargs.setdefault("capsize", 3)
 
         z = inv_normal_cdf(1 - self.alpha / 2)
+        user_supplied_columns = True
 
         if columns is None:
+            user_supplied_columns = False
             columns = self.hazards_.index
 
         yaxis_locations = list(range(len(columns)))
         log_hazards = self.hazards_.loc[columns].values.copy()
 
-        order = np.argsort(log_hazards)
+        order = list(range(len(columns) - 1, -1, -1)) if user_supplied_columns else np.argsort(log_hazards)
 
         if hazard_ratios:
             exp_log_hazards = np.exp(log_hazards)
