+++ tests/utils/test_utils.py
@@ -13,6 +13,7 @@ from numpy.random import randn
 from lifelines import CoxPHFitter, WeibullAFTFitter
 from lifelines.datasets import load_regression_dataset, load_larynx, load_waltons, load_rossi
 from lifelines import utils
+from lifelines import metrics
 from lifelines.utils.sklearn_adapter import sklearn_adapter
 
 
@@ -75,35 +76,35 @@ def test_lstsq_returns_correct_values():
 def test_l1_log_loss_with_no_observed():
     actual = np.array([1, 1, 1])
     predicted = np.array([1, 1, 1])
-    assert utils.l1_log_loss(actual, predicted) == 0.0
+    assert metrics.uncensored_l1_log_loss(actual, predicted) == 0.0
     predicted = predicted + 1
-    assert utils.l1_log_loss(actual, predicted) == np.log(2)
+    assert metrics.uncensored_l1_log_loss(actual, predicted) == np.log(2)
 
 
 def test_l1_log_loss_with_observed():
     E = np.array([0, 1, 1])
     actual = np.array([1, 1, 1])
     predicted = np.array([1, 1, 1])
-    assert utils.l1_log_loss(actual, predicted, E) == 0.0
+    assert metrics.uncensored_l1_log_loss(actual, predicted, E) == 0.0
     predicted = np.array([2, 1, 1])
-    assert utils.l1_log_loss(actual, predicted, E) == 0.0
+    assert metrics.uncensored_l1_log_loss(actual, predicted, E) == 0.0
 
 
 def test_l2_log_loss_with_no_observed():
     actual = np.array([1, 1, 1])
     predicted = np.array([1, 1, 1])
-    assert utils.l2_log_loss(actual, predicted) == 0.0
+    assert metrics.uncensored_l2_log_loss(actual, predicted) == 0.0
     predicted = predicted + 1
-    assert abs(utils.l2_log_loss(actual, predicted) - np.log(2) ** 2) < 10e-8
+    assert abs(metrics.uncensored_l2_log_loss(actual, predicted) - np.log(2) ** 2) < 10e-8
 
 
 def test_l2_log_loss_with_observed():
     E = np.array([0, 1, 1])
     actual = np.array([1, 1, 1])
     predicted = np.array([1, 1, 1])
-    assert utils.l2_log_loss(actual, predicted, E) == 0.0
+    assert metrics.uncensored_l2_log_loss(actual, predicted, E) == 0.0
     predicted = np.array([2, 1, 1])
-    assert utils.l2_log_loss(actual, predicted, E) == 0.0
+    assert metrics.uncensored_l2_log_loss(actual, predicted, E) == 0.0
 
 
 def test_unnormalize():
