### 2019-06-23 6e9025c
V0.21.5 (#749)
* small updates

* better docs on convergence + inversion

* playing around with metrics

* small refactor

* remove necessary dep on sklearn

* v0.21.5

* this is fine

* fail this test

* fail this test
