### 2019-06-23 bac01ee
some doc improvements
