### 2019-06-24 #1317 @becarioprecario
Hi,

I am done with my review now. The package and its documentation are excellent!! I had some troubles with the installation, but due to my local configuration. Installing the package with pip run smoothly.

Best,

Virgilio
