### 2019-06-27 1c14d20
generalizing AFT models
