+++ experiments/generalized_aft.py
@@ -51,790 +51,6 @@ from lifelines.utils import (
 )
 
 
-class DataframeSliceDict:
-    def __init__(self, df, mappings):
-        self.df = df
-        self.mappings = mappings
-        self.size = sum(len(v) for v in self.mappings.values())
-
-    def __getitem__(self, key):
-        return self.df[self.mappings[key]].values
-
-    def __iter__(self):
-        for k in self.mappings:
-            yield (k, self[k])
-
-    def filter(self, ix):
-        return DataframeSliceDict(self.df.loc[ix], self.mappings)
-
-
-class ParametericRegressionFitter(BaseFitter):
-    def __init__(self, alpha=0.05, penalizer=0.0):
-        super(ParametericRegressionFitter, self).__init__(alpha=alpha)
-        self._hazard = egrad(self._cumulative_hazard, argnum=1)  # pylint: disable=unexpected-keyword-arg
-        self.penalizer = penalizer
-
-    def _check_values(self, df, T, E, weights, entries):
-        check_for_numeric_dtypes_or_raise(df)
-        check_nans_or_infs(df)
-        check_nans_or_infs(T)
-        check_nans_or_infs(E)
-        check_positivity(T)
-        check_complete_separation(df, E, T, self.event_col)
-
-        if self.weights_col:
-            if (weights.astype(int) != weights).any() and not self.robust:
-                warnings.warn(
-                    dedent(
-                        """It appears your weights are not integers, possibly propensity or sampling scores then?
-                                        It's important to know that the naive variance estimates of the coefficients are biased. Instead a) set `robust=True` in the call to `fit`, or b) use Monte Carlo to
-                                        estimate the variances. See paper "Variance estimation when using inverse probability of treatment weighting (IPTW) with survival analysis"""
-                    ),
-                    StatisticalWarning,
-                )
-            if (weights <= 0).any():
-                raise ValueError("values in weight column %s must be positive." % self.weights_col)
-
-        if self.entry_col:
-            count_invalid_rows = (entries > T).sum()
-            if count_invalid_rows:
-                warnings.warn("""There exist %d rows where entry > duration.""")
-
-    def _log_hazard(self, params, T, *Xs):
-        # can be overwritten to improve convergence, see WeibullAFTFitter
-        hz = self._hazard(params, T, *Xs)
-        hz = anp.clip(hz, 1e-20, np.inf)
-        return anp.log(hz)
-
-    def _log_1m_sf(self, params, T, *Xs):
-        # equal to log(cdf), but often easier to express with sf.
-        cum_haz = self._cumulative_hazard(params, T, *Xs)
-        return anp.log1p(-anp.exp(-cum_haz))
-
-    def _survival_function(self, params, T, Xs):
-        return anp.clip(anp.exp(-self._cumulative_hazard(params, T, Xs)), 1e-25, 1.0)
-
-    def _log_likelihood_right_censoring(self, params, Ts, E, W, entries, Xs):
-        warnings.simplefilter(action="ignore", category=FutureWarning)
-
-        T = Ts[0]
-        non_zero_entries = entries > 0
-
-        log_hz = self._log_hazard(params, T, Xs)
-        cum_hz = self._cumulative_hazard(params, T, Xs)
-        delayed_entries = self._cumulative_hazard(params, entries[non_zero_entries], Xs.filter(non_zero_entries))
-
-        ll = 0
-        ll = ll + (W * E * log_hz).sum()
-        ll = ll + -(W * cum_hz).sum()
-        ll = ll + (W[non_zero_entries] * delayed_entries).sum()
-        ll = ll / anp.sum(W)
-        return ll
-
-    @CensoringType.right_censoring
-    def fit(
-        self,
-        df,
[…]
