+++ lifelines/fitters/__init__.py
@@ -47,6 +47,7 @@ from lifelines.utils import (
     normalize,
     concordance_index,
     CensoringType,
+    DataframeSliceDict,
 )
 
 __all__ = []
@@ -1070,15 +1071,14 @@ class KnownModelParametericUnivariateFitter(ParametericUnivariateFitter):
     _KNOWN_MODEL = True
 
 
-class ParametericAFTRegressionFitter(BaseFitter):
-    def __init__(self, alpha=0.05, penalizer=0.0, l1_ratio=0.0, fit_intercept=True, model_ancillary=False):
-        super(ParametericAFTRegressionFitter, self).__init__(alpha=alpha)
+class ParametericRegressionFitter(BaseFitter):
+
+    _KNOWN_MODEL = False
+
+    def __init__(self, alpha=0.05, penalizer=0.0):
+        super(ParametericRegressionFitter, self).__init__(alpha=alpha)
         self._hazard = egrad(self._cumulative_hazard, argnum=1)  # pylint: disable=unexpected-keyword-arg
-        self._fitted_parameter_names = [self._primary_parameter_name, self._ancillary_parameter_name]
         self.penalizer = penalizer
-        self.l1_ratio = l1_ratio
-        self.fit_intercept = fit_intercept
-        self.model_ancillary = model_ancillary
 
     def _check_values(self, df, T, E, weights, entries):
         check_for_numeric_dtypes_or_raise(df)
@@ -1106,9 +1106,6 @@ class ParametericAFTRegressionFitter(BaseFitter):
             if count_invalid_rows:
                 warnings.warn("""There exist %d rows where entry > duration.""")
 
-        if self.fit_intercept:
-            check_low_var(df)
-
     def _log_hazard(self, params, T, *Xs):
         # can be overwritten to improve convergence, see WeibullAFTFitter
         hz = self._hazard(params, T, *Xs)
@@ -1120,8 +1117,8 @@ class ParametericAFTRegressionFitter(BaseFitter):
         cum_haz = self._cumulative_hazard(params, T, *Xs)
         return anp.log1p(-anp.exp(-cum_haz))
 
-    def _survival_function(self, params, T, *Xs):
-        return anp.clip(anp.exp(-self._cumulative_hazard(params, T, *Xs)), 1e-25, 1.0)
+    def _survival_function(self, params, T, Xs):
+        return anp.clip(anp.exp(-self._cumulative_hazard(params, T, Xs)), 1e-25, 1.0)
 
     def _log_likelihood_right_censoring(self, params, Ts, E, W, entries, Xs):
         warnings.simplefilter(action="ignore", category=FutureWarning)
@@ -1129,60 +1126,735 @@ class ParametericAFTRegressionFitter(BaseFitter):
         T = Ts[0]
         non_zero_entries = entries > 0
 
-        log_hz = self._log_hazard(params, T, *Xs)
-        cum_hz = self._cumulative_hazard(params, T, *Xs)
-        delayed_entries = self._cumulative_hazard(
-            params, entries[non_zero_entries], *(X[non_zero_entries, :] for X in Xs)
-        )
+        log_hz = self._log_hazard(params, T, Xs)
+        cum_hz = self._cumulative_hazard(params, T, Xs)
+        delayed_entries = self._cumulative_hazard(params, entries[non_zero_entries], Xs.filter(non_zero_entries))
 
         ll = 0
         ll = ll + (W * E * log_hz).sum()
         ll = ll + -(W * cum_hz).sum()
         ll = ll + (W[non_zero_entries] * delayed_entries).sum()
-        ll = ll / np.sum(W)
+        ll = ll / anp.sum(W)
         return ll
 
-    def _log_likelihood_left_censoring(self, params, Ts, E, W, entries, Xs):
[…]

[…]
