+++ lifelines/utils/__init__.py
@@ -1411,6 +1411,11 @@ class DataframeSliceDict:
     def __init__(self, df, mappings):
         self.df = df
         self.mappings = mappings
+        if len(self.mappings["lambda_"]) == 3:
+            import pdb
+
+            pdb.set_trace()
+
         self.size = sum(len(v) for v in self.mappings.values())
 
     def __getitem__(self, key):
@@ -1421,7 +1426,7 @@ class DataframeSliceDict:
             yield (k, self[k])
 
     def filter(self, ix):
-        return DataframeSliceDict(self.df.values[ix], self.mappings)
+        return DataframeSliceDict(self.df[ix], self.mappings)
 
     def iterdicts(self):
         for _, x in self.df.iterrows():
