### 2019-06-28 d026974
things work. Next need to improve perf and edge cases
