+++ lifelines/fitters/__init__.py
@@ -1312,7 +1312,8 @@ class ParametericRegressionFitter(BaseFitter):
         self.weights = weights.copy()
         self.regressors = {
             name: list(df.columns.intersection(cols)) for name, cols in regressors.items()
-        }  # the .intersection perserves order, important!
+        }  # the .intersection preserves order, important!
+        self._LOOKUP_SLICE = self._create_slicer()
 
         df = df.astype(float)
         self._check_values(df, coalesce(Ts[1], Ts[0]), E, weights, entries)
@@ -1321,21 +1322,18 @@ class ParametericRegressionFitter(BaseFitter):
 
         _norm_std = df.std(0)
         _norm_std[_norm_std < 1e-8] = 1.0
-        df_normalized = normalize(df, 0, _norm_std)
 
-        Xs = self._create_Xs_dict(df_normalized)
-
-        self._LOOKUP_SLICE = self._create_slicer(Xs)
         _index = pd.MultiIndex.from_tuples(
             sum(([(name, col) for col in columns] for name, columns in self.regressors.items()), [])
         )
 
+        self._norm_mean = df.mean(0)
         self._norm_std = pd.Series([_norm_std.loc[variable_name] for _, variable_name in _index], index=_index)
 
         _params, self._log_likelihood, self._hessian_ = self._fit_model(
             log_likelihood_function,
             Ts,
-            Xs,
+            self._create_Xs_dict(normalize(df, 0, _norm_std)),
             E.values,
             weights.values,
             entries.values,
@@ -1345,23 +1343,25 @@ class ParametericRegressionFitter(BaseFitter):
         self.params_ = _params / self._norm_std
 
         self.variance_matrix_ = self._compute_variance_matrix()
-        self.standard_errors_ = self._compute_standard_errors(Ts, E.values, weights.values, entries.values, Xs)
+        self.standard_errors_ = self._compute_standard_errors(
+            Ts, E.values, weights.values, entries.values, self._create_Xs_dict(df)
+        )
         self.confidence_intervals_ = self._compute_confidence_intervals()
-        # self._predicted_median = self.predict_median(df)
+        self._predicted_median = self.predict_median(df)
 
     def _create_initial_point(self, Ts, E, entries, weights, Xs):
         return np.zeros(Xs.size)
 
-    def _add_penalty(self, ll, *args):
-        return ll
+    def _add_penalty(self, params, neg_ll, *args):
+        return neg_ll
 
-    def _wrap_ll(self, ll):
+    def _wrap_function_to_covert_array_to_dict(self, function):
         def f(params_array, *args):
             params_dict = {
                 parameter_name: params_array[self._LOOKUP_SLICE[parameter_name]]
                 for parameter_name in self._fitted_parameter_names
             }
-            return ll(params_dict, *args)
+            return function(params_dict, *args)
 
         return f
 
@@ -1372,9 +1372,9 @@ class ParametericRegressionFitter(BaseFitter):
 
         assert initial_point.shape[0] == Xs.size, "initial_point is not the correct shape."
 
-        self._neg_likelihood_with_penalty_function = lambda *args: self._add_penalty(
-            -self._wrap_ll(likelihood)(*args), *args
-        )
+        self._neg_likelihood_with_penalty_function = lambda params, *args: self._wrap_function_to_covert_array_to_dict(
+            self._add_penalty
+        )(params, -self._wrap_function_to_covert_array_to_dict(likelihood)(params, *args), *args)
 
         results = minimize(
             # using value_and_grad is much faster (takes advantage of shared computations) than splitting.
[…]
