+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Changelog
 
+
+- changed ancillary_df to ancillary_df
+
+
 #### 0.21.5 - 2019-06-22
 
 I'm skipping 0.21.4 version because of deployment issues.

[…]
