+++ lifelines/fitters/piecewise_exponential_regression_fitter.py
@@ -1,102 +1,12 @@
 # -*- coding: utf-8 -*-
-from textwrap import dedent
-import warnings
-from datetime import datetime
-from scipy.integrate import trapz
-from scipy.optimize import minimize
-from scipy import stats
+import autograd.numpy as np
+from lifelines.utils import coalesce, _get_index, CensoringType
+from lifelines.fitters import ParametericRegressionFitter
 import pandas as pd
-from autograd import numpy as np
-from autograd import hessian, value_and_grad, elementwise_grad as egrad, grad
-
-from lifelines.fitters import BaseFitter
-from lifelines.plotting import set_kwargs_drawstyle, set_kwargs_ax
-
-from lifelines.utils import (
-    _to_list,
-    _get_index,
-    qth_survival_times,
-    dataframe_interpolate_at_times,
-    ConvergenceError,
-    inv_normal_cdf,
-    string_justify,
-    format_floats,
-    format_p_value,
-    format_exp_floats,
-    coalesce,
-    check_nans_or_infs,
-    pass_for_numeric_dtypes_or_raise_array,
-    check_for_numeric_dtypes_or_raise,
-    check_complete_separation,
-    check_low_var,
-    StatisticalWarning,
-    StatError,
-    median_survival_times,
-    normalize,
-    CensoringType,
-    concordance_index,
-)
 
 
-class PiecewiseExponentialRegressionFitter(BaseFitter):
-    r"""
-    This class implements an Piecewise Exponential model for univariate data. The model has parameterized
-    hazard rate:
-
-    .. math::  h(t\;|\;x) = \begin{cases}
-                        1/\lambda_0(x),  & \text{if $t \le \tau_0$} \\
-                        1/\lambda_1(x) & \text{if $\tau_0 < t \le \tau_1$} \\
-                        1/\lambda_2(x) & \text{if $\tau_1 < t \le \tau_2$} \\
-                        ...
-                      \end{cases}
-
-    and $\lambda_i(x) = \exp(\mathbf{\beta}_i x^T), \;\; \mathbf{\beta}_i = (\beta_{i,1}, \beta_{i,2}, ...)$. That is, each period has a hazard rate, $\lambda_i$ the is the exponential of a linear model. The parameters of each linear model are unique to that period - different periods have different parameters (later we will generalize this).
-
-    Why do I want a model like this? Well, it offers lots of flexibility (at the cost of efficiency though), but importantly I can see:
-
-    1. Influence of variables over time.
-    2. Looking at important variables at specific "drops" (or regime changes). For example, what variables cause the large drop at the start? What variables prevent death at the second billing?
-    3. Predictive power: since we model the hazard more accurately (we hope) than a simpler parametric form, we have better estimates of a subjects survival curve.
-
-    After calling the `.fit` method, you have access to properties like: ``params_``
-    A summary of the fit is available with the method ``print_summary()``
-
-
-    Parameters
-    -----------
-    breakpoints: list
-        a list of times when a new exponential model is constructed.
-
-    alpha: float, optional (default=0.05)
-        the level in the confidence intervals.
-
-    fit_intercept: boolean, optional (default=True)
-        Allow lifelines to add an intercept column of 1s to df.
-
-    penalizer: float, optional (default=0.0)
-        the penalizer coefficient to the size of the coefficients. See `l1_ratio`. Must be equal to or greater than 0.
-
-    Attributes
-    ----------
-    params_ : DataFrame
-        The estimated coefficients
-    confidence_intervals_ : DataFrame
-        The lower and upper confidence intervals for the coefficients
-    durations: Series
-        The event_observed variable provided
-    event_observed: Series
-        The event_observed variable provided
-    weights: Series
[…]
