+++ tests/conftest.py
@@ -6,7 +6,7 @@ import pytest
 def pytest_runtest_setup(item):
     random_seed = np.random.randint(1000)
     print("Seed used in np.random.seed(): %d" % random_seed)
-    np.random.seed(812)
+    np.random.seed(random_seed)
 
 
 def pytest_addoption(parser):

[…]
