### 2019-06-30 d79b594
perf improvements
