### 2019-06-30 24db830
some QOL doc updates
