+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 ### Changelog
 
+#### 0.22.0
+
+##### New features
+ - Ability to create custom parametric regression models by specifying the cumulative hazard. This enables lots more AFT models.
+
+##### API changes
+ - In regression models, the column names in `confidence_intervals_` has changed to include the alpha value.
+ - In regression models, some column names in `.summary` and `.print_summary` has changed to include the alpha value.
+ - Significant changes to internal AFT code.
+
+##### Bug fixes
+
 #### 0.21.5 - 2019-06-22
 
-I'm skipping 0.21.4 version because of deployment issues.
+I'm skipping 0 .21.4 version because of deployment issues.
 
 ##### New features
  - `scoring_method` now a kwarg on `sklearn_adapter`
