### 2019-07-01 f726998
interesting print_summary trials
