### 2019-07-01 #1317 @trallard
Thanks for your time and reviews @becarioprecario  and @sunhwan 🙏🏼

@CameronDevine I can confirm your paper has been accepted now! 🎉  Let's move this to publication, please complete the following actions and let me know once this is done

- Final check on article: DOIS are present, authors are correctly listed
- Create a  new tag for the software and add let me know what final version is
- Deposit the software in Zenodo and share the DOI here
