### 2019-07-01 15f8321
example of cure model
