+++ paper/paper.bib
@@ -150,6 +150,7 @@
         },
     YEAR = {1996},
     JOURNAL = {Statistics in medicine},
+    DOI = "https://doi.org/10.1002/(SICI)1097-0258(19960229)15:4<361::AID-SIM168>3.0.CO;2-4",
 }
 @ARTICLE{flexsurv,
     AUTHOR = {{Christopher Jackson}},
