### 2019-07-02 efad7dd
Fix typo in smoothing hazard curve docs (#758)
There is no such method `hazard_confidence_intervals_`,
but there is a `smoothed_hazard_confidence_intervals_`.
