+++ lifelines/fitters/__init__.py
@@ -24,6 +24,7 @@ from lifelines.plotting import _plot_estimate, set_kwargs_drawstyle, set_kwargs_
 
 from lifelines.utils import (
     qth_survival_times,
+    qth_survival_time,
     _to_array,
     _to_list,
     safe_zip,
@@ -48,6 +49,7 @@ from lifelines.utils import (
     concordance_index,
     CensoringType,
     DataframeSliceDict,
+    _get_index,
 )
 
 __all__ = []
@@ -230,7 +232,7 @@ class UnivariateFitter(BaseFitter):
 
         """
         age = self.survival_function_.index.values[:, None]
-        columns = ["%s - Conditional time remaining to event" % self._label]
+        columns = ["%s - Conditional median duration remaining to event" % self._label]
         return (
             pd.DataFrame(
                 qth_survival_times(self.survival_function_[self._label] * 0.5, self.survival_function_)
@@ -265,6 +267,28 @@ class UnivariateFitter(BaseFitter):
 
     def plot_cumulative_density(self, **kwargs):
         raise NotImplementedError()
+
+    @property
+    def median_(self):
+        """
+        Return the unique time point, t, such that S(t) = 0.5. This is the "half-life" of the population, and a
+        robust summary statistic for the population, if it exists.
+
+        For known parametric models, this should be overwritten by something more accurate.
+        """
+        return self.percentile(0.5)
+
+    def percentile(self, p):
+        """
+        Return the unique time point, t, such that S(t) = p.
+
+        For known parametric models, this should be overwritten by something more accurate.
+        """
+        warnings.warn(
+            "Approximating using `survival_function_`. Try using or increasing the resolution of the timeline kwarg in `.fit(..., timeline=timeline)`.",
+            StatisticalWarning,
+        )
+        return qth_survival_times(p, self.survival_function_)
 
 
 class ParametericUnivariateFitter(UnivariateFitter):
@@ -601,7 +625,7 @@ class ParametericUnivariateFitter(UnivariateFitter):
             "{} = {}".format(
                 justify("hypothesis"),
                 ", ".join(
-                    "%s != %d" % (name, iv) for (name, iv) in zip(self._fitted_parameter_names, self._initial_values)
+                    "%s != %g" % (name, iv) for (name, iv) in zip(self._fitted_parameter_names, self._initial_values)
                 ),
             )
         )
@@ -980,20 +1004,6 @@ class ParametericUnivariateFitter(UnivariateFitter):
         return pd.Series(self._hazard(self._fitted_parameters_, times), index=_to_array(times), name=label)
 
     @property
-    def median_(self):
-        """
-        Return the unique time point, t, such that S(t) = 0.5. This is the "half-life" of the population, and a
-        robust summary statistic for the population, if it exists.
-
-        For known parametric models, this should be overwritten by something more accurate.
-        """
-        warnings.warn(
-            "Approximating the median using `survival_function_`. Try using or increasing the resolution of the timeline kwarg in `.fit(..., timeline=timeline)`.",
-            StatisticalWarning,
-        )
-        return median_survival_times(self.survival_function_)
-
-    @property
     def confidence_interval_(self):
         """
         The confidence interval of the cumulative hazard. This is an alias for ``confidence_interval_cumulative_hazard_``.
@@ -1065,18 +1075,38 @@ class ParametericUnivariateFitter(UnivariateFitter):
             self, estimate=getattr(self, "hazard_"), confidence_intervals=self.confidence_interval_hazard_, **kwargs
         )
 
+    def _conditional_time_to_event_(self):
+        """
[…]
