### 2019-07-02 bc905ac
Generalize aft (#752)
* generalizing AFT models

* generalizing AFT models

* things work. Next need to improve perf and edge cases

* nice, PERF is now much less a snowflake

* wip checkpoint

* perf improvements

* example of cure model

* add percentile to univarite fitters; make conditional_time_to_event_ work for parametric mdoels
