### 2019-07-02 290bf9b
changing coxph hazards_ to params_
