+++ tests/test_estimation.py
@@ -1409,8 +1409,8 @@ class TestRegressionFitters:
 
             # we drop indexes since aaf will have a different "time" index.
             try:
-                hazards = fitter.fit(rossi, duration_col="week", event_col="arrest").hazards_
-                hazards_norm = fitter.fit(normalized_rossi, duration_col="week", event_col="arrest").hazards_
+                hazards = fitter.fit(rossi, duration_col="week", event_col="arrest").params_
+                hazards_norm = fitter.fit(normalized_rossi, duration_col="week", event_col="arrest").params_
             except AttributeError:
                 hazards = fitter.fit(rossi, duration_col="week", event_col="arrest").params_
                 hazards_norm = fitter.fit(normalized_rossi, duration_col="week", event_col="arrest").params_
@@ -1572,7 +1572,7 @@ class TestAFTFitters:
             LogNormalAFTFitter(fit_intercept=False),
             LogLogisticAFTFitter(fit_intercept=False),
         ]:
-            fitter.fit(rossi, "week", "arrest")
+            fitter.fit(rossi, "week", "arrest", ancillary_df=rossi[["intercept"]])
 
     def test_warning_is_present_if_entry_greater_than_duration(self, rossi, models):
         rossi["start"] = 10
@@ -1681,14 +1681,6 @@ class TestAFTFitters:
             accelerated_survival = aft.predict_expectation(subject).squeeze()
             factor = aft.summary.loc[(aft._primary_parameter_name, "prio"), "exp(coef)"]
             npt.assert_allclose(accelerated_survival, baseline_survival * factor)
-
-    def test_aft_predict_acccepts_numpy_arrays_and_dataframes(self, models, rossi):
-        for aft in models:
-            aft.fit(rossi, "week", "arrest")
-
-            subject = aft._norm_mean.to_frame().T
-
-            assert_frame_equal(aft.predict_expectation(subject), aft.predict_expectation(subject.values))
 
     def test_aft_models_can_do_left_censoring(self, models):
         N = 100
@@ -2242,7 +2234,7 @@ class TestCoxPHFitter:
 
         cph.fit(regression_dataset, "T", "E")
 
-        results = cph.compute_residuals(regression_dataset, "scaled_schoenfeld") - cph.hazards_.values
+        results = cph.compute_residuals(regression_dataset, "scaled_schoenfeld") - cph.params_.values
         npt.assert_allclose(results.iloc[0].values, [0.785518935413, 0.862926592959, 2.479586809860], rtol=5)
         npt.assert_allclose(results.iloc[1].values, [-0.888580165064, -1.037904485796, -0.915334612372], rtol=5)
         npt.assert_allclose(
@@ -2306,12 +2298,6 @@ class TestCoxPHFitter:
         rossi["matched_pairs"] = np.floor(rossi.index / 2.0).astype(int)
         cph.fit(rossi, duration_col="week", event_col="arrest", strata=["matched_pairs"], show_progress=True)
         assert cph.baseline_cumulative_hazard_.shape[1] == 216
-
-    def test_summary(self, rossi, cph):
-        cph.fit(rossi, duration_col="week", event_col="arrest")
-        summary = cph.summary
-        expected_columns = ["coef", "exp(coef)", "se(coef)", "z", "p", "lower 0.95", "upper 0.95", "-log2(p)"]
-        assert all([col in summary.columns for col in expected_columns])
 
     def test_print_summary_with_decimals(self, rossi, cph):
         import sys
[…]
