+++ lifelines/fitters/cox_time_varying_fitter.py
@@ -60,8 +60,10 @@ class CoxTimeVaryingFitter(BaseFitter):
 
     Attributes
     ----------
-    hazards_ : Series
-        The estimated hazards
+    params_ : Series
+        The estimated coefficients
+    hazard_ratios_ : Series
+        The exp(coefficients)
     confidence_intervals_ : DataFrame
         The lower and upper confidence intervals for the hazard coefficients
     event_observed: Series
@@ -199,7 +201,7 @@ class CoxTimeVaryingFitter(BaseFitter):
         self._norm_mean = df.mean(0)
         self._norm_std = df.std(0)
 
-        hazards_ = self._newton_rhaphson(
+        params_ = self._newton_rhaphson(
             normalize(df, self._norm_mean, self._norm_std),
             events,
             start,
@@ -210,7 +212,8 @@ class CoxTimeVaryingFitter(BaseFitter):
             step_size=step_size,
         )
 
-        self.hazards_ = pd.Series(hazards_, index=df.columns, name="coef") / self._norm_std
+        self.params_ = pd.Series(params_, index=df.columns, name="coef") / self._norm_std
+        self.hazard_ratios_ = pd.Series(np.exp(self.params_), index=df.columns, name="exp(coef)")
         self.variance_matrix_ = -inv(self._hessian_) / np.outer(self._norm_std, self._norm_std)
         self.standard_errors_ = self._compute_standard_errors(
             normalize(df, self._norm_mean, self._norm_std), events, start, stop, weights
@@ -257,7 +260,7 @@ class CoxTimeVaryingFitter(BaseFitter):
             yield function(stratified_X, stratified_events, stratified_start, stratified_stop, stratified_W, *args)
 
     def _compute_z_values(self):
-        return self.hazards_ / self.standard_errors_
+        return self.params_ / self.standard_errors_
 
     def _compute_p_values(self):
         U = self._compute_z_values() ** 2
@@ -267,34 +270,36 @@ class CoxTimeVaryingFitter(BaseFitter):
         ci = 100 * (1 - self.alpha)
         z = inv_normal_cdf(1 - self.alpha / 2)
         se = self.standard_errors_
-        hazards = self.hazards_.values
+        hazards = self.params_.values
         return pd.DataFrame(
             np.c_[hazards - z * se, hazards + z * se],
             columns=["%g%% lower-bound" % ci, "%g%% upper-bound" % ci],
-            index=self.hazards_.index,
+            index=self.params_.index,
         )
 
     @property
     def summary(self):
-        """
-        Summary statistics describing the fit.
+        """Summary statistics describing the fit.
+        Set alpha property in the object before calling.
 
         Returns
         -------
-        df: DataFrame
-            contains columns coef, exp(coef), se(coef), z, p, lower, upper
-        """
+        df : DataFrame
+            Contains columns coef, np.exp(coef), se(coef), z, p, lower, upper"""
         ci = 100 * (1 - self.alpha)
+        z = inv_normal_cdf(1 - self.alpha / 2)
         with np.errstate(invalid="ignore", divide="ignore"):
-            df = pd.DataFrame(index=self.hazards_.index)
-            df["coef"] = self.hazards_
-            df["exp(coef)"] = np.exp(self.hazards_)
+            df = pd.DataFrame(index=self.params_.index)
+            df["coef"] = self.params_
+            df["exp(coef)"] = self.hazard_ratios_
             df["se(coef)"] = self.standard_errors_
+            df["coef lower %g%%" % ci] = self.confidence_intervals_["%g%% lower-bound" % ci]
+            df["coef upper %g%%" % ci] = self.confidence_intervals_["%g%% upper-bound" % ci]
[…]
