+++ lifelines/statistics.py
@@ -730,7 +730,7 @@ def proportional_hazard_test(
             result += StatisticalResult(
                 p_values,
                 T,
-                name=[(c, transform_name) for c in fitted_cox_model.hazards_.index],
+                name=[(c, transform_name) for c in fitted_cox_model.params_.index],
                 test_name="proportional_hazard_test",
                 null_distribution="chi squared",
                 degrees_of_freedom=1,
@@ -751,7 +751,7 @@ def proportional_hazard_test(
         result = StatisticalResult(
             p_values,
             T,
-            name=fitted_cox_model.hazards_.index.tolist(),
+            name=fitted_cox_model.params_.index.tolist(),
             test_name="proportional_hazard_test",
             time_transform=time_transform,
             null_distribution="chi squared",
