+++ tests/utils/test_utils.py
@@ -957,11 +957,7 @@ class TestSklearnAdapter:
 
         base_model = sklearn_adapter(WeibullAFTFitter, event_col="E")
 
-        grid_params = {
-            "penalizer": 10.0 ** np.arange(-2, 3),
-            "l1_ratio": [0.05, 0.5, 0.95],
-            "model_ancillary": [True, False],
-        }
+        grid_params = {"penalizer": 10.0 ** np.arange(-2, 3), "model_ancillary": [True, False]}
         clf = GridSearchCV(base_model(), grid_params, cv=4)
         clf.fit(X, Y)
 
