### 2019-07-03 d5835ce
make this more robust, it was failing ci
