### 2019-07-03 71a56b9
V0.22.0 (#757)
* wip

* some QOL doc updates

* v0.22.0

* interesting print_summary trials

* Generalize aft (#752)

* generalizing AFT models

* generalizing AFT models

* things work. Next need to improve perf and edge cases

* nice, PERF is now much less a snowflake

* wip checkpoint

* perf improvements

* example of cure model

* add percentile to univarite fitters; make conditional_time_to_event_ work for parametric mdoels

* fixing tests

* changing coxph hazards_ to params_

* more work

* lint

* fix plotting

* docs changelog

* make this more robust, it was failing ci

* forgot to 
[…]
