+++ experiments/generalized_aft.py
@@ -1,867 +1,39 @@
 # -*- coding: utf-8 -*-
-import collections
-from functools import wraps
-import sys
-import warnings
-from datetime import datetime
-
-# pylint: disable=wrong-import-position
-warnings.simplefilter(action="ignore", category=FutureWarning)
-
-from textwrap import dedent
-
+from lifelines.fitters import ParametricRegressionFitter
 import numpy as np
 import autograd.numpy as anp
-from autograd import hessian, value_and_grad, elementwise_grad as egrad, grad
-from autograd.differential_operators import make_jvp_reversemode
-from scipy.optimize import minimize
-from scipy import stats
-import pandas as pd
-from numpy.linalg import inv, pinv
+import matplotlib.pyplot as plt
 
 
-from lifelines.plotting import _plot_estimate, set_kwargs_drawstyle, set_kwargs_ax
-from lifelines.fitters import BaseFitter
+class Custom(ParametricRegressionFitter):
 
-from lifelines.utils import (
-    qth_survival_times,
-    _to_array,
-    _to_list,
-    safe_zip,
-    dataframe_interpolate_at_times,
-    ConvergenceError,
-    inv_normal_cdf,
-    string_justify,
-    format_floats,
-    format_p_value,
-    format_exp_floats,
-    coalesce,
-    check_nans_or_infs,
-    pass_for_numeric_dtypes_or_raise_array,
-    check_for_numeric_dtypes_or_raise,
-    check_complete_separation,
-    check_low_var,
-    check_positivity,
-    StatisticalWarning,
-    StatError,
-    median_survival_times,
-    normalize,
-    concordance_index,
-    CensoringType,
-)
-
-
-class DataframeSliceDict:
-    def __init__(self, df, mappings):
-        self.df = df
-        self.mappings = mappings
-        self.size = sum(len(v) for v in self.mappings.values())
-
-    def __getitem__(self, key):
-        return self.df[self.mappings[key]].values
-
-    def __iter__(self):
-        for k in self.mappings:
-            yield (k, self[k])
-
-    def filter(self, ix):
-        return DataframeSliceDict(self.df.loc[ix], self.mappings)
-
-
-class ParametericRegressionFitter(BaseFitter):
-    def __init__(self, alpha=0.05, penalizer=0.0):
-        super(ParametericRegressionFitter, self).__init__(alpha=alpha)
-        self._hazard = egrad(self._cumulative_hazard, argnum=1)  # pylint: disable=unexpected-keyword-arg
-        self.penalizer = penalizer
-
-    def _check_values(self, df, T, E, weights, entries):
-        check_for_numeric_dtypes_or_raise(df)
-        check_nans_or_infs(df)
-        check_nans_or_infs(T)
-        check_nans_or_infs(E)
-        check_positivity(T)
-        check_complete_separation(df, E, T, self.event_col)
-
-        if self.weights_col:
-            if (weights.astype(int) != weights).any() and not self.robust:
-                warnings.warn(
-                    dedent(
-                        """It appears your weights are not integers, possibly propensity or sampling scores then?
-                                        It's important to know that the naive variance estimates of the coefficients are biased. Instead a) set `robust=True` in the call to `fit`, or b) use Monte Carlo to
-                                        estimate the variances. See paper "Variance estimation when using inverse probability of treatment weighting (IPTW) with survival analysis"""
-                    ),
-                    StatisticalWarning,
-                )
-            if (weights <= 0).any():
-                raise ValueError("values in weight column %s must be positive." % self.weights_col)
-
-        if self.entry_col:
-            count_invalid_rows = (entries > T).sum()
-            if count_invalid_rows:
-                warnings.warn("""There exist %d rows where entry > duration.""")
-
-    def _log_hazard(self, params, T, *Xs):
[…]
