### 2019-07-04 d6a0050
docs improvement, closes #761
