### 2019-07-05 4464d13
experiments with gamma
