### 2019-07-05 faf594d
experiments with gamma
