### 2019-07-05 44d32e1
experiments with gamma
