+++ lifelines/__init__.py
@@ -17,6 +17,7 @@ from lifelines.fitters.weibull_aft_fitter import WeibullAFTFitter
 from lifelines.fitters.log_logistic_aft_fitter import LogLogisticAFTFitter
 from lifelines.fitters.log_normal_aft_fitter import LogNormalAFTFitter
 from lifelines.fitters.piecewise_exponential_regression_fitter import PiecewiseExponentialRegressionFitter
+from lifelines.fitters.generalized_gamma_fitter import GeneralizedGammaFitter
 
 
 from lifelines.version import __version__
@@ -33,6 +34,7 @@ __all__ = [
     "CoxTimeVaryingFitter",
     "AalenJohansenFitter",
     "LogNormalFitter",
+    "GeneralizedGammaFitter",
     "LogLogisticFitter",
     "WeibullAFTFitter",
     "LogLogisticAFTFitter",
