+++ .travis.yml
@@ -7,6 +7,10 @@ python:
 env:
 - export PANDAS_VERSION=0.23.4
 - export PANDAS_VERSION=0.24.1
+- export PANDAS_VERSION=0.25.0rc0
+- export NUMPY_VERSION=1.16.4
+- export NUMPY_VERSION=1.15.4
+- export NUMPY_VERSION=1.17.0rc1
 # Enable newer 3.7 without globally enabling sudo and dist: xenial for other build jobs
 matrix:
   include:
