+++ .travis.yml
@@ -1,27 +1,18 @@
 language: python
 cache: pip
-dist: trusty
+dist: xenial
 python:
    - "3.5"
    - "3.6"
+   - "3.7"
 env:
-- export PANDAS_VERSION=0.23.4
-- export PANDAS_VERSION=0.24.1
-- export PANDAS_VERSION=0.25.0rc0
-- export NUMPY_VERSION=1.16.4
-- export NUMPY_VERSION=1.15.4
-- export NUMPY_VERSION=1.17.0rc1
-# Enable newer 3.7 without globally enabling sudo and dist: xenial for other build jobs
-matrix:
-  include:
-  - python: 3.7
-    dist: xenial
-    sudo: true
-    env: export PANDAS_VERSION=0.24.1
-  - python: 3.7
-    dist: xenial
-    sudo: true
-    env: export PANDAS_VERSION=0.23.4
+  matrix:
+    - export PANDAS_VERSION=0.23.4
+    - export PANDAS_VERSION=0.24.1
+    - export PANDAS_VERSION=0.25.0rc0
+    - export NUMPY_VERSION=1.16.4
+    - export NUMPY_VERSION=1.15.4
+    - export NUMPY_VERSION=1.17.0rc1
 before_install:
   - ls
   # - sudo apt-get update
