### 2019-07-06 b1f29d4
some doc improvements
