+++ lifelines/fitters/log_normal_fitter.py
@@ -12,12 +12,12 @@ class LogNormalFitter(KnownModelParametericUnivariateFitter):
     This class implements an Log Normal model for univariate data. The model has parameterized
     form:
 
-    .. math::  S(t) = 1 - \Phi((\log(t) - \mu)/\sigma),   \sigma >0
+    .. math::  S(t) = 1 - \Phi(\frac{\log(t) - \mu}{\sigma}),  \;\; \sigma >0
 
     where :math:`\Phi` is the CDF of a standard normal random variable.
     This implies the cumulative hazard rate is
 
-    .. math::  H(t) = -\log(1 - \Phi((\log(t) - \mu)/\sigma))
+    .. math::  H(t) = -\log(1 - \Phi(\frac{\log(t) - \mu}{\sigma})
 
     After calling the `.fit` method, you have access to properties like: ``survival_function_``, ``mu_``, ``sigma_``.
     A summary of the fit is available with the method ``print_summary()``
@@ -44,7 +44,7 @@ class LogNormalFitter(KnownModelParametericUnivariateFitter):
         The lower and upper confidence intervals for the survival function
     cumumlative_density_ : DataFrame
         The estimated cumulative density function (with custom timeline if provided)
-    confidence_interval_cumumlative_density_ : DataFrame
+    confidence_interval_cumulative_density_ : DataFrame
         The lower and upper confidence intervals for the cumulative density
     variance_matrix_ : numpy array
         The variance matrix of the coefficients
