+++ lifelines/fitters/piecewise_exponential_fitter.py
@@ -50,7 +50,7 @@ class PiecewiseExponentialFitter(KnownModelParametericUnivariateFitter):
         The lower and upper confidence intervals for the survival function
     cumumlative_density_ : DataFrame
         The estimated cumulative density function (with custom timeline if provided)
-    confidence_interval_cumumlative_density_ : DataFrame
+    confidence_interval_cumulative_density_ : DataFrame
         The lower and upper confidence intervals for the cumulative density
     variance_matrix_ : numpy array
         The variance matrix of the coefficients
