+++ lifelines/fitters/exponential_fitter.py
@@ -63,7 +63,7 @@ class ExponentialFitter(KnownModelParametericUnivariateFitter):
         The entry array provided, or None
     cumumlative_density_ : DataFrame
         The estimated cumulative density function (with custom timeline if provided)
-    confidence_interval_cumumlative_density_ : DataFrame
+    confidence_interval_cumulative_density_ : DataFrame
         The lower and upper confidence intervals for the cumulative density
     """
 
