### 2019-07-06 57b44f2
try this 4th order approximation. variance feels more stable
