### 2019-07-06 230592d
try this 4th order approximation. variance feels more stable
