+++ lifelines/utils/gamma.py
@@ -6,13 +6,23 @@ from autograd.extend import primitive, defvjp
 from autograd.numpy.numpy_vjps import unbroadcast_f
 
 
-delta = 1e-9
+delta = 1e-6
 gammainc = primitive(_scipy_gammainc)
 
 
 defvjp(
     gammainc,
-    lambda ans, a, x: unbroadcast_f(a, lambda g: g * (gammainc(a + delta, x) - gammainc(a - delta, x)) / (2 * delta)),
+    lambda ans, a, x: unbroadcast_f(
+        a,
+        lambda g: g
+        * (
+            -gammainc(a + 2 * delta, x)
+            + 8 * gammainc(a + delta, x)
+            - 8 * gammainc(a - delta, x)
+            + gammainc(a - 2 * delta, x)
+        )
+        / (12 * delta),
+    ),
     lambda ans, a, x: unbroadcast_f(x, lambda g: g * np.exp(-x) * np.power(x, a - 1) / gamma(a)),
 )
 
@@ -22,6 +32,16 @@ gammaincc = primitive(_scipy_gammaincc)
 
 defvjp(
     gammaincc,
-    lambda ans, a, x: unbroadcast_f(a, lambda g: g * (gammaincc(a + delta, x) - gammaincc(a - delta, x)) / (2 * delta)),
+    lambda ans, a, x: unbroadcast_f(
+        a,
+        lambda g: g
+        * (
+            -gammaincc(a + 2 * delta, x)
+            + 8 * gammaincc(a + delta, x)
+            - 8 * gammaincc(a - delta, x)
+            + gammaincc(a - 2 * delta, x)
+        )
+        / (12 * delta),
+    ),
     lambda ans, a, x: unbroadcast_f(x, lambda g: -g * np.exp(-x) * np.power(x, a - 1) / gamma(a)),
 )
