### 2019-07-07 a993787
use autograd-gamma
