### 2019-07-07 17ad92c
use autograd-gamma
