### 2019-07-07 7fdab12
use autograd-gamma
