+++ lifelines/utils/gamma.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
-from autograd_gamma import gammainc, gammaincc, gamma
+from autograd_gamma import gammainc, gammaincc
+from autograd.scipy.special import gamma, gammaln
