+++ lifelines/fitters/generalized_gamma_fitter.py
@@ -2,7 +2,7 @@
 import autograd.numpy as np
 from scipy.special import gammainccinv
 from lifelines.fitters import KnownModelParametericUnivariateFitter
-from lifelines.utils.gamma import gammaincc
+from lifelines.utils.gamma import gammaincc, gammainc, gammaln
 
 
 class GeneralizedGammaFitter(KnownModelParametericUnivariateFitter):
@@ -12,10 +12,7 @@ class GeneralizedGammaFitter(KnownModelParametericUnivariateFitter):
     This class implements a Generalized Gamma model for univariate data. The model has parameterized
     form:
 
-    TODO:
-    .. math::  S(t) = \exp\left(-\left(\frac{t}{\lambda}\right)^\rho\right),   \lambda > 0, \rho > 0,
-
-    which implies the cumulative hazard rate is
+    The cumulative hazard rate is
 
     .. math:: H(t) = \left(\frac{t}{\lambda}\right)^\rho,
 
@@ -23,7 +20,7 @@ class GeneralizedGammaFitter(KnownModelParametericUnivariateFitter):
 
     .. math::  h(t) = \frac{\rho}{\lambda}\left(\frac{t}{\lambda}\right)^{\rho-1}
 
-    After calling the `.fit` method, you have access to properties like: ``cumulative_hazard_``, ``survival_function_``, ``lambda_`` and ``rho_``.
+    After calling the `.fit` method, you have access to properties like: ``cumulative_hazard_``, ``survival_function_``, ``alpha_``, ``lambda_`` and ``rho_``.
     A summary of the fit is available with the method ``print_summary()``.
 
     Parameters
@@ -35,13 +32,13 @@ class GeneralizedGammaFitter(KnownModelParametericUnivariateFitter):
     Examples
     --------
 
-    >>> from lifelines import WeibullFitter
+    >>> from lifelines import GeneralizedGammaFitter
     >>> from lifelines.datasets import load_waltons
     >>> waltons = load_waltons()
-    >>> wbf = WeibullFitter()
-    >>> wbf.fit(waltons['T'], waltons['E'])
-    >>> wbf.plot()
-    >>> print(wbf.lambda_)
+    >>> ggf = GeneralizedGammaFitter()
+    >>> ggf.fit(waltons['T'], waltons['E'])
+    >>> ggf.plot()
+    >>> ggf.summary
 
     Attributes
     ----------
@@ -61,6 +58,8 @@ class GeneralizedGammaFitter(KnownModelParametericUnivariateFitter):
         The fitted parameter in the model
     rho_: float
         The fitted parameter in the model
+    alpha_: float
+        The fitted parameter in the model
     durations: array
         The durations provided
     event_observed: array
@@ -74,11 +73,21 @@ class GeneralizedGammaFitter(KnownModelParametericUnivariateFitter):
 
     _fitted_parameter_names = ["alpha_", "lambda_", "rho_"]
     _bounds = [(0.0, None), (0.0, None), (0.0, None)]
+    _initial_point = np.array([1.0, 1.0, 1.0])
+
+    def _survival_function(self, params, times):
+        alpha_, lambda_, rho_ = params
+        ug = gammaincc(alpha_ / rho_, (times / lambda_) ** rho_)
+        ug = np.clip(ug, 1e-17, 1 - 1e-17)
+        return ug
 
     def _cumulative_hazard(self, params, times):
-        alpha_, lambda_, rho_ = params
-        lg = np.clip(gammaincc(alpha_ / rho_, (times / lambda_) ** rho_), 1e-20, 1 - 1e-20)
-        return -np.log(lg)
+        sf = self._survival_function(params, times)
+        return -np.log(sf)
+
[…]
