### 2019-07-08 b346201
docstrings and some unit tests for gg
