+++ tests/test_estimation.py
@@ -910,6 +910,33 @@ class TestWeibullFitter:
             assert abs(1 - wf.lambda_ / lambda_) < 5 / np.sqrt(N)
 
 
+class TestGeneralizedGammaFitter:
+    def test_exponential_data_inference(self):
+        T = np.random.exponential(5, size=5000)
+        gg = GeneralizedGammaFitter().fit(T)
+        assert gg.summary.loc["alpha_"]["lower 0.95"] < 1 < gg.summary.loc["alpha_"]["upper 0.95"]
+        assert gg.summary.loc["rho_"]["lower 0.95"] < 1 < gg.summary.loc["rho_"]["upper 0.95"]
+
+    def test_weibull_data_inference(self):
+        T = 5 * np.random.exponential(1, size=10000) ** 0.5
+        gg = GeneralizedGammaFitter().fit(T)
+        assert gg.summary.loc["alpha_"]["lower 0.95"] < 1 < gg.summary.loc["alpha_"]["upper 0.95"]
+
+    def test_gamma_data_inference(self):
+        shape = 2
+        T = np.random.gamma(shape=shape, scale=2, size=10000)
+        gg = GeneralizedGammaFitter().fit(T)
+        assert gg.summary.loc["rho_"]["lower 0.95"] < 1 / shape < gg.summary.loc["rho_"]["upper 0.95"]
+        assert gg.summary.loc["alpha_"]["lower 0.95"] < shape < gg.summary.loc["alpha_"]["upper 0.95"]
+
+    def test_lognormal_data_inference(self):
+        T = np.exp(0.5 * np.random.randn(10000) + 1)
+        gg = GeneralizedGammaFitter().fit(T)
+        assert gg.summary.loc["rho_"]["coef"] < 0.05
+        assert gg.summary.loc["lambda_"]["coef"] < 0.05
+        assert gg.summary.loc["alpha_"]["coef"] > 10.0
+
+
 class TestExponentialFitter:
     def test_fit_computes_correct_lambda_(self):
         T = np.array([10, 10, 10, 10], dtype=float)
