### 2019-07-08 #1317 @CamDavidsonPilon
hi @trallard, I have addressed the missing DOI. Anything else? Is there a list I missed?
