### 2019-07-08 2f343ae
GG improvements
