### 2019-07-08 20f20a0
GG improvements
