### 2019-07-09 c75099e
fixing #769 and #768
