### 2019-07-09 3f6b50d
fixing #769 and #768
