### 2019-07-09 336c459
maybe revert: remove any covariate rows that occur at the same time as the death event
