### 2019-07-09 576873d
maybe revert: remove any covariate rows that occur at the same time as the death event
