### 2019-07-10 a4ed7ac
kinda best effort for GG stability to far - I'm going to test different parameterizations next
