+++ experiments/gamma_experiments.py
@@ -2,29 +2,22 @@
 from autograd import numpy as np
 from lifelines.fitters import KnownModelParametericUnivariateFitter, ParametericUnivariateFitter
 from lifelines.utils.gamma import gammainc, gammaincc
-
-
-class GG(ParametericUnivariateFitter):
-
-    _fitted_parameter_names = ["alpha_", "lambda_", "rho_"]
-    _bounds = [(0.0, None), (0.0, None), (0.0, None)]
-
-    def _cumulative_hazard(self, params, times):
-        alpha_, lambda_, rho_ = params
-        lg = np.clip(gammaincc(alpha_ / rho_, (times / lambda_) ** rho_), 1e-20, 1 - 1e-20)
-        return -np.log(lg)
+from lifelines import GeneralizedGammaFitter
 
 
 from lifelines.datasets import load_waltons
 
 df = load_waltons()
 
-gg = GG()
-gg.fit(df["T"], df["E"])
+T = np.arange(1, 100)
+
+gg = GeneralizedGammaFitter()
+gg.fit(T, initial_point=np.array([1.0, 1.0, 1.0]), show_progress=True)
 
 gg.print_summary(3)
+print(gg.variance_matrix_)
 
-
+"""
 class LogGammaFitter(ParametericUnivariateFitter):
 
     _fitted_parameter_names = ["lambda_", "delta_", "kappa_"]
@@ -44,3 +37,4 @@ lg = LogGammaFitter()
 lg.fit(df["T"], df["E"])
 
 lg.print_summary(3)
+"""
