### 2019-07-11 88e970b
maybe stable gamma model
