### 2019-07-11 33ab639
maybe stable gamma model
