### 2019-07-11 503e9b9
fix potential overflow bug
