### 2019-07-11 5e5efc0
better, direct sol to #772
