### 2019-07-11 b5758dc
allow fitters to specify their minimi routine
