### 2019-07-11 f3206cc
allow fitters to specify their minimi routine
