+++ experiments/gamma_experiments.py
@@ -2,7 +2,7 @@
 from autograd import numpy as np
 from lifelines.fitters import KnownModelParametericUnivariateFitter, ParametericUnivariateFitter
 from lifelines.utils.gamma import gammainc, gammaincc
-from lifelines import GeneralizedGammaFitter
+from lifelines import *
 
 
 from lifelines.datasets import load_waltons
@@ -13,7 +13,14 @@ T = np.arange(1, 100)
 
 gg = GeneralizedGammaFitter()
 # gg.fit(T, show_progress=True)
-gg.fit(T, initial_point=np.array([0.0, 0.6, 1.0]), show_progress=True)
+gg.fit(T, show_progress=True)  # initial_point=np.array([4.5, 0.1, 9]))
 
 gg.print_summary(3)
-print(gg.variance_matrix_)
+
+
+lg = LogNormalFitter().fit(T)
+lg.print_summary()
+
+
+lg = WeibullFitter().fit(T)
+lg.print_summary()
