+++ docs/lifelines.fitters.rst
@@ -1,15 +1,14 @@
 Univariate Models
 ==================
 
-
-aalen\_johansen\_fitter
+AalenJohansenFitter
 ------------------------------------------------
 
 .. automodule:: lifelines.fitters.aalen_johansen_fitter
     :members:
     :undoc-members:
 
-breslow\_fleming\_harrington\_fitter
+BreslowFlemingHarringtonFitter
 -------------------------------------------------------------
 
 .. automodule:: lifelines.fitters.breslow_fleming_harrington_fitter
@@ -17,42 +16,42 @@ breslow\_fleming\_harrington\_fitter
     :undoc-members:
 
 
-exponential\_fitter
+ExponentialFitter
 --------------------------------------------
 
 .. automodule:: lifelines.fitters.exponential_fitter
     :members:
     :undoc-members:
 
-kaplan\_meier\_fitter
+KaplanMeierFitter
 ----------------------------------------------
 
 .. automodule:: lifelines.fitters.kaplan_meier_fitter
     :members:
     :undoc-members:
 
-log\_logistic\_fitter
+LogLogisticFitter
 ----------------------------------------------
 
 .. automodule:: lifelines.fitters.log_logistic_fitter
     :members:
     :undoc-members:
 
-log\_normal\_fitter
+LogNormalFitter
 --------------------------------------------
 
 .. automodule:: lifelines.fitters.log_normal_fitter
     :members:
     :undoc-members:
 
-nelson\_aalen\_fitter
+NelsonAalenFitter
 ----------------------------------------------
 
 .. automodule:: lifelines.fitters.nelson_aalen_fitter
     :members:
     :undoc-members:
 
-piecewise\_exponential\_fitter
+PiecewiseExponentialFitter
 -------------------------------------------------------
 
 .. automodule:: lifelines.fitters.piecewise_exponential_fitter
@@ -61,10 +60,18 @@ piecewise\_exponential\_fitter
 
 
 
-weibull\_fitter
+WeibullFitter
 ------------------------------------------
 
 .. automodule:: lifelines.fitters.weibull_fitter
+    :members:
+    :undoc-members:
+
+
+GeneralizedGammaFitter
+------------------------------------------
+
+.. automodule:: lifelines.fitters.generalized_gamma_fitter
     :members:
     :undoc-members:
 
@@ -73,7 +80,7 @@ Regression Models
 ==================
 
 
-aalen\_additive\_fitter
+AalenAdditiveFitter
 ------------------------------------------------
 
 .. automodule:: lifelines.fitters.aalen_additive_fitter
@@ -81,14 +88,14 @@ aalen\_additive\_fitter
     :undoc-members:
 
 
-cox\_time\_varying\_fitter
+CoxTimeVaryingFitter
 ---------------------------------------------------
 
 .. automodule:: lifelines.fitters.cox_time_varying_fitter
     :members:
     :undoc-members:
 
-coxph\_fitter
+CoxPHFitter
 --------------------------------------
 
 .. automodule:: lifelines.fitters.coxph_fitter
@@ -96,21 +103,21 @@ coxph\_fitter
     :undoc-members:
 
 
-weibull\_aft\_fitter
+WeibullAFTFitter
 -------------------------------------------------
 
 .. automodule:: lifelines.fitters.weibull_aft_fitter
     :members:
     :undoc-members:
 
-log\_normal\_aft\_fitter
+LogNormalAFTFitter
 ---------------------------------------------------
 
 .. automodule:: lifelines.fitters.log_normal_aft_fitter
     :members:
     :undoc-members:
 
-log\_logistic\_aft\_fitter
+LogLogisticAFTFitter
 -----------------------------------------------------
 
[…]
