+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ##### Bug fixes
  - when using `censors_show` in plotting functions, the censor ticks are now reactive to the estimate being shown.
  - fixed an overflow bug in `KaplanMeierFitter` confidence intervals
+ - improvements in data validation for `CoxTimeVaryingFitter`
 
 
 #### 0.22.0
