+++ docs/Changelog.rst
@@ -1,8 +1,47 @@
 Changelog
 ~~~~~~~~~
 
+0.22.1
+~~~~~~
+
+New features
+''''''''''''
+
+-  New univariate model, ``GeneralizedGammaFitter``. This model contains
+   many sub-models, so it is a good model to check fits.
+-  added a warning when a time-varying dataset had instantaneous deaths.
+-  added a ``initial_point`` option in univariate parametric fitters.
+-  ``initial_point`` kwarg is present in parametric univariate fitters
+   ``.fit``
+-  ``event_table`` is now an attribute on all univariate fitters (if
+   right censoring)
+-  improvements to ``lifelines.utils.gamma``
+
+API changes
+'''''''''''
+
+-  In AFT models, the column names in ``confidence_intervals_`` has
+   changed to include the alpha value.
+-  In AFT models, some column names in ``.summary`` and
+   ``.print_summary`` has changed to include the alpha value.
+-  In AFT models, some column names in ``.summary`` and
+   ``.print_summary`` includes confidence intervals for the exponential
+   of the value.
+
+Bug fixes
+'''''''''
+
+-  when using ``censors_show`` in plotting functions, the censor ticks
+   are now reactive to the estimate being shown.
+-  fixed an overflow bug in ``KaplanMeierFitter`` confidence intervals
+-  improvements in data validation for ``CoxTimeVaryingFitter``
+
+.. _section-1:
+
 0.22.0
 ^^^^^^
+
+.. _new-features-1:
 
 New features
 ''''''''''''
@@ -10,9 +49,11 @@ New features
 -  Ability to create custom parametric regression models by specifying
    the cumulative hazard. This enables new and extensions of AFT models.
 -  ``percentile(p)`` method added to univariate models that solves the
-   equation ``p = S(t⃰)`` for ``t⃰``
+   equation ``p = S(t)`` for ``t``
 -  for parametric univariate models, the ``conditional_time_to_event_``
    is now exact instead of an approximation.
+
+.. _api-changes-1:
 
 API changes
 '''''''''''
@@ -22,11 +63,11 @@ API changes
    is more clear (what is a hazard anyways?)
 -  In Cox models, a new ``hazard_ratios_`` attribute is available which
    is the exponentiation of ``params_``.
--  In regression models, the column names in ``confidence_intervals_``
-   has changed to include the alpha value.
--  In regression models, some column names in ``.summary`` and
+-  In Cox models, the column names in ``confidence_intervals_`` has
+   changed to include the alpha value.
+-  In Cox models, some column names in ``.summary`` and
    ``.print_summary`` has changed to include the alpha value.
--  In regression models, some column names in ``.summary`` and
+-  In Cox models, some column names in ``.summary`` and
    ``.print_summary`` includes confidence intervals for the exponential
[…]
