### 2019-07-14 f232b55
slight perf improvement, and rerun the script to look at large counts
