+++ perf_tests/cp_perf_test.py
@@ -13,7 +13,7 @@ if __name__ == "__main__":
     df = load_rossi()
     df = pd.concat([df] * 16)
     # df = df.reset_index()
-    # df['week'] = np.random.exponential(1, size=df.shape[0])
+    df["week"] = np.random.exponential(1, size=df.shape[0])
     cp = CoxPHFitter()
     start_time = time.time()
     cp.fit(df, duration_col="week", event_col="arrest", batch_mode=True)
