### 2019-07-15 028c98f
setup.py should look at requirements.txt
