### 2019-07-16 #1317 @trallard
Can you please paste the DOI here so I can set it as the archive? As well as the latest version of the package and I can finalise the acceptance
