### 2019-07-17 155e9bb
remove errant print statement
