### 2019-07-18 #1317 @CamDavidsonPilon
Missing DOI entry: https://github.com/CamDavidsonPilon/lifelines/blob/master/paper/paper.bib#L144-L153

Latest version is [0.21.1](https://github.com/CamDavidsonPilon/lifelines/blob/155e9bb9e913e19858a91885a36ac844931c1eee/lifelines/version.py#L4)
