### 2019-07-19 #1317 @trallard
@CamDavidsonPilon Lifelines has been accepted to JOSS 🎉🎉

@becarioprecario and @sunhwan thanks for the time and effort put in this review. Your contribution to JOSS is deeply appreciated. 🙏🏼🙌🏼🙂

@openjournals/joss-eics this paper is ready for acceptance 👾
