### 2019-07-19 #1317 @danielskatz
@trallard - you've jumped the gun slightly.  Let's say that you believe Lifelines is ready to be accepted to JOSS - now @openjournals/joss-eics need to do some final checks.
