### 2019-07-19 b8a1565
some wording and spacing changes
