### 2019-07-19 #1317 @danielskatz
@CamDavidsonPilon - please accept the changes in https://github.com/CamDavidsonPilon/lifelines/pull/781, or disagree with the ones that are not adding needed spaces :)
