### 2019-07-22 2ac3f3e
improve stability (I think) of GG
