+++ docs/Changelog.rst
@@ -1,8 +1,8 @@
 Changelog
 ~~~~~~~~~
 
-0.22.1
-~~~~~~
+0.22.1 - 2019-07-14
+^^^^^^^^^^^^^^^^^^^
 
 New features
 ''''''''''''
@@ -38,8 +38,8 @@ Bug fixes
 
 .. _section-1:
 
-0.22.0
-^^^^^^
+0.22.0 - 2019-07-03
+^^^^^^^^^^^^^^^^^^^
 
 .. _new-features-1:
 
