### 2019-07-22 1c55c56
I think this is more stable
