+++ lifelines/fitters/weibull_fitter.py
@@ -73,7 +73,7 @@ class WeibullFitter(KnownModelParametericUnivariateFitter):
 
     Notes
     ----------
-    Looking for a 3-parameter Weibull model? See notes here: https://lifelines.readthedocs.io/en/latest/jupyter_notebooks/Piecewise%20Exponential%20Models%20and%20Creating%20Custom%20Models.html#3-parameter-Weibull-distribution
+    Looking for a 3-parameter Weibull model? See notes `here <https://lifelines.readthedocs.io/en/latest/jupyter_notebooks/Piecewise%20Exponential%20Models%20and%20Creating%20Custom%20Models.html#3-parameter-Weibull-distribution>`_.
     """
 
     _fitted_parameter_names = ["lambda_", "rho_"]
