### 2019-07-24 f8fd36c
some more attempts to stabilize GG fitter
