+++ docs/Changelog.rst
@@ -1,8 +1,29 @@
 Changelog
 ~~~~~~~~~
 
+0.22.2
+^^^^^^
+
+New features
+''''''''''''
+
+-  lifelines is now compatible with scipy>=1.3.0
+
+Bug fixes
+'''''''''
+
+-  fixed printing error when using robust=True in regression models
+-  ``GeneralizedGammaFitter`` is more stable, maybe.
+-  lifelines was allowing old version of numpy (1.6), but this caused
+   errors when using the library. The correcty numpy has been pinned (to
+   1.14.0+)
+
+.. _section-1:
+
 0.22.1 - 2019-07-14
 ^^^^^^^^^^^^^^^^^^^
+
+.. _new-features-1:
 
 New features
 ''''''''''''
@@ -28,6 +49,8 @@ API changes
    ``.print_summary`` includes confidence intervals for the exponential
    of the value.
 
+.. _bug-fixes-1:
+
 Bug fixes
 '''''''''
 
@@ -36,12 +59,12 @@ Bug fixes
 -  fixed an overflow bug in ``KaplanMeierFitter`` confidence intervals
 -  improvements in data validation for ``CoxTimeVaryingFitter``
 
-.. _section-1:
+.. _section-2:
 
 0.22.0 - 2019-07-03
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-1:
+.. _new-features-2:
 
 New features
 ''''''''''''
@@ -75,7 +98,7 @@ API changes
    could set ``fit_intercept`` to False and not have to set
    ``ancillary_df`` - now one must specify a DataFrame.
 
-.. _bug-fixes-1:
+.. _bug-fixes-2:
 
 Bug fixes
 '''''''''
@@ -84,21 +107,21 @@ Bug fixes
    is now exact instead of an approximation.
 -  fixed a name error bug in ``CoxTimeVaryingFitter.plot``
 
-.. _section-2:
+.. _section-3:
 
 0.21.5 - 2019-06-22
 ^^^^^^^^^^^^^^^^^^^
 
 I’m skipping 0.21.4 version because of deployment issues.
 
-.. _new-features-2:
+.. _new-features-3:
 
 New features
 ''''''''''''
 
 -  ``scoring_method`` now a kwarg on ``sklearn_adapter``
 
-.. _bug-fixes-2:
+.. _bug-fixes-3:
 
 Bug fixes
 '''''''''
@@ -108,12 +131,12 @@ Bug fixes
 -  fixed visual bug that misaligned x-axis ticks and at-risk counts.
    Thanks @christopherahern!
 
-.. _section-3:
+.. _section-4:
 
 0.21.3 - 2019-06-04
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-3:
+.. _new-features-4:
 
 New features
 ''''''''''''
@@ -127,19 +150,19 @@ New features
 -  ``CoxPHFitter.check_assumptions`` now accepts a ``columns`` parameter
    to specify only checking a subset of columns.
 
-.. _bug-fixes-3:
+.. _bug-fixes-4:
 
 Bug fixes
 '''''''''
 
 -  ``covariates_from_event_matrix`` handle nulls better
 
-.. _section-4:
+.. _section-5:
 
 0.21.2 - 2019-05-16
 ^^^^^^^^^^^^^^^^^^^
 
-.. _new-features-4:
+.. _new-features-5:
 
 New features
 ''''''''''''
@@ -163,17 +186,17 @@ API changes
[…]
