+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 #### 0.22.2
 
+##### New features
+ - lifelines is now compatible with scipy>=1.3.0
+
 ##### Bug fixes
  - fixed printing error when using robust=True in regression models
- - `GeneralizedGammaFitter` is more stable.
+ - `GeneralizedGammaFitter` is more stable, maybe.
+ - lifelines was allowing old version of numpy (1.6), but this caused errors when using the library. The correcty numpy has been pinned (to 1.14.0+)
+
 
 
 #### 0.22.1 - 2019-07-14
