+++ reqs/base-requirements.txt
@@ -1,5 +1,5 @@
 numpy>=1.14.0
-scipy>=1.0,<=1.2.1
+scipy>=1.0
 pandas>=0.23.0
 matplotlib>=3.0
 bottleneck>=1.0
