### 2019-07-25 56eed75
remove logsf
