### 2019-07-26 8ed9a14
update travis to new pandas, new numpy, and drop py35
