+++ .travis.yml
@@ -2,17 +2,16 @@ language: python
 cache: pip
 dist: xenial
 python:
-   - "3.5"
    - "3.6"
    - "3.7"
 env:
   matrix:
-    - export PANDAS_VERSION=0.23.4     && export NUMPY_VERSION=1.16.4
-    - export PANDAS_VERSION=0.24.1     && export NUMPY_VERSION=1.16.4
-    - export PANDAS_VERSION=0.25.0rc0  && export NUMPY_VERSION=1.16.4
-    - export NUMPY_VERSION=1.16.4      && export PANDAS_VERSION=0.24.1
-    - export NUMPY_VERSION=1.15.4      && export PANDAS_VERSION=0.24.1
-    - export NUMPY_VERSION=1.17.0rc1   && export PANDAS_VERSION=0.24.1
+    - export PANDAS_VERSION=0.23.4     && export NUMPY_VERSION=1.17.0
+    - export PANDAS_VERSION=0.24.1     && export NUMPY_VERSION=1.17.0
+    - export PANDAS_VERSION=0.25.0     && export NUMPY_VERSION=1.17.0
+    - export NUMPY_VERSION=1.16.4      && export PANDAS_VERSION=0.25.0
+    - export NUMPY_VERSION=1.15.4      && export PANDAS_VERSION=0.25.0
+    - export NUMPY_VERSION=1.17.0      && export PANDAS_VERSION=0.25.0
 before_install:
   - ls
   # - sudo apt-get update
