### 2019-07-28 e1c1e31
small refactor of regression internals
