### 2019-07-28 aaf009b
more cleanup
