+++ lifelines/utils/gamma.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from autograd_gamma import gammaincc, gammainc, gamma, gammaln, gammainccln, gammaincln
