+++ lifelines/fitters/__init__.py
@@ -4,13 +4,10 @@ from functools import wraps
 import sys
 import warnings
 from datetime import datetime
-
-# pylint: disable=wrong-import-position
-warnings.simplefilter(action="ignore", category=FutureWarning)
-
 from textwrap import dedent
 
 import numpy as np
+from numpy.linalg import inv, pinv
 import autograd.numpy as anp
 from autograd.misc import flatten
 from autograd import hessian, value_and_grad, elementwise_grad as egrad, grad
@@ -19,42 +16,10 @@ from scipy.optimize import minimize
 from scipy.integrate import trapz
 from scipy import stats
 import pandas as pd
-from numpy.linalg import inv, pinv
 
 
 from lifelines.plotting import _plot_estimate, set_kwargs_drawstyle, set_kwargs_ax
-
-from lifelines.utils import (
-    qth_survival_times,
-    qth_survival_time,
-    _to_array,
-    _to_list,
-    safe_zip,
-    dataframe_interpolate_at_times,
-    ConvergenceError,
-    inv_normal_cdf,
-    string_justify,
-    format_floats,
-    format_p_value,
-    format_exp_floats,
-    coalesce,
-    check_nans_or_infs,
-    check_entry_times,
-    pass_for_numeric_dtypes_or_raise_array,
-    check_for_numeric_dtypes_or_raise,
-    check_complete_separation,
-    check_low_var,
-    check_positivity,
-    StatisticalWarning,
-    StatError,
-    median_survival_times,
-    normalize,
-    concordance_index,
-    CensoringType,
-    DataframeSliceDict,
-    _get_index,
-    survival_table_from_events,
-)
+from lifelines import utils
 
 __all__ = []
 
@@ -78,11 +43,11 @@ class BaseFitter(object):
             s = """<lifelines.%s>""" % classname
         return s
 
-    @CensoringType.right_censoring
+    @utils.CensoringType.right_censoring
     def fit(*args, **kwargs):
         raise NotImplementedError()
 
-    @CensoringType.right_censoring
+    @utils.CensoringType.right_censoring
     def fit_right_censoring(self, *args, **kwargs):
         """ Alias for ``fit``
 
@@ -211,12 +176,16 @@ class UnivariateFitter(BaseFitter):
         predictions: a scalar if time is a scalar, a numpy array if time in an array.
         """
         if callable(self._estimation_method):
-            return pd.DataFrame(self._estimation_method(_to_array(times)), index=_to_array(times)).loc[times].squeeze()
+            return (
+                pd.DataFrame(self._estimation_method(utils._to_array(times)), index=utils._to_array(times))
+                .loc[times]
+                .squeeze()
+            )
 
         estimate = getattr(self, self._estimation_method)
         if not interpolate:
             return estimate.asof(times).squeeze()
-        return dataframe_interpolate_at_times(estimate, times)
+        return utils.dataframe_interpolate_at_times(estimate, times)
 
     @property
     def conditional_time_to_event_(self):
@@ -250,7 +219,7 @@ class UnivariateFitter(BaseFitter):
         columns = ["%s - Conditional median duration remaining to event" % self._label]
         return (
             pd.DataFrame(
-                qth_survival_times(self.survival_function_[self._label] * 0.5, self.survival_function_)
+                utils.qth_survival_times(self.survival_function_[self._label] * 0.5, self.survival_function_)
                 .sort_index(ascending=False)
                 .values,
                 index=self.survival_function_.index,
@@ -301,9 +270,9 @@ class UnivariateFitter(BaseFitter):
         """
         warnings.warn(
             "Approximating using `survival_function_`. Try using or increasing the resolution of the timeline kwarg in `.fit(..., timeline=timeline)`.",
-            StatisticalWarning,
+            utils.StatisticalWarning,
         )
[…]
