### 2019-07-28 bac18f4
minor change to default plot
