+++ reqs/base-requirements.txt
@@ -3,5 +3,5 @@ scipy>=1.0
 pandas>=0.23.0
 matplotlib>=3.0
 bottleneck>=1.0
-autograd>=1.2
+autograd>=1.3
 autograd-gamma>=0.3
