+++ README.md
@@ -3,10 +3,7 @@
 [![PyPI version](https://badge.fury.io/py/lifelines.svg)](https://badge.fury.io/py/lifelines)
 [![Build Status](https://travis-ci.org/CamDavidsonPilon/lifelines.svg?branch=master)](https://travis-ci.org/CamDavidsonPilon/lifelines)
 [![Coverage Status](https://coveralls.io/repos/github/CamDavidsonPilon/lifelines/badge.svg?branch=master)](https://coveralls.io/github/CamDavidsonPilon/lifelines?branch=master)
-[![Total alerts](https://img.shields.io/lgtm/alerts/g/CamDavidsonPilon/lifelines.svg?logo=lgtm&logoWidth=18)](https://lgtm.com/projects/g/CamDavidsonPilon/lifelines/alerts/)
-[![Language grade: Python](https://img.shields.io/lgtm/grade/python/g/CamDavidsonPilon/lifelines.svg?logo=lgtm&logoWidth=18)](https://lgtm.com/projects/g/CamDavidsonPilon/lifelines/context:python)
 [![Join the chat at https://gitter.im/python-lifelines/Lobby](https://badges.gitter.im/python-lifelines/Lobby.svg)](https://gitter.im/python-lifelines/Lobby)
-[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/ambv/black)
 [![DOI](https://zenodo.org/badge/12420595.svg)](https://zenodo.org/badge/latestdoi/12420595)
 
 
@@ -14,7 +11,7 @@
  Survival analysis was originally developed and applied heavily by the actuarial and medical community. Its purpose was to answer *why do events occur now versus later* under uncertainty (where *events* might refer to deaths, disease remission, etc.). This is great for researchers who are interested in measuring lifetimes: they can answer questions like *what factors might influence deaths?*
 
 But outside of medicine and actuarial science, there are many other interesting and exciting applications of this survival analysis. For example:
-- SaaS providers are interested in measuring customer lifetimes, or time to first behaviors
+- SaaS providers are interested in measuring subscriber lifetimes, or time to some first action
 - inventory stock out is a censoring event for true "demand" of a good.
 - sociologists are interested in measuring political parties' lifetimes, or relationships, or marriages
 - analyzing [Godwin's law](https://raw.githubusercontent.com/lukashalim/GODWIN/master/Kaplan-Meier-Godwin.png) in Reddit comments
@@ -28,7 +25,6 @@ You can install *lifelines* using
 
        pip install lifelines
 
-
 Or getting the bleeding edge version with:
 
        pip install --upgrade --no-deps git+https://github.com/CamDavidsonPilon/lifelines.git
@@ -40,7 +36,7 @@ from the command line.
 See the common [problems/solutions for installing lifelines](https://github.com/CamDavidsonPilon/lifelines/issues?utf8=%E2%9C%93&q=label%3Ainstallation+).
 
 
-## *lifelines* Documentation and an intro to survival analysis
+## *lifelines* documentation and an intro to survival analysis
 
 If you are new to survival analysis, wondering why it is useful, or are interested in *lifelines* examples, API, and syntax, please check out the [Documentation and Tutorials page](http://lifelines.readthedocs.org/en/latest/index.html)
 
