+++ docs/Examples.rst
@@ -33,7 +33,7 @@ Logrank test
 .. note:: The logrank test has maximum power when the assumption of proportional hazards is true. As a consequence, if the survival functions cross, the logrank test will give an inaccurate assessment of differences.
 
 
-The ``lifelines.statistics.logrank_test`` function compares whether the "death" generation process of the two populations are equal:
+The :func:`lifelines.statistics.logrank_test` function compares whether the "death" generation process of the two populations are equal:
 
 .. code-block:: python
 
@@ -61,8 +61,8 @@ The ``lifelines.statistics.logrank_test`` function compares whether the "death" 
     print(results.test_statistic) # 0.528
 
 
-If you have more than two populations, you can use ``pairwise_logrank_test`` (which compares
-each pair in the same manner as above), or ``multivariate_logrank_test`` (which tests the
+If you have more than two populations, you can use :func:`lifelines.statistics.pairwise_logrank_test` (which compares
+each pair in the same manner as above), or :func:`lifelines.statistics.multivariate_logrank_test` (which tests the
 hypothesis that all the populations have the same "death" generation process).
 
 
@@ -95,7 +95,7 @@ Survival differences at a point in time
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 Often analysts want to compare the survival-ness of groups at specific times, rather than comparing the entire survival curves against each other.  For example, analysts may be interested in 5-year survival. Statistically comparing the naive Kaplan-Meier points at a specific time
-actually has reduced power. By transforming the Kaplan-Meier curve, we can recover more power. The function ``statistics.survival_difference_at_fixed_point_in_time_test`` uses
+actually has reduced power. By transforming the Kaplan-Meier curve, we can recover more power. The function :func:`lifelines.statistics.survival_difference_at_fixed_point_in_time_test` uses
 the log(-log) transformation implicitly and compares the survival-ness of populations at a specific point in time.
 
 
@@ -141,7 +141,7 @@ From these results, Aalen's Additive model with a penalizer of 10 is best model 
 Selecting a parametric model using QQ plots
 ###############################################
 
-QQ plots normally are constructed by sorting the values. However, this isn't appropriate when there is censored data. In *lifelines*, there are routines to still create QQ plots with censored data. These are available under ``lifelines.plotting.qq_plots``, and accepts fitted a parametric lifelines model.
+QQ plots normally are constructed by sorting the values. However, this isn't appropriate when there is censored data. In *lifelines*, there are routines to still create QQ plots with censored data. These are available under :func:`lifelines.plotting.qq_plots`, and accepts fitted a parametric lifelines model.
 
 .. code-block:: python
 
@@ -167,7 +167,7 @@ QQ plots normally are constructed by sorting the values. However, this isn't app
 
 This graphical test can be used to invalidate models. For example, in the above figure, we can see that only the log-normal parametric model is appropriate (we expect deviance in the tails, but not too much). Another use case is choosing the correct parametric AFT model.
 
-The ``qq_plot`` also works with left censorship as well.
+The :func:`~lifelines.plotting.qq_plots` also works with left censorship as well.
 
 
 Plotting multiple figures on a plot
@@ -368,7 +368,7 @@ Set the index/timeline of a estimate
 ##############################################
 
 Suppose your dataset has lifetimes grouped near time 60, thus after fitting
-`KaplanMeierFitter`, you survival function might look something like:
+``KaplanMeierFitter``, you survival function might look something like:
 
 .. code-block:: python
 
[…]

[…]
