### 2019-07-31 061901e
more docs cleaup
