### 2019-07-31 b6948ba
more docs cleaup
