### 2019-07-31 5999b94
some fixes
