### 2019-07-31 2698ee8
some fixes
