### 2019-08-02 #1317 @danielskatz
Sorry for the delay in seeing this - I was expecting you to say you had merged the PR, which would have led to a GH notification.  @labarba - this was in the midst of acceptance, perhaps as the on-duty @openjournals/joss-eics you can continue the process
