### 2019-08-03 #1317 @labarba
@CamDavidsonPilon — the listing on the bottom of page 2 bleeds into the margin. Could you add some line breaks to make that render better?
