### 2019-08-03 #1317 @labarba
Also, I had a look at the Zenodo archive here: http://doi.org/10.5281/zenodo.805993
It shows a long list of authors, possibly an automatic listing from the commits.

We request that authors modify the Zenodo-archive metadata so that the title and author list match the JOSS paper. You may also want to add your ORCID to the metadata.
