### 2019-08-03 #1317 @labarba
I see your latest tagged release is v0.22.2, but here we have 0.21.1 — could you double check which version we should be associating the JOSS paper with?
