### 2019-08-03 #1317 @labarba
Also, the Zenodo archive shows the repo name as title. Could you edit the metadata to match the JOSS paper title?
