### 2019-08-03 #1317 @CamDavidsonPilon
> I see your latest tagged release is v0.22.2, but here we have 0.21.1 — could you double check which version we should be associating the JOSS paper with?

If I change the JOSS paper to be associated to 0.22.2, does it need to be re-reviewed, etc.? If not, v0.22.2 is correct.
