### 2019-08-03 a4d76aa
some image improvements
