### 2019-08-04 #1317 @labarba
Huge thanks to our editor: @trallard, and the reviewers: @becarioprecario, @sunhwan — we greatly appreciate your contribution to JOSS 🙏
